### 2017-11-09 #157 @kenarab
Rpolyhedra
Submitting Author: Alejandro Baranek" (@kenarab)
Repository:  https://github.com/qbotics/Rpolyhedra
Version submitted: 0.1.0
Archive: N/A
Version accepted: 0.4.1

A 142 polyhedra Database scraped from PHD as R6 objects and RGL visualizing capabilities. The PHD format was created to describe the geometric polyhedron definitions derived mathematically by Andrew Hume and by the Kaleido program of Zvi Har'El.

```
Package: Rpolyhedra
Title: Polyhedra Database
Version: 0.1.0
    person("Alejandro", "Baranek",,"abaranek@dc.uba.ar", c("aut","com","cre", "cph")),
    person("Leonardo","Belen",,"leobelen@gmail.com", c("aut","com","cph"))
    )
Maintainer: Alejandro Baranek <abaranek@dc.uba.ar>
Description: A 142 polyhedra database scraped from PHD files <http://paulbourke.net/dataformats/phd/> as R6 objects and 'rgl' visualizing capabilities. The PHD format was created to describe the geometric polyhedra definitions derived mathematically <http://www.netlib.org/polyhedra/> by Andrew Hume and by the Kaleido program of Zvi Har'El.
Suggests: knitr, rmarkdown,
    covr
Imports: futile.logger,
         rgl,
         R6,
Collate:
  'polyhedra-lib.R'
  'zzz.R'
```

https://github.com/qbotics/Rpolyhedra

[e.g., "data extraction, because the package parses a scientific data file format"]

data retrieval, because includes the code for downloading the phd files
data extraction, because the package parses a scientific data file format
database access, because it gives access to PHD polyhedra database

Other categories:
legacy sourcecode: as poly is an opensource inciative (from '80s) than no longer compiles, the project gives access to original poly outputs.

    * Scientists and general public who wants to explore polyhedra through RGL.
    * Scientists and general public who wants to study polyhedra's properties and make taxonomies
    * Scientists and general public who wants to make models over polyhedra definitions

RGL includes 5 regular solids. Rpolyhedra provides version of the 5 regular solids and 137 more polyhedra.

* Delete the file `inst/extdata/polyhedra.RDS` for scraping and regenerating it.
* Compile the package with a small number of polyhedra first:
```
.polyhedra <- scrapePolyhedra(max.quant = 5, test = FALSE)
```
* We have to find a way for accelerate the scraping of the whole database. Maybe a reviewer know a tweak.
