### 2017-11-09 #157 @sckott
thanks for your submission @kenarab

Please change the URL to https://github.com/qbotics/Rpolyhedra/ instead of https://github.com/cran/Rpolyhedra/

We need some clarification to assess fit

- What are the scientific uses of these data? You say scientists are a target audience, but what kind?  What would they do with these data?
- I know little about this area. Are there alternative ways to specify polyhedra?  If so, what is most widely used?
