### 2017-11-10 #157 @kenarab
Thank you @sckott.

I corrected the URL, and extended the users with this paragraph:
- Scientists and general public who wants to explore polyhedra through RGL.
- Scientists and general public who wants to study polyhedra's properties and make taxonomies
- Scientists and general public who wants to make models over polyhedra definitions

About
* I know little about this area. Are there alternative ways to specify polyhedra? If so, what is most widely used?

PHD is a legacy open source specification for polyhedra. As we found online that database with solid definition we chose to use it for further research over polyhedra.
For specifying polyhedra solids there is needed to specify only the vertex and the edges. But there are different properties the specialists uses: Schaffi symbols, symmetry/rotation groups, etc.

For example, in Wikipedia there are a lot of information. `https://en.wikipedia.org/wiki/Icosahedron`
The visualizations of wikipedia are generally made in POV-ray. But there is currently no way for accessing a large database of solid (and net) models of polyhedra in R. Thats the target of Rpolyhedra.

Rpolyhedra includes algorithms for triangulating polygon faces.

We didn't work with polyhedra formulation and generation in Rpolyhedra, yet (Is not really our objective neither work). Only gave access to other people work (compilation work). There are private projects with polyhedra databases but RGL includes only models for 5 regular solids as RGL models.
