### 2017-11-16 #157 @noamross
### Editor checks:

---

#### Editor comments

Hi @kenarab, thanks for your submission, I think this is a really neat package.  I can think of a few people who would be interested already who I might ask to review. A few things before we can assign reviewers, though:

-  We require that packages have continuous integration (typically Travis-CI), and code coverage reporting (typically codecov or coveralls).  Looking at your package tests, I see you are testing the scraping functions but not necessarily what goes on in the R6.  We don't require 100% code coverage, but we'd like coverage of the major functionality.  As far as I can tell there are three major functionalities: scraping the data from netlib, parsing the PHD format, and providing R6 objects with polyhedral data. All three should have associated tests. You can use `skip_on_cran()` or `skip_on_travis()` if long-running scraping calls or RGL calls will break that.  Let us know if you'd like some pointers to setting this infrastructure set up.
-  It looks like you have some debugging infrastructure hard-coded into the package right now via the `futile.logger` calls in you `.onLoad()`.  This breaks other testing tools and also results in unneeded/confusing messages bring printed out to the user.  This should be removed. You might keep it on a development branch.
-  I note that the code doesn't run in your vignette because there isn't a `library(rgl)` call.  You also  should consider using [the WebGL capability](https://cran.r-project.org/web/packages/rgl/vignettes/WebGL.html) in the **rgl** package so that you can make a vignette that compiles and shows.
-  Documentation is very sparse - most functions have a single sentence. We try to aim documentation to be sufficient for familiars unfamiliar with the original data source.  R6 classes can be a bit difficult to document, but docs should detail the arguments and formats of the member functions.  We want documentation to be as complete as possible at submission so reviewers can give feedback on it.
