### 2018-03-01 #157 @kenarab
Dear Noam:

We worked hard this week on Rpolyhedra. We are including another polyhedra
database with +700 polyhedra. We already working with TRAVIS. If everything
ok we setup tomorrow the V0.2.0 release including both databases.

All the best, Ale.

2018-02-22 19:32 GMT-03:00 alejandro baranek <alejandrobaranek@gmail.com>:

> Hello Noam:
>
> We are up to push, but couldn’t made the timeframe to concentrate on the
> release. We hope will do it in the next days.
>
> Best regards, Ale.
>
> Enviado desde el móvil
>
> El 22 feb. 2018, a la(s) 13:33, Noam Ross <notifications@github.com>
> escribió:
>
> Hi @kenarab <https://github.com/kenarab>! Just checking to see if you
> have a time frame for following up on this.
>
