### 2018-03-01 #157 @noamross
Great, Ale. Let me know when you want me to look again so I can assign reviewers.  I'd suggest running [`goodpractice::gp`](https://github.com/MangoTheCat/goodpractice) before you do as that's what we run to check packages before sending out to review.
