+++ inst/extdata/retrieve_netlib_polyedra.sh
@@ -1,2 +1,2 @@
-#curl http://www.netlib.org/polyhedra/ /usr/local/qubo/polyhedra
-#wget -r -np -k http://www.netlib.org/polyhedra/
+#!/usr/local/bin/bash
+wget -r -np -k http://www.netlib.org/polyhedra/
