+++ R/db-lib.R
@@ -1,0 +1,928 @@
+#' Gets the path of package data.
+#'
+#' This function is used internally to determine whether the package
+#' is compiled in source or package directory.
+getDataDir <- function() {
+  home.dir <- find.package("Rpolyhedra", lib.loc = NULL, quiet = TRUE)
+  data.subdir <- "inst/extdata/"
+  if (!dir.exists(paste(home.dir, "/", data.subdir, sep = "")))
+    data.subdir <- "extdata/"
+  paste(home.dir, "/", data.subdir, sep = "")
+}
+
+#' Gets the path of Polyhedra RDS database file
+#'
+#' @param polyhedra_rds_filename filename of polyhedra database
+#' @return the path to the Polyhedra database file
+getPolyhedraRDSPath <- function(polyhedra_rds_filename = "polyhedra.RDS") {
+  paste(getDataDir(), polyhedra_rds_filename, sep = "")
+}
+
+
+#' PolyhedronScraperConfiguration
+#'
+#' Abstract class for configuring specific scrapers for Diferent sources
+#'
+#' @section Methods:
+#' \describe{
+#'   \item{\code{initialize()}}{Initializes the object}
+#' }
+#' @import     futile.logger
+#' @importFrom R6 R6Class
+PolyhedronScraperConfiguration.class <- R6::R6Class("PolyhedronScraperConfiguration",
+  public = list(
+    name     = NA,
+    base.dir = NA,
+    initialize = function(name,base.dir) {
+      self$name     <- name
+      self$base.dir <- base.dir
+      self
+    },
+    getName    = function(){
+      self$name
+    },
+    getBaseDir = function(home.dir.data) {
+      paste(home.dir.data,self$base.dir,sep="")
+    },
+    getPolyhedraFiles = function(home.dir.data){
+      stop(gettext("rpoly.abstract_class", domain = "R-Rpolyhedra"))
+    }))
+
+
+#' PolyhedronScraperConfigurationNetlib
+#'
+#' Scraper configuration for Netlib source
+#'
+#' @section Methods:
+#' \describe{
+#'   \item{\code{initialize()}}{initializes the object}
+#'   \item{\code{getPolyhedraFiles(home.dir.data)}}{returns the file names on the netlib database}
+#'   \item{\code{scrape(polyhedron.number, polyhedron.filename)}}{scrapes the object}
+#' }
+#'
+#' @format \code{\link{R6Class}} object.
+#' @docType class
+#' @import futile.logger
+#' @importFrom R6 R6Class
+PolyhedronScraperConfigurationNetlib.class <- R6::R6Class("PolyhedronScraperConfigurationNetlib",
+  inherit = PolyhedronScraperConfiguration.class,
+  public = list(
+    initialize = function(name) {
+      super$initialize(name = "netlib", base.dir = "www.netlib.org/polyhedra/")
+    },
+    getPolyhedraFiles = function(home.dir.data){
+      # wget -r -np -k http://www.netlib.org/polyhedra/ data/www.netlib.org/
+      polyhedra.dir   <- self$getBaseDir(home.dir.data)
+      polyhedra.files <- dir(polyhedra.dir)
+      polyhedra.files <- polyhedra.files[grep("[0-9]+", polyhedra.files)]
+      polyhedra.files <- polyhedra.files[order(as.numeric(polyhedra.files))]
+      polyhedra.files
+    },
+    scrape = function(polyhedron.number, polyhedron.filename){
+      polyhedra.netlib.lines <- readLines(polyhedron.filename)
+      polyhedron.number <- polyhedron.filename
+      polyhedron.number <- strsplit(polyhedron.number,split = "/")[[1]]
[…]
