+++ inst/extdata/dmccooey.com/polyhedra/SquareAntiprismTrapezohedronToroid.txt
@@ -1,0 +1,42 @@
+Square Antiprism-Trapezohedron Toroid
+
+C0 = 0.0721375078778507478291882437893 = sqrt(2 * (17 * sqrt(2) - 24)) / 4
+C1 = 0.112085382291991279528598381347  = sqrt(2 * (10 - 7 * sqrt(2))) / 4
+C2 = 0.270598050073098492199861602683  = sqrt(2 * (2 - sqrt(2))) / 4
+C3 = 0.420448207626857271515562738117  = sqrt(2 * sqrt(2)) / 4
+C4 = 0.653281482438188263928321586714  = sqrt(2 * (2 + sqrt(2))) / 4
+
+V0  = ( C4,  C2,  C3)
+V1  = ( C4, -C2, -C3)
+V2  = (-C4, -C2,  C3)
+V3  = (-C4,  C2, -C3)
+V4  = ( C2, -C4,  C3)
+V5  = ( C2,  C4, -C3)
+V6  = (-C2,  C4,  C3)
+V7  = (-C2, -C4, -C3)
+V8  = ( C2, -C1,  C0)
+V9  = ( C2,  C1, -C0)
+V10 = (-C2,  C1,  C0)
+V11 = (-C2, -C1, -C0)
+V12 = ( C1,  C2,  C0)
+V13 = ( C1, -C2, -C0)
+V14 = (-C1, -C2,  C0)
+V15 = (-C1,  C2, -C0)
+
+Faces:
+{  0,  6, 15, 12,  9 }
+{  0,  9,  8, 13,  4 }
+{  1,  7, 14, 13,  8 }
+{  1,  8,  9, 12,  5 }
+{  2,  4, 13, 14, 11 }
+{  2, 11, 10, 15,  6 }
+{  3,  5, 12, 15, 10 }
+{  3, 10, 11, 14,  7 }
+{  0,  4,  1 }
+{  0,  1,  5 }
+{  2,  6,  3 }
+{  2,  3,  7 }
+{  4,  2,  7 }
+{  4,  7,  1 }
+{  6,  0,  5 }
+{  6,  5,  3 }

[…]
