### 2018-03-13 0f6749d
Changes to v0.2.0
