+++ tests/testthat.R
@@ -1,5 +1,6 @@
 library(testthat)
 library(Rpolyhedra)
+library(stringr)
 library(futile.logger)
 library(rgl)
 
