+++ inst/extdata/dmccooey.com/polyhedra/PentagonalAntiprismTrapezohedronToroid.txt
@@ -1,0 +1,48 @@
+Pentagonal Antiprism-Trapezohedron Toroid
+
+C0 = 0.190983005625052575897706582817 = (3 - sqrt(5)) / 4
+C1 = 0.309016994374947424102293417183 = (sqrt(5) - 1) / 4
+C2 = 0.809016994374947424102293417183 = (1 + sqrt(5)) / 4
+
+V0  = ( 0.5,  0.0,   C2)
+V1  = ( 0.5,  0.0,  -C2)
+V2  = (-0.5,  0.0,   C2)
+V3  = (-0.5,  0.0,  -C2)
+V4  = ( 0.0,   C2,  0.5)
+V5  = ( 0.0,   C2, -0.5)
+V6  = ( 0.0,  -C2,  0.5)
+V7  = ( 0.0,  -C2, -0.5)
+V8  = (  C2,  0.5,  0.0)
+V9  = ( -C2, -0.5,  0.0)
+V10 = ( 0.0,   C0,  0.5)
+V11 = ( 0.0,   C0, -0.5)
+V12 = ( 0.0,  -C0,  0.5)
+V13 = ( 0.0,  -C0, -0.5)
+V14 = (  C1,   C1,   C1)
+V15 = (  C1,   C1,  -C1)
+V16 = ( -C1,  -C1,   C1)
+V17 = ( -C1,  -C1,  -C1)
+V18 = (  C0,  0.5,  0.0)
+V19 = ( -C0, -0.5,  0.0)
+
+Faces:
+{  0,  6, 16, 12, 10 }
+{  1,  8, 18, 15, 11 }
+{  2,  4, 14, 10, 12 }
+{  3,  9, 19, 17, 13 }
+{  4,  5, 15, 18, 14 }
+{  5,  3, 13, 11, 15 }
+{  6,  7, 17, 19, 16 }
+{  7,  1, 11, 13, 17 }
+{  8,  0, 10, 14, 18 }
+{  9,  2, 12, 16, 19 }
+{  2,  6,  0 }
+{  2,  0,  4 }
+{  3,  1,  7 }
+{  3,  7,  9 }
+{  4,  0,  8 }
+{  4,  8,  5 }
+{  5,  8,  1 }
+{  5,  1,  3 }
+{  6,  2,  9 }
+{  6,  9,  7 }
