+++ inst/extdata/dmccooey.com/polyhedra/SquarePrism.txt
@@ -1,0 +1,18 @@
+Square Prism (Cube)
+
+V0 = ( 0.5,  0.5,  0.5)
+V1 = ( 0.5,  0.5, -0.5)
+V2 = ( 0.5, -0.5,  0.5)
+V3 = ( 0.5, -0.5, -0.5)
+V4 = (-0.5,  0.5,  0.5)
+V5 = (-0.5,  0.5, -0.5)
+V6 = (-0.5, -0.5,  0.5)
+V7 = (-0.5, -0.5, -0.5)
+
+Faces:
+{ 0, 1, 5, 4 }
+{ 0, 4, 6, 2 }
+{ 0, 2, 3, 1 }
+{ 7, 3, 2, 6 }
+{ 7, 6, 4, 5 }
+{ 7, 5, 1, 3 }

[…]
