### 2018-03-13 a2b07a4
[bugfix] in getAvailablePolyhedra when no data for a source
