### 2018-03-13 ff8c480
[bugfix] in getAvailablePolyhedra when no data for a source
