### 2018-03-13 b2a4f2f
removed lowercase rpolyhedra vignettes
