### 2018-03-13 e56c386
added Rpolyhedra vignettes (uppercase)
