### 2018-03-13 d9975fb
[refactoring] Renamed record as ledger and splitted in different code. Added required documentation
