+++ R/ledger-lib.R
@@ -1,0 +1,217 @@
+#' ScraperLedger
+#'
+#' Ledger of scraping status of each objects. Allows different type of states:
+#' queued, scraping, scraped, failed, exception, skipped
+#'
+#' For flexible and reproducible configuration for package development
+#'
+#' @section Methods:
+#' \describe{
+#'   \item{\code{initialize()}}{initializes the object}
+#'   \item{\code{addFilename(source, filename)}}{add filename to the ledger}
+#'   \item{\code{getIdFilename(source, filename)}}{Returns id/row of source and filenames parameters in the ledger}
+#'   \item{\code{updateStatus(source, filename, status, status.field = 'status', scraped.name = NA, obs ='')}}{Updates status of source and filenames parameters in Ledger }
+#'   \item{\code{savePreloadedComplexities()}}{Internal method which saves a file with an estimation of time required time to scrape each filename}
+#'   \item{\code{loadPreloadedComplexities()}}{Load a file with an estimation of time required time to scrape each filename}
+#'   \item{\code{getSizeToTimeScrape(sources, time2scrape = 60)}}{Estimates how much filenames could be scraped in a time frame, considering data retrieved with loadPreloadedComplexities}
+#'   \item{\code{resetStatesMetrics()}}{Reset metrics of application of different status values}
+#'   \item{\code{countStatusUse(status.field,status)}}{Add an use to the metrics of status.field and status parameters}
+#'   \item{\code{getFilenamesStatusMode(mode,sources = sort(unique(self$df$source)),max.quant = 0,order.by.time2scrape = FALSE)}}{Get a list of the filenames in the ledger with a defined mode (status agrupation)}
+#'   \item{\code{getFilenamesStatus(status,sources = sort(unique(self$df$source)),max.quant = 0,order.by.time2scrape = FALSE)}}{Get a list of the filenames in the ledger with specified status}
+#' }
+#'
+#' @format \code{\link{R6Class}} object.
+#' @docType class
+#' @import futile.logger
+#' @importFrom R6 R6Class
+#'
+
+ScraperLedger.class <- R6::R6Class("ScraperLedger",
+ public = list(
+   states = NULL,
+   df = NA,
+   preloaded.complexities.filename = NA,
+   preloaded.complexities = NA,
+   initialize = function() {
+     self$df <- data.frame(id           = character(),
+                           source       = character(),
+                           number       = numeric(),
+                           filename     = character(),
+                           start.scrape = as.POSIXct(character()),
+                           end.scrape   = as.POSIXct(character()),
+                           status       = character(),
+                           scraped.name = character(),
+                           status.test  = character(),
+                           obs          = character(),
+                           git.commit   = character(),
+                           time.scraped = numeric(),
+                           preloaded.time2scrape = numeric(),
+                           stringsAsFactors = FALSE)
+     self$resetStatesMetrics()
+     self$loadPreloadedComplexities()
+     self
+   },
+   addFilename = function(source, filename){
+     r <- NULL
+     default.status <- "queued"
+     if (is.null(self$getIdFilename(source, filename))){
+       r <- nrow(self$df)+1
+       status.field <- "status"
+       self$countStatusUse(status.field = status.field, status = default.status)
+       states.row <- which(self$states$status.field %in% status.field &
+                             self$states$status %in% default.status)
+       number <- self$states[states.row,"count"]
+
+       #obtain preloaded.time2scrape
+       row.preloaded.t2s <- which(self$preloaded.complexities$source == source &
+                                    self$preloaded.complexities$filename==filename)
+       if (length(row.preloaded.t2s)==1){
[…]
