+++ R/db-lib.R
@@ -83,7 +83,6 @@ PolyhedronScraperConfigurationNetlib.class <- R6::R6Class("PolyhedronScraperConf
       polyhedron.number <- polyhedron.filename
       polyhedron.number <- strsplit(polyhedron.number,split = "/")[[1]]
       polyhedron.number <- as.numeric(polyhedron.number[length(polyhedron.number)])
-      #TODO check polyhedron number
       current.polyhedron <- Polyhedron.class$new(number=polyhedron.number)
       current.polyhedron$scrapeNetlib(netlib.p3.lines = polyhedra.netlib.lines)
       futile.logger::flog.debug(paste("parsed", polyhedron.filename, "with name",
@@ -135,7 +134,7 @@ PolyhedronScraperConfigurationDmccoey.class <- R6::R6Class("PolyhedronScraperCon
       geodesic.order  <- geodesic.order[1:50,]
       non.geodesic.files <- polyhedra.files[-which(polyhedra.files %in% geodesic.order$filename)]
       polyhedra.files <- c(geodesic.order$filename,non.geodesic.files[order(non.geodesic.files)])
-      #TODO remove geodesic files from proirity when everything works ok.
+      #TODO remove geodesic files from priority when everything works ok.
       polyhedra.files
     },
     scrape = function(polyhedron.number, polyhedron.filename){
@@ -165,198 +164,23 @@ getGitCommit <- function(){
   git_id
 }
 
-#' ScraperRecord
-#' TODO add doc
 
-ScraperRecord.class <- R6::R6Class("ScraperRecord",
-  public = list(
-    states = NULL,
-    df = NA,
-    preloaded.complexities.filename = NA,
-    preloaded.complexities = NA,
-    initialize = function() {
-      self$df <- data.frame(id           = character(),
-                            source       = character(),
-                            number       = numeric(),
-                            filename     = character(),
-                            start.scrape = as.POSIXct(character()),
-                            end.scrape   = as.POSIXct(character()),
-                            status       = character(),
-                            scraped.name = character(),
-                            status.test  = character(),
-                            obs          = character(),
-                            git.commit   = character(),
-                            time.scraped = numeric(),
-                            preloaded.time2scrape = numeric(),
-                            stringsAsFactors = FALSE)
-      self$resetStatesMetrics()
-      self$loadPreloadedComplexities()
-      self
-    },
-    addFilename = function(source, filename){
-      r <- NULL
-      default.status <- "queued"
-      if (is.null(self$getIdFilename(source, filename))){
-        r <- nrow(self$df)+1
-        status.field <- "status"
-        self$countStatusUse(status.field = status.field, status = default.status)
-        #TODO make abstraction as method
-        states.row <- which(self$states$status.field %in% status.field &
-                                    self$states$status %in% default.status)
-        number <- self$states[states.row,"count"]
-
-        #obtain preloaded.time2scrape
-        row.preloaded.t2s <- which(self$preloaded.complexities$source == source &
-                                     self$preloaded.complexities$filename==filename)
-        if (length(row.preloaded.t2s)==1){
-          preloaded.time2scrape <- as.numeric(self$preloaded.complexities[row.preloaded.t2s,"time2scrape"])
-        }
-        else{
-          preloaded.time2scrape <- NA
-        }
-        self$df[r,c("id","source","filename","status")]<- c(r,source,filename,default.status)
-        self$df[r,c("number","preloaded.time2scrape")]<- c(number,preloaded.time2scrape)
-      }
-      r
-    },
-    getIdFilename = function(source, filename){
[…]
