### 2018-03-13 0be3b32
[refactoring] Renamed record as ledger and splitted in different code. Added required documentation
