+++ man/PolyhedronTestTaskEdgesConsistency.class.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/db-lib.R
+\docType{data}
+\name{PolyhedronTestTaskEdgesConsistency.class}
+\alias{PolyhedronTestTaskEdgesConsistency.class}
+\title{PolyhedronTestTaskEdgesConsistency}
+\format{An object of class \code{R6ClassGenerator} of length 24.}
+\usage{
+PolyhedronTestTaskEdgesConsistency.class
+}
+\description{
+A Test task for running edges consistency test for current polyhedron
+}
+\keyword{datasets}
