### 2018-03-13 #157 @kenarab
Noam:

We just pushed release V0.2.0.
https://github.com/qbotics/Rpolyhedra/releases/tag/v0.2.0

We have problems with goodpractice we didn't have time to attend yet.

We are uploading this release in CRAN while waiting the review. After the
reviewers work we hope can make a new release directly from Ropensci.

Thank you very much for your advisor.

Best regards, Ale.

2018-03-08 16:45 GMT-03:00 alejandro baranek <alejandrobaranek@gmail.com>:

> Noam:
>
> We worked hard including the new source. We have persisting small issues.
> For example, Travis fail because it takes a long time to scrape all the
> database. We will make a prescraped polyhedron list for the simpler
> polyhedron. And (without testing) bigger poyhedra will be loaded on demand.
>
> But the problem that's stopped us for publication is goodpractice::gp
> gives a strange error we cannot reproduce with our tests. I hope in the
> next days we'll be able to make the submission.
>
> By the way, this versions includes some exciting feature like a
> reproducible research compliant scraping scheduler (For problems with
> scraping specific models don't stop us to publicate), polyhedra slicing
> (For big polyhedra, like geodesic ones), and exporting Vertices and Edges
> models, and of Course, original Faces models.
>
> We are making a parallel project for making a simple Shinny interface for
> Rpolyhedra. If there is somewhere to host it we will be able to reach
> broader users, like  people who doesn't use R but want to use polyhedrons.
> Architect, desgingers and artists, for example.
>
> Best regards, Ale.
>
> 2018-03-01 18:26 GMT-03:00 Noam Ross <notifications@github.com>:
>
>> Great, Ale. Let me know when you want me to look again so I can assign
>> reviewers. I'd suggest running goodpractice::gp
>> <https://github.com/MangoTheCat/goodpractice> before you do as that's
>> what we run to check packages before sending out to review.
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/ropensci/onboarding/issues/157#issuecomment-369736166>,
>> or mute the thread
>> <https://github.com/notifications/unsubscribe-auth/ACzA1s_cJA2imDk-JSAemkpGrzN7mYneks5taGdogaJpZM4QYpu6>
>> .
>>
>
>
>
> --
>  alejandro baranek
> @ken4rab <https://twitter.com/ken4rab>
> qbotics <http://qbotics.tumblr.com/> | surferinvaders
> <http://surferinvaders.tumblr.com> | algebraic-soundscapes
> <http://imaginary.org/content/algebraic-soundscapes> | surfer-shuffle
> <http://imaginary.org/program/surfer-shuffle>
>

--
 alejandro baranek
@ken4rab <https://twitter.com/ken4rab>
qbotics <http://qbotics.tumblr.com/> | surferinvaders
<http://surferinvaders.tumblr.com> | algebraic-soundscapes
<http://imaginary.org/content/algebraic-soundscapes> | surfer-shuffle
<http://imaginary.org/program/surfer-shuffle>
