+++ vignettes/Rpolyhedra.Rmd
@@ -8,23 +8,35 @@ vignette: >
   \usepackage[utf8]{inputenc}
 ---
 #Introduction
-This package is a curation made based on based on the poly package found on http://www.netlib.org/polyhedra/ ([Original Help message](poly_original_help_message.html)), and the polyhedra database found on https://www.dmccooey.com/polyhedra, both of which provide a library of polyhedron and its representation. As such, Rpolyhedra provides with the following:
+This package is a curation made based on the poly package found on http://www.netlib.org/polyhedra/ ([Original Help message](poly_original_help_message.html)), and the polyhedra database found on http://dmccooey.com/polyhedra, both of which provide polyhedra databases on its own format. As such, Rpolyhedra provides with the following:
 
-1. A module to scrape the polyhedra PHD files provided by the original poly command. 
-1. A library of the scraped polyhedra.
+1. A module to scrape the polyhedra for the different sources found with features for incremental correction of issues found and to be found in scraping process.
+1. A database of the scraped polyhedra.
 1. An R6 polyhedron representation with 'rgl' package visualizing capabilites. 
 
 #Usage
-The package provides a simplified set of functions to scrape, show the scrapped polyhedra and access each individual polyhedron. 
+
+For final users, the package provides a common interface for accesing public polyhedra databases, analize properties, compare and visualize them with RGL. 
+
+For advanced users, the package provides a simplified set of R6 objects to scrape and compare polyhedra databases. 
 
 ##Scrape polyhedra
 In order to scrape, the following code can be executed to retrieve all polyhedra from the original files produced by the poly command. 
 
-``` R
+```R
 library(rgl)
 library(Rpolyhedra)
 polyhedra <- scrapePolyhedraSources(sources.config = PolyhedronScraperConfigurationNetlib.class$new(), max.quant = 5)
 ```
+
+With following code, based on preloaded estimated scraping time, a batch of maximum filenames that could be scraped in 30 seconds will be scraped.
+
+```R
+library(rgl)
+library(Rpolyhedra)
+polyhedra <- scrapePolyhedraSources(sources.config = PolyhedronScraperConfigurationNetlib.class$new(), time2scrape.source = 30)
+```
+
 
 ##Get available polyhedra
 Once the original files had been processed, a simple call to ```getAvailablePolyhedra()``` retrieves a list of the available polyhedra:
