### 2018-03-13 146003c
Removed bigger polyhedron files while waiting for a fancy solution to package size
