+++ inst/extdata/132Pentagons.txt
@@ -1,363 +1,0 @@
-132-Pentagon Polyhedron
-
-C0  = 0.0821365024859431746289276542569
-C1  = 0.0968707119007493965352481928744
-C2  = 0.1328996527392962888499762586852
-C3  = 0.225272416370950346143386900849
-C4  = 0.229752920547361183523937369381
-C5  = 0.260581638424572725834669320746
-C6  = 0.357452350325322122369917513620
-C7  = 0.358570173636287162330608606811
-C8  = 0.364498426416015900399539018979
-C9  = 0.421629947815094272969598125911
-C10 = 0.426193644046761302125467331647
-C11 = 0.559093296786057590975443590333
-C12 = 0.578370052184905727030401874089
-C13 = 0.580178728295464104708673924149
-C14 = 0.594251346963377083923476388361
-C15 = 0.597020450831134836275447090312
-C16 = 0.641229799272000765604371244589
-C17 = 0.736246460876200390531599208443
-C18 = 0.771732304342777240799276217258
-C19 = 0.779082298140416395339515639531
-C20 = 0.819523763334327430066863289210
-C21 = 0.822495454596130355020324821686
-C22 = 0.83895169060947845286507119483452
-C23 = 0.904631957082073529649252475943
-C24 = 0.938748901931751267039282530960
-C25 = 0.961518877247150736674986109291
-
-V0   = (  C1,  0.0,  1.0)
-V1   = (  C1,  0.0, -1.0)
-V2   = ( -C1,  0.0,  1.0)
-V3   = ( -C1,  0.0, -1.0)
-V4   = ( 1.0,   C1,  0.0)
-V5   = ( 1.0,  -C1,  0.0)
-V6   = (-1.0,   C1,  0.0)
-V7   = (-1.0,  -C1,  0.0)
-V8   = ( 0.0,  1.0,   C1)
-V9   = ( 0.0,  1.0,  -C1)
-V10  = ( 0.0, -1.0,   C1)
-V11  = ( 0.0, -1.0,  -C1)
-V12  = (  C4,   C3,  C25)
-V13  = (  C4,   C3, -C25)
-V14  = (  C4,  -C3,  C25)
-V15  = (  C4,  -C3, -C25)
-V16  = ( -C4,   C3,  C25)
-V17  = ( -C4,   C3, -C25)
-V18  = ( -C4,  -C3,  C25)
-V19  = ( -C4,  -C3, -C25)
-V20  = ( C25,   C4,   C3)
-V21  = ( C25,   C4,  -C3)
-V22  = ( C25,  -C4,   C3)
-V23  = ( C25,  -C4,  -C3)
-V24  = (-C25,   C4,   C3)
-V25  = (-C25,   C4,  -C3)
-V26  = (-C25,  -C4,   C3)
-V27  = (-C25,  -C4,  -C3)
-V28  = (  C3,  C25,   C4)
-V29  = (  C3,  C25,  -C4)
-V30  = (  C3, -C25,   C4)
-V31  = (  C3, -C25,  -C4)
-V32  = ( -C3,  C25,   C4)
-V33  = ( -C3,  C25,  -C4)
-V34  = ( -C3, -C25,   C4)
-V35  = ( -C3, -C25,  -C4)
-V36  = ( 0.0,   C7,  C24)
-V37  = ( 0.0,   C7, -C24)
[…]
