### 2018-03-14 2ca93fd
Badges added to README.md
