### 2018-03-14 65f04cb
Minor changes to README.md
