### 2018-03-14 #157 @kenarab
Great Noam.

We already included Badges, and submitted a new release on CRAN with minor
corrections.

At this stage, as the package is a compilation of previous work, we will
not write a paper so we don't need at these time a joint JOSS publication.
We have the plan to implement [[Conway-Hart polyhedra notation]](
https://en.wikipedia.org/wiki/Conway_polyhedron_notation) in a not defined
yet future version. At that moment we will be looking forward to publish a
paper.

We will need some time for comply with the suggestions made. I will keep
you posted.

Best, Ale.

2018-03-13 23:59 GMT-03:00 Noam Ross <notifications@github.com>:

> Editor checks:
>
>    - *Fit*: The package meets criteria for fit <http://policies.md#fit>
>    and overlap <http://policies.md#fit>
>    - *Automated tests:* Package has a testing suite and is tested via
>    Travis-CI or another CI service.
>    - *License:* The package has a CRAN or OSI accepted license
>    - *Repository:* The repository link resolves correctly
>    - *Archive* (JOSS only, may be post-review): The repository DOI
>    resolves correctly
>    - *Version* (JOSS only, may be post-review): Does the release version
>    given match the GitHub release (v1.0.0)?
>
> ------------------------------
> Editor comments
>
> Glad we're getting this going again!
>
> You must include a paper.md <http://joss.theoj.org/about#paper_structure>
> so our reviewers can review it on behalf of JOSS if you want joint JOSS
> publication. This *does* need to happen before reviewers are assigned,
> but I will go ahead and look for reviewers and have them start once you
> have the paper in place.
>
> I note that while you have CI set up, you don't have Travis-CI or Codecov
> badges in your package README. Please add these - they make it easier to
> see package status at a glance. Also, you may now add the ROpenSci status
> badge:
>
>
> It appears goodpractice::gp() bug is actually in the *covr* package. For
> some reason it fails under the the first test, even though your tests pass
> with flying colors. I suggest you submit a bug report to *covr*.
> Otherwise, Here are goodpractice checks
> without using the covr. Nothing serious here - please fix the small
> stuff. Reviewers should consider whether changes to code style or package
> imports are necessary.
>
> Regarding the large file issue: if you have a lot of large polyhedra you
> would like to serve to users but cannot get on CRAN, may I suggest adding
> an after-the-fact download function like download_polyhedra(). This is a
> strategy used by several others; the recently accepted *rtika* packages uses
> this approach
> <https://github.com/ropensci/onboarding/issues/191#issuecomment-370656143>
> to deliver a large .jar after installation. Your decision on this can
> wait until after reviews are in an reviewers have weighed in.
>
> ── GP Rpolyhedra ─────────────────────────────────────
>
> It is good practice to
>
>   ✖ add a "URL" field to DESCRIPTION. It helps users find information about your package online. If your package
>     does not have a homepage, add an URL to GitHub, or the CRAN package package page.
>   ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80
>     characters wide. Try make your lines shorter than 80 characters
>
>     R/db-lib.R:32:1
>     R/db-lib.R:67:1
>     R/db-lib.R:85:1
>     R/db-lib.R:88:1
>     R/db-lib.R:109:1
>     ... and 134 more lines
>
>   ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data.
>     Consider using vapply() instead.
>
>     R/polyhedra-lib.R:142:38
>
>   ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...) expressions. They are error prone
>     and result 1:0 if the expression on the right hand side is zero. Use seq_len() or seq_along() instead.
>
>     R/db-lib.R:435:21
>     R/polyhedra-lib.R:162:17
>     R/polyhedra-lib.R:300:21
>     R/polyhedra-lib.R:473:26
>     R/polyhedra-lib.R:477:28
>     ... and 2 more lines
>
>   ✖ not import packages as a whole, as this can cause name clashes between the imported packages. Instead, import
>     only the specific functions you need.
>   ✖ fix this R CMD check NOTE: installed size is 13.6Mb sub-directories of 1Mb or more: extdata 13.4Mb
> ──────────────────────────────────────────────
>
> ------------------------------
>
> Reviewers:
> Due date:
>
