+++ README.md
@@ -7,21 +7,6 @@
  [![](https://badges.ropensci.org/157_status.svg)](https://github.com/ropensci/onboarding/issues/157)| [![minimal R version](https://img.shields.io/badge/R%3E%3D-3.0.0-blue.svg)](https://cran.r-project.org/) | [![Travis](https://travis-ci.org/qbotics/Rpolyhedra.svg?branch=master)](https://travis-ci.org/qbotics/Rpolyhedra) |
 | [![CRAN](http://www.r-pkg.org/badges/version/Rpolyhedra)](http://cran.r-project.org/package=Rpolyhedra) 
 
-## sources 
-### netlib
- Includes 142 polyhedra definitions.
- The PHD format was created to describe the geometric polyhedron definitions derived mathematically by Andrew Hume and by the Kaleido program of Zvi Har'El.
-
- PHD files were generated using [poly2](http://www.netlib.org/poly2/readme) library (no longer mantained). Althought the code is available, specific programming skills are required to run it.
- 
-PDH files can be found in `extdata/www.netlib.org/polyhedra/index.html`
-
-### Dmccooey
-Includes 767 polyhedra definitions.
-The [polyhedra database](http://dmccooey.com/polyhedra/) built by David Mccooey has an open format which has been scraped to feed RPolyhedra database
-
-dmccooney files can be found in `extdata/dmccooey.com/polyhedra/`
-
 # How to get started
 ```R
 install.packages("Rpolyhedra")
@@ -63,6 +48,20 @@ for (polyhedron.name in polyhedra.2.draw) {
 }
 
 ```
+## sources 
+### netlib
+ Includes 142 polyhedra definitions.
+ The PHD format was created to describe the geometric polyhedron definitions derived mathematically by Andrew Hume and by the Kaleido program of Zvi Har'El.
+
+ PHD files were generated using [poly2](http://www.netlib.org/poly2/readme) library (no longer mantained). Althought the code is available, specific programming skills are required to run it.
+ 
+PDH files can be found in `extdata/www.netlib.org/polyhedra/index.html`
+
+### Dmccooey
+Includes 767 polyhedra definitions.
+The [polyhedra database](http://dmccooey.com/polyhedra/) built by David Mccooey has an open format which has been scraped to feed RPolyhedra database
+
+dmccooney files can be found in `extdata/dmccooey.com/polyhedra/`
 
 # Troubleshooting
 
