### 2018-03-15 84abf90
Update README.md
fix CRAN url
