+++ README.md
@@ -5,7 +5,7 @@
   | Release | Usage | Development |
 |:--------|:------|:------------|
  [![](https://badges.ropensci.org/157_status.svg)](https://github.com/ropensci/onboarding/issues/157)| [![minimal R version](https://img.shields.io/badge/R%3E%3D-3.0.0-blue.svg)](https://cran.r-project.org/) | [![Travis](https://travis-ci.org/qbotics/Rpolyhedra.svg?branch=master)](https://travis-ci.org/qbotics/Rpolyhedra) |
-| [![CRAN](http://www.r-pkg.org/badges/version/Rpolyhedra)](http://cran.r-project.org/package=Rpolyhedra) 
+| [![CRAN](http://www.r-pkg.org/badges/version/Rpolyhedra)](https://cran.r-project.org/package=Rpolyhedra) 
 
 # How to get started
 ```R
