### 2018-03-16 65d6f97
Scraped vertices and edges
