+++ R/zzz.R
@@ -44,7 +44,7 @@
   assign(".polyhedra", value = .polyhedra, envir = parent.env(environment()))
   scrapePolyhedraSources(max.quant.config.schedule = 0,
                          max.quant.scrape = 0,
-                         time2scrape.source = 40, #40 seconds of building scraping polyhedra
+                         time2scrape.source = 0.5, #40 seconds of building scraping polyhedra
                          sources.config = .available.sources,
                          retry.scrape = FALSE)
 }
