### 2018-03-16 f903663
Scraped vertices and edges
