+++ R/ledger-lib.R
@@ -10,7 +10,7 @@
 #'   \item{\code{initialize()}}{initializes the object}
 #'   \item{\code{addFilename(source, filename)}}{add filename to the ledger}
 #'   \item{\code{getIdFilename(source, filename)}}{Returns id/row of source and filenames parameters in the ledger}
-#'   \item{\code{updateStatus(source, filename, status, status.field = 'status', scraped.name = NA, obs ='')}}{Updates status of source and filenames parameters in Ledger }
+#'   \item{\code{updateStatus(source, filename, status, status.field = 'status', scraped.polyhedron = NA, obs ='')}}{Updates status of source and filenames parameters in Ledger }
 #'   \item{\code{savePreloadedComplexities()}}{Internal method which saves a file with an estimation of time required time to scrape each filename}
 #'   \item{\code{loadPreloadedComplexities()}}{Load a file with an estimation of time required time to scrape each filename}
 #'   \item{\code{getSizeToTimeScrape(sources, time2scrape = 60)}}{Estimates how much filenames could be scraped in a time frame, considering data retrieved with loadPreloadedComplexities}
@@ -30,8 +30,8 @@ ScraperLedger.class <- R6::R6Class("ScraperLedger",
  public = list(
    states = NULL,
    df = NA,
-   preloaded.complexities.filename = NA,
-   preloaded.complexities = NA,
+   preloaded.data.filename = NA,
+   preloaded.data = NA,
    initialize = function() {
      self$df <- data.frame(id           = character(),
                            source       = character(),
@@ -41,10 +41,14 @@ ScraperLedger.class <- R6::R6Class("ScraperLedger",
                            end.scrape   = as.POSIXct(character()),
                            status       = character(),
                            scraped.name = character(),
-                           status.test  = character(),
-                           obs          = character(),
-                           git.commit   = character(),
-                           time.scraped = numeric(),
+                           scraped.vertices   = numeric(),
+                           scraped.faces      = numeric(),
+                           status.test        = character(),
+                           obs                = character(),
+                           git.commit         = character(),
+                           time.scraped       = numeric(),
+                           preloaded.vertices    = numeric(),
+                           preloaded.faces       = numeric(),
                            preloaded.time2scrape = numeric(),
                            stringsAsFactors = FALSE)
      self$resetStatesMetrics()
@@ -63,10 +67,10 @@ ScraperLedger.class <- R6::R6Class("ScraperLedger",
        number <- self$states[states.row,"count"]
 
        #obtain preloaded.time2scrape
-       row.preloaded.t2s <- which(self$preloaded.complexities$source == source &
-                                    self$preloaded.complexities$filename==filename)
+       row.preloaded.t2s <- which(self$preloaded.data$source == source &
+                                    self$preloaded.data$filename==filename)
        if (length(row.preloaded.t2s)==1){
-         preloaded.time2scrape <- as.numeric(self$preloaded.complexities[row.preloaded.t2s,"time2scrape"])
+         preloaded.time2scrape <- as.numeric(self$preloaded.data[row.preloaded.t2s,"time2scrape"])
        }
        else{
          preloaded.time2scrape <- NA
@@ -87,7 +91,7 @@ ScraperLedger.class <- R6::R6Class("ScraperLedger",
      r
    },
    updateStatus = function(source, filename, status, status.field = "status",
-                           scraped.name = NA, obs =""){
+                           scraped.polyhedron = NA, obs =""){
      if (is.null(obs)){
        obs <- ""
      }
@@ -105,12 +109,26 @@ ScraperLedger.class <- R6::R6Class("ScraperLedger",
[…]
