### 2018-03-17 57f6686
Included preloaded data of vertices & faces
