### 2018-03-17 a2ee9d9
getAvailablePolyhedra Gives more Info
