### 2018-03-17 982b028
getAvailablePolyhedra Gives more Info
