+++ man/maxWithoutNA.Rd
@@ -1,0 +1,13 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/ledger-lib.R
+\name{maxWithoutNA}
+\alias{maxWithoutNA}
+\title{maxWithoutNA
+Function that returns NA if all elements are NA, and the max value not NA, if not.}
+\usage{
+maxWithoutNA(x)
+}
+\description{
+maxWithoutNA
+Function that returns NA if all elements are NA, and the max value not NA, if not.
+}
