### 2018-03-17 ee871ce
[bugfix] updateCalculatedFields. enabled Exception in scraper
