### 2018-03-17 0aea2ef
[bugfix] updateCalculatedFields. enabled Exception in scraper
