### 2018-03-17 4cb18bc
Minimum setup for passing tests in Travis
