### 2018-03-17 84ee835
Considering vertices and faces for ordering scraping
