### 2018-03-17 dd1f816
Considering vertices and faces for ordering scraping
