+++ R/zzz.R
@@ -42,10 +42,9 @@
     .polyhedra <- PolyhedronDatabase.class$new()
   }
   assign(".polyhedra", value = .polyhedra, envir = parent.env(environment()))
-  scrapePolyhedraSources(max.quant.config.schedule = 20, #change
+  scrapePolyhedraSources(max.quant.config.schedule = 0,
                          max.quant.scrape = 0,
-                         time2scrape.source = 4.2, #4.2 seconds for scraping all Platonic solids
-                         #time2scrape.source = 40, #40 seconds of building scraping polyhedra
+                         time2scrape.source = 80, #80 seconds of building scraping polyhedra
                          sources.config = .available.sources,
                          retry.scrape = FALSE)
 }
