### 2018-03-17 ca1a150
Release version including vertices and faces and skipped polyhedra in getAvailablePolyhedra
