### 2018-03-17 396a9f2
Release version including vertices and faces and skipped polyhedra in getAvailablePolyhedra
