### 2018-03-21 200237a
New Release. Solving unnecesary RDS file saving
