+++ R/polyhedra-lib.R
@@ -405,7 +405,7 @@ norm <- function(vector){
 #'   \item{\code{calculateMassCenter(size = 1, vertices.3d)}}{Calculates the object's Mass Center for parameter
 #'         vertices}
 #'   \item{\code{getNormalizedSize()}}{Normalizes the volume of the object to a tetrahedron bounding box}#'
-#'   \item{\code{getPositionedVertices(size,origin)}}{Returns the vertices adjusted to size and origin parameters}#'
+#'   \item{\code{getTransformedVertices(size,origin)}}{Returns the vertices adjusted to size and origin parameters}#'
 #'   \item{\code{buildRGL(size = 1, origin = c(0, 0, 0), normalize.size = TRUE)}}{
 #'   Builds the RGL model, taking the object's size, the origin}
 #'   \item{\code{exportToXML()}}{Gets an XML representation out of the polyhedron object}
@@ -422,27 +422,36 @@ norm <- function(vector){
 #' @field solid polyhedron list of edges which generate a solid (netlib|dmccooey)
 #' @field dih Dih attribute (netlib)
 #' @field vertices Polyhedron vertices list (netlib|dmccooey)
+#' @field mass.center polyhedron mass center
+#' @field transformation.matrix transformation matrix for calculations and visualizing polyhedron
+# private
+#' @field edges.cont  Edges count
+#' @field edges.check Edges check degree property
+#' @field vertices.centered centered vertices for applying transformation matrices
 #' @field vertices.rgl Polyhedron triangulated vertices list for RGL
 #' @field solid.triangulated Polyhedron solid (triangulated) for RGL visualization
-#' @field mass.center polyhedron mass center
+
 #' @format \code{\link{R6Class}} object.
 #' @docType class
 #' @import rgl testthat XML
 #' @importFrom R6 R6Class
-PolyhedronStateDefined.class <- R6::R6Class("PolyhedronStateDefined", inherit = PolyhedronState.class,
-private = list(mass.center = NA,
-#infered state
-edges.cont = 0,
-edges.check = NULL,
-#rgl aux members
-vertices.rgl = NULL,
-solid.triangulated = NULL
-),
-public = list(number = NA, source = NA, name = NA, symbol = NA, dual = NA, sfaces = NA, svertices = NA,
-net = NA, hinges = NA, solid = NA, dih = NA, vertices = NA, edges = NULL,
-initialize = function(source, number,
-name, vertices, solid, symbol=NULL, dual=NULL, sfaces=NULL, svertices = NULL, net = NULL,  hinges = NULL, dih = NULL) {
-    super$initialize(source = source,number = number)
+PolyhedronStateDefined.class <- R6::R6Class("PolyhedronStateDefined",
+  inherit = PolyhedronState.class,
+  private = list(mass.center = NULL,
+                  #infered state
+                  edges.cont = 0,
+                  edges.check = NULL,
+                  #rgl aux members
+                  vertices.centered = NULL,
+                  vertices.rgl = NULL,
+                  solid.triangulated = NULL
+                  ),
+  public = list(name = NA, symbol = NA, dual = NA, sfaces = NA, svertices = NA,
+                net = NA, hinges = NA, solid = NA, dih = NA, vertices = NA, edges = NULL,
+                transformation.matrix = NA,
+  initialize = function(source, number,name, vertices, solid, symbol=NULL, dual=NULL,
+                        sfaces=NULL, svertices = NULL, net = NULL,  hinges = NULL, dih = NULL) {
+    super$initialize(source = source, number = number)
     self$name <- name
     self$vertices <- vertices
     self$solid <- solid
[…]
