### 2018-04-10 13ef66a
[bugfix] in building extra vertex for RGL
