### 2018-04-10 e3e6340
[bugfix] in building extra vertex for RGL
