### 2018-04-10 5539ec0
save private vertices definition with normalized size included
