### 2018-04-10 e70831b
[bugfix] test cases. Removed exception catching for easy developing. Expected to pass Travis CI for #12
