+++ man/PolyhedronStateNetlibScraper.class.Rd
@@ -34,7 +34,8 @@ Scrapes polyhedra from a PHD file format.
   \item{\code{setupLabelsOrder()}}{Sets up the order of labels included in PHD file}
   \item{\code{getDataFromLabel(label)}}{Gets data from the Label}
   \item{\code{scrape()}}{Scrapes the data from the PHD file}
-  \item{\code{buildRGL(size = 1, origin = c(0, 0, 0), normalize.size = TRUE)}}{Builds the \code{RGL} mmodel}
+  \item{\code{applyTransformationMatrix(transformation.matrix)}}{Apply transformation matrix to polyhedron}
+  \item{\code{buildRGL(transformation.matrix)}}{Builds the \code{RGL} mmodel}
 }
 }
 
