+++ man/PolyhedronState.class.Rd
@@ -27,7 +27,8 @@ This abstract class provide the basis from which polyhedron state class derivate
   \item{\code{addError(current.error)}}{Adds an error to the error string and log it as info}
   \item{\code{scrape()}}{Scrapes the polyhedra folder files}
   \item{\code{geSolid()}}{returns the object corresponding to the solid}
-  \item{\code{buildRGL(size = 1, origin = c(0, 0, 0), normalize.size = TRUE)}}{creates a RGL representation of the object}
+  \item{\code{applyTransformationMatrix(transformation.matrix)}}{Apply transformation matrix to polyhedron}
+  \item{\code{buildRGL(transformation.matrix)}}{creates a RGL representation of the object}
   \item{\code{exportToXML()}}{Gets an XML representation out of the polyhedron object}
 }
 }
