### 2018-05-18 c5519c3
configuration for FullDatabse scraping on first run
