+++ DESCRIPTION
@@ -1,4 +1,4 @@
-Package: Rpolyhedra
+Package: RpolyhedraFullDB
 Type: Package
 Title: Polyhedra Database
 Version: 0.2.4
