### 2018-05-18 6da0fa1
configuration for FullDatabse scraping on first run
