+++ R/zzz.R
@@ -42,9 +42,9 @@
     .polyhedra <- PolyhedronDatabase.class$new()
   }
   assign(".polyhedra", value = .polyhedra, envir = parent.env(environment()))
-  scrapePolyhedraSources(max.quant.config.schedule = 0,
-                         max.quant.scrape = 0,
-                         time2scrape.source = 80, #80 seconds of building scraping polyhedra
+  scrapePolyhedraSources(max.quant.config.schedule = 0, #All polyhedra
+                         max.quant.scrape = 0,          #All polyhedra
+                         time2scrape.source = 0,        #All polyhedra
                          sources.config = .available.sources,
                          retry.scrape = FALSE)
 }
