+++ man/RpolyhedraFullDB.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/zzz.R
 \docType{package}
-\name{Rpolyhedra}
-\alias{Rpolyhedra}
+\name{RpolyhedraFullDB}
+\alias{RpolyhedraFullDB}
 \alias{.onLoad}
-\alias{Rpolyhedra-package}
+\alias{RpolyhedraFullDB-package}
 \title{Rpolyhedra}
 \usage{
 .onLoad(libname, pkgname)
