### 2018-05-18 7c3924f
It is seemed not needed to change Package name in DESCRIPTION
