### 2018-05-22 e36d0d6
checkVertices
