### 2018-05-22 3133091
removed debug code
