### 2018-05-23 56ab224
Solved Directory problem for RpolyhedraFullDB
