+++ NAMESPACE
@@ -2,8 +2,6 @@
 
 export(getAvailablePolyhedra)
 export(getAvailableSources)
-export(getDataDir)
-export(getPolyhedraRDSPath)
 export(getPolyhedron)
 export(polyhedronToXML)
 export(scrapePolyhedraSources)
