### 2018-05-23 2a9a695
changes to db-lib.R
