### 2018-05-23 a6d6546
now it knows about the real package name
