### 2018-05-23 42f61f6
now it knows about the real package name
