### 2018-05-28 3d38dd0
supporting files removed to an external project and now they will be living on the home
