+++ inst/extdata/dmccooey.com/polyhedra/GeodesicIcosahedron3.txt
@@ -1,252 +1,0 @@
-Geodesic Icosahedron Pattern 3 [2,1] (Pentakis Snub Dodecahedron)
-
-C0  = 0.0919831947610306166536978645902
-C1  = 0.157802827551188752848878549936
-C2  = 0.1787372607291718755716205007246
-C3  = 0.2707204554902024922253183653148
-C4  = 0.306634763068337552375856521956
-C5  = 0.347313533259692355979268469051
-C6  = 0.404162274015810538852548576480
-C7  = 0.5260507939888642315508889697754
-C8  = 0.5335632139999518363605442871496
-C9  = 0.595837725984189461147451423520
-C10 = 0.6748827295060130310778669417953
-C11 = 0.693365236931662447624143478044
-C12 = 0.785348431692693064277841342634
-C13 = 0.832685557057201783926745491732
-C14 = 0.863323415398555806964420242008
-C15 = 0.943151259243881817126719892570
-
-C0  = phi * (3 - (x^2))
-C1  = x * phi - phi - 1
-C2  = (phi + 1 - x) / (x^3)
-C3  = x * phi * (x - phi)
-C4  = (phi + 1 - x) / (x^2)
-C5  = (phi^2) * (x - 1 - 1 / x)
-C6  = ((x^2) + x - (phi + 1)) / (x^3)
-C7  = (x^2) * phi - 2 * phi - 1
-C8  = (x * (23*phi+290) + 4 * (290*phi-267) + (4/x) * (21*phi-118)) / 2201
-C9  = (phi + 1) * ((x^2) - x - 1)
-C10 = phi + 2 - (x^2)
-C11 = ((x^2) + x - (phi + 1)) / (x^2)
-C12 = phi * (phi - x + 1 / x)
-C13 = x * phi + 1 - (x^2)
-C14 = (x * (313*phi + 23) + 4 * (23*phi + 290) - (4/x) * (97*phi - 21)) / 2201
-C15 = phi / x
-WHERE:  phi = (1 + sqrt(5)) / 2
-        x = cbrt((phi + sqrt(phi-5/27))/2) + cbrt((phi - sqrt(phi-5/27))/2)
-
-V0  = (  C2,  -C1,  1.0)
-V1  = (  C2,   C1, -1.0)
-V2  = ( -C2,   C1,  1.0)
-V3  = ( -C2,  -C1, -1.0)
-V4  = ( 1.0,  -C2,   C1)
-V5  = ( 1.0,   C2,  -C1)
-V6  = (-1.0,   C2,   C1)
-V7  = (-1.0,  -C2,  -C1)
-V8  = (  C1, -1.0,   C2)
-V9  = (  C1,  1.0,  -C2)
-V10 = ( -C1,  1.0,   C2)
-V11 = ( -C1, -1.0,  -C2)
-V12 = (  C3,   C4,  C15)
-V13 = (  C3,  -C4, -C15)
-V14 = ( -C3,  -C4,  C15)
-V15 = ( -C3,   C4, -C15)
-V16 = ( C15,   C3,   C4)
-V17 = ( C15,  -C3,  -C4)
[…]
