+++ inst/extdata/dmccooey.com/polyhedra/DualGeodesicIcosahedron13.txt
@@ -1,1089 +1,0 @@
-Dual Geodesic Icosahedron Pattern 13 [5,1]
-
-C0   = 0.0110695394112015749404588628354
-C1   = 0.0161126703275763129164581298316
-C2   = 0.02051220906909919091160251673348
-C3   = 0.0233756106644306655745083769091
-C4   = 0.0364364261181300363664056060852
-C5   = 0.0369166542453332048930280144377
-C6   = 0.04009013326463827983864447596147
-C7   = 0.0574863677901060120445312292230
-C8   = 0.0575508904696384399300720159937
-C9   = 0.0608638633196675702191367340081
-C10  = 0.0629875024848735813250490626826
-C11  = 0.0758450716474535376498707212481
-C12  = 0.0807276286945397211195814691186
-C13  = 0.0813978477130066706547253343978
-C14  = 0.08348156559974550576017669094853
-C15  = 0.0858032495594176011654336396613
-C16  = 0.102251697716096960142540391102
-C17  = 0.109815538949832982323277047956
-C18  = 0.1227639067851961510541429078351
-C19  = 0.141731778197124618977100371746
-C20  = 0.141996080761873503072105696594
-C21  = 0.142436941487454295837669749728
-C22  = 0.150845309874856877951493548754
-C23  = 0.1515976027284089400214343170152
-C24  = 0.161222672687516567423628612304
-C25  = 0.168756200881987524700742761701
-C26  = 0.171512436692572264799923683040
-C27  = 0.177933082777971807021505099322
-C28  = 0.184736111870261759183237962886
-C29  = 0.193411898712717034763599861011
-C30  = 0.1940313864621510185110111357346
-C31  = 0.201628428924336586236930985958
-C32  = 0.202830435924001690982260544177
-C33  = 0.222723709456413224191687165713
-C34  = 0.223129625910131289631825706144
-C35  = 0.233141310148288157567629856134
-C36  = 0.235171642303823711278286593738
-C37  = 0.241718562188974866075575461919
-C38  = 0.255142377832230150852522902021
-C39  = 0.258279096948414554467369359053
-C40  = 0.259673796433696974202770622487
-C41  = 0.261983729428159222661313252133
-C42  = 0.270700171487886407892738860737
-C43  = 0.271052379640669201172426285510
-C44  = 0.275662526581753593474637208972
-C45  = 0.295159026816619720134171953817
-C46  = 0.297496328996530303782679477871
-C47  = 0.315356158929208330836570304854
-C48  = 0.3205376597533645444219073564947
-C49  = 0.339155755465488374445133711626
-C50  = 0.345452666127445383497005665721
-C51  = 0.352645394606725732178703183040
-C52  = 0.353037159210466476338860460847
-C53  = 0.357597147503827145838016177354
-C54  = 0.358360192316197874001816211879
-C55  = 0.366044264707588624839135141440
-C56  = 0.377814224181525716203985482928
[…]
