+++ inst/extdata/dmccooey.com/polyhedra/OverarchingHexagonalToroid8.txt
@@ -1,28 +1,0 @@
-Hexagonal Toroid with 8 overarching faces
-
-V0  = ( 4, -4,  4)
-V1  = ( 4,  4, -4)
-V2  = (-4,  4,  4)
-V3  = (-4, -4, -4)
-V4  = ( 3, -3,  4)
-V5  = ( 3,  3, -4)
-V6  = (-3,  3,  4)
-V7  = (-3, -3, -4)
-V8  = ( 3, -3,  2)
-V9  = ( 3,  3, -2)
-V10 = (-3,  3,  2)
-V11 = (-3, -3, -2)
-V12 = ( 2, -2,  2)
-V13 = ( 2,  2, -2)
-V14 = (-2,  2,  2)
-V15 = (-2, -2, -2)
-
-Faces:
-{  0,  1,  2,  6,  9,  4 }
-{  0,  4, 11,  6,  2,  3 }
-{  0,  3,  7,  8,  5,  1 }
-{ 10,  5,  8, 12, 13, 14 }
-{ 10, 14, 15, 12,  8,  7 }
-{ 10,  7,  3,  2,  1,  5 }
-{  9,  6, 11, 15, 14, 13 }
-{  9, 13, 12, 15, 11,  4 }

[…]
