### 2018-05-28 f6fd348
supporting files removed to an external project and now they will be living on the home
