+++ R/polyhedra-lib.R
@@ -1,9 +1,3 @@
-
-
-
-
-
-
 #' Polyhedron State
 #'
 #' This abstract class provide the basis from which polyhedron state class derivate.
@@ -624,6 +618,9 @@ buildRGL = function(size = 1, origin = c(0, 0, 0), normalize.size = TRUE) {
         }
         self$calculateMassCenter(size, vertices.3d = sort(unique(unlist(triangulated.solid))))
         positioned.vertices     <- self$getPositionedVertices(size,origin)
+        positioned.vertices <- checkVertices(vertices = self$getVertices()[,1:3],
+                                             positioned.vertices = positioned.vertices,
+                                             triangulated.solid)
         vertices <- as.matrix(cbind(positioned.vertices, 1))
         ret <- rgl::tmesh3d(c(t(vertices)), unlist(triangulated.solid))
     } else {
