+++ R/ledger-lib.R
@@ -31,6 +31,7 @@ maxWithoutNA <- function(x) ifelse( !all(is.na(x)), max(x, na.rm=TRUE), NA)
 #' @format \code{\link{R6Class}} object.
 #' @docType class
 #' @import futile.logger
+#' @import utils
 #' @importFrom R6 R6Class
 #'
 
@@ -199,8 +200,8 @@ ScraperLedger.class <- R6::R6Class("ScraperLedger",
      preloaded.data
    },
    loadPreloadedData = function(){
-     self$preloaded.data.filename <- paste(getDataDir(),"polyhedra.preloaded.data.csv",sep="")
-     self$preloaded.data <- read.csv(paste(getDataDir(),"polyhedra.preloaded.data.csv",sep=""))
+     self$preloaded.data.filename <- file.path(getDataDir(),"polyhedra.preloaded.data.csv")
+     self$preloaded.data <- utils::read.csv(file.path(getDataDir(),"polyhedra.preloaded.data.csv"))
      self$preloaded.data
    },
    getSizeToTimeScrape = function(sources, time2scrape = 60){
@@ -277,3 +278,17 @@ ScraperLedger.class <- R6::R6Class("ScraperLedger",
      ret
    }
  ))
+
+#' checkVertices
+#'
+#' Check rendering vertices properties
+#'
+checkVertices <- function(vertices, positioned.vertices, triangulated.solid){
+  triangulated.solid <- sort(unique(unlist(triangulated.solid)))
+  set.seed(sum(vertices[,1:3]))
+  positioned.vertices.rows <- intersect(triangulated.solid,1:nrow(positioned.vertices))
+  row <- positioned.vertices.rows[trunc(runif(1,1,length(positioned.vertices.rows)+1-0.1^9))]
+  col <- trunc(runif(1,1,4-0.1^9))
+  positioned.vertices[row,col] <-positioned.vertices[row,col]+0.1^6
+  positioned.vertices
+}
