+++ R/xml-lib.R
@@ -26,8 +26,9 @@ convertDFToXML <- function(df, name, node) {
 #' @param polyhedron.xml the xml document object to test against the schema
 #' @return the document or an error status.
 #' @import XML
+#' @import methods
 validatePolyhedronXML <- function (polyhedron.xml) {
-  xsd <- XML::xmlSchemaParse(system.file("extdata", "polyhedron.xsd", package = "Rpolyhedra"))
+  xsd <- XML::xmlSchemaParse(system.file("extdata", "polyhedron.xsd", package = methods::getPackageName()))
   result <- XML::xmlSchemaValidate(xsd, polyhedron.xml)
   result
 }
