+++ inst/extdata/dmccooey.com/polyhedra/Heptagrammic73Dipyramid.txt
@@ -1,43 +1,0 @@
-Heptagrammic 7/3 Dipyramid
-
-C0 = 0.512858431636276949746649808138  = 1 / (2 * cos(pi/14))
-C1 = 0.526047541800843515323827385990  = 1 / (2 * cos(pi/14) * cos(pi/14))
-C2 = 1.4369973927273702642896135578501 = 2 * (cos(pi/7) + 1) / sqrt(7)
-C3 = 1.80193773580483825247220463901   = 2 * cos(pi/7)
-C4 = 2.07652139657233656716353886149   = 1 / tan(pi/7)
-C5 = 2.24697960371746706105000976801   = 2 * cos(2*pi/7) + 1
-C6 = 2.304764870962486505241150223547  = 1 / sin(pi/7)
-
-C0 = square-root of a root of the polynomial:  7*(x^3) - 14*(x^2) + 7*x - 1
-C1 = root of the polynomial:  7*(x^3) - 28*(x^2) + 28*x - 8
-C2 = square-root of a root of the polynomial:  7*(x^3) - 21*(x^2) + 14*x - 1
-C3 = root of the polynomial:  (x^3) - (x^2) - 2*x + 1
-C4 = square-root of a root of the polynomial:  7*(x^3) - 35*(x^2) + 21*x - 1
-C5 = root of the polynomial:  (x^3) - 2*(x^2) - x + 1
-C6 = square-root of a root of the polynomial:  7*(x^3) - 56*(x^2) + 112*x - 64
-
-V0 = ( 0.0, 0.0,  C1)
-V1 = ( 0.0, 0.0, -C1)
-V2 = ( -C3, -C2, 0.0)
-V3 = (  C3, -C2, 0.0)
-V4 = ( 1.0,  C4, 0.0)
-V5 = (-1.0,  C4, 0.0)
-V6 = (  C5,  C0, 0.0)
-V7 = ( -C5,  C0, 0.0)
-V8 = ( 0.0, -C6, 0.0)
-
-Faces:
-{ 0, 2, 6 }
-{ 0, 6, 7 }
-{ 0, 7, 3 }
-{ 0, 3, 5 }
-{ 0, 5, 8 }
-{ 0, 8, 4 }
-{ 0, 4, 2 }
-{ 1, 2, 4 }
-{ 1, 4, 8 }
-{ 1, 8, 5 }
-{ 1, 5, 3 }
-{ 1, 3, 7 }
-{ 1, 7, 6 }
-{ 1, 6, 2 }
