+++ inst/extdata/dmccooey.com/polyhedra/HexagonalAntiprism.txt
@@ -1,33 +1,0 @@
-Hexagonal Antiprism
-
-C0 = 0.427799838583676096484617883106 = sqrt(sqrt(3) - 1) / 2
-C1 = 0.866025403784438646763723170753 = sqrt(3) / 2
-
-V0  = ( 1.0,  0.0,  C0)
-V1  = (-1.0,  0.0,  C0)
-V2  = ( 0.5,   C1,  C0)
-V3  = ( 0.5,  -C1,  C0)
-V4  = (-0.5,   C1,  C0)
-V5  = (-0.5,  -C1,  C0)
-V6  = (  C1,  0.5, -C0)
-V7  = (  C1, -0.5, -C0)
-V8  = ( -C1,  0.5, -C0)
-V9  = ( -C1, -0.5, -C0)
-V10 = ( 0.0,  1.0, -C0)
-V11 = ( 0.0, -1.0, -C0)
-
-Faces:
-{  0,  2,  4,  1,  5,  3 }
-{  6,  7, 11,  9,  8, 10 }
-{  0,  3,  7 }
-{  0,  7,  6 }
-{  1,  4,  8 }
-{  1,  8,  9 }
-{  2,  0,  6 }
-{  2,  6, 10 }
-{  3,  5, 11 }
-{  3, 11,  7 }
-{  4,  2, 10 }
-{  4, 10,  8 }
-{  5,  1,  9 }
-{  5,  9, 11 }
