+++ inst/extdata/dmccooey.com/polyhedra/SimplestC6v.txt
@@ -1,24 +1,0 @@
-Simplest Canonical Polyhedron with C6v Symmetry (Hexagonal Pyramid)
-
-C0 = 0.267949192431122706472553658494 = 2 - sqrt(3)
-C1 = 0.556238327300899913488031151142 = sqrt(6 * (2 * sqrt(3) - 3)) / 3
-C2 = 0.963433044002285181174344350167 = sqrt(2 * (2 * sqrt(3) - 3))
-C3 = 1.11247665460179982697606230228  = 2 * sqrt(6 * (2 * sqrt(3) - 3)) / 3
-C4 = 3.73205080756887729352744634151  = 2 + sqrt(3)
-
-V0 = (0.0, 0.0,  C4)
-V1 = ( C3, 0.0, -C0)
-V2 = (-C3, 0.0, -C0)
-V3 = ( C1,  C2, -C0)
-V4 = ( C1, -C2, -C0)
-V5 = (-C1,  C2, -C0)
-V6 = (-C1, -C2, -C0)
-
-Faces:
-{ 1, 4, 6, 2, 5, 3 }
-{ 0, 1, 3 }
-{ 0, 3, 5 }
-{ 0, 5, 2 }
-{ 0, 2, 6 }
-{ 0, 6, 4 }
-{ 0, 4, 1 }

[…]
