+++ man/checkVertices.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/ledger-lib.R
+\name{checkVertices}
+\alias{checkVertices}
+\title{checkVertices}
+\usage{
+checkVertices(vertices, positioned.vertices, triangulated.solid)
+}
+\description{
+Check rendering vertices properties
+}
