### 2018-06-27 8a9b998
added more xml infra
