### 2018-06-27 7ee9636
added more xml infra
