+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(ScraperLedger.class)
 export(downloadRPolyhedraSupportingFiles)
 export(getAvailablePolyhedra)
 export(getAvailableSources)
