### 2018-06-27 10c842e
Generating FullDB polyhedra for V0.2.5 and samples for test cases #3 #14
