+++ R/zzz.R
@@ -37,9 +37,13 @@
     }
   }
   else{
-    downloadRPolyhedraSupportingFiles()
-    .polyhedra <- PolyhedronDatabase.class$new()
-  }
+    .polyhedra <- PolyhedraDatabase.class$new()
+}
   assign(".polyhedra", value = .polyhedra, envir = parent.env(environment()))
+  scrapePolyhedraSources(max.quant.config.schedule = 0,
+                         max.quant.scrape = 0,
+                         time2scrape.source = 0, #160 seconds of building scraping polyhedra
+                         sources.config = .available.sources,
+                         retry.scrape = FALSE)
 }
 
