### 2018-06-27 097d0e6
finnishing #3 compabilities issues and dev coding and documentation
