+++ man/PolyhedraDatabase.class.Rd
@@ -1,12 +1,12 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/db-lib.R
 \docType{class}
-\name{PolyhedronDatabase.class}
-\alias{PolyhedronDatabase.class}
-\title{PolyhedronDatabase}
+\name{PolyhedraDatabase.class}
+\alias{PolyhedraDatabase.class}
+\title{PolyhedraDatabase}
 \format{\code{\link{R6Class}} object.}
 \usage{
-PolyhedronDatabase.class
+PolyhedraDatabase.class
 }
 \description{
 Scrapes all polyhedra in data folder to save a representation which is accesible by the final users upon call to \code{getPolyhedron()}.
