### 2018-06-27 f27136e
finnishing #3 compabilities issues and dev coding and documentation
