+++ man/isCompatiblePolyhedraRDS.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/db-lib.R
-\name{compatiblePolyhedraRDS}
-\alias{compatiblePolyhedraRDS}
-\title{compatiblePolyhedraRDS()}
+\name{isCompatiblePolyhedraRDS}
+\alias{isCompatiblePolyhedraRDS}
+\title{isCompatiblePolyhedraRDS()}
 \usage{
-compatiblePolyhedraRDS(.polyhedra = .polyhedra)
+isCompatiblePolyhedraRDS(.polyhedra.candidate = .polyhedra)
 }
 \arguments{
 \item{.polyhedra}{current polyhedra database}
