### 2018-06-27 285fd0c
bugfix XML testcase. But generated an issue #15
