### 2018-06-27 8c220ff
bugfix XML testcase. But generated an issue #15
