### 2018-06-27 4b8a30b
[bugfix] Test first for preloaded data csv than RDS file
