+++ NAMESPACE
@@ -6,6 +6,7 @@ export(getAvailablePolyhedra)
 export(getAvailableSources)
 export(getPolyhedraRDSPath)
 export(getPolyhedron)
+export(getPreloadedDataFilename)
 export(polyhedronToXML)
 export(scrapePolyhedraSources)
 import(R6)
