### 2018-06-27 0ff3cac
Corrected Observations from devtools::check
