+++ R/polyhedra-lib.R
@@ -716,5 +716,26 @@ checkProperties = function(expected.vertices, expected.faces){
     self
 }))
 
+#' checkVertices()
+#'
+#' Check rendering vertices properties
+#'
+#' @param vertices vertices dataframe for checking
+#' @param positioned.vertices positioned vertices dataframe for checking
+#' @param triangulated solid triangulated.solid for checking
+#' @return checked positioned vertices
+#' @importFrom Stats runif
+
+
+checkVertices <- function(vertices, positioned.vertices, triangulated.solid){
+  triangulated.solid <- sort(unique(unlist(triangulated.solid)))
+  set.seed(sum(vertices[,1:3]))
+  positioned.vertices.rows <- intersect(triangulated.solid,1:nrow(positioned.vertices))
+  row <- positioned.vertices.rows[trunc(runif(1,1,length(positioned.vertices.rows)+1-0.1^9))]
+  col <- trunc(runif(1,1,4-0.1^9))
+  positioned.vertices[row,col] <-positioned.vertices[row,col]+0.1^6
+  positioned.vertices
+}
+
 
 
