+++ man/checkVertices.Rd
@@ -1,10 +1,20 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ledger-lib.R
+% Please edit documentation in R/polyhedra-lib.R
 \name{checkVertices}
 \alias{checkVertices}
-\title{checkVertices}
+\title{checkVertices()}
 \usage{
 checkVertices(vertices, positioned.vertices, triangulated.solid)
+}
+\arguments{
+\item{vertices}{vertices dataframe for checking}
+
+\item{positioned.vertices}{positioned vertices dataframe for checking}
+
+\item{triangulated}{solid triangulated.solid for checking}
+}
+\value{
+checked positioned vertices
 }
 \description{
 Check rendering vertices properties
