### 2018-06-27 3607f70
Corrected Observations from devtools::check
