+++ NAMESPACE
@@ -12,9 +12,9 @@ export(scrapePolyhedraSources)
 import(R6)
 import(XML)
 import(futile.logger)
-import(methods)
 import(rgl)
 import(stringr)
 import(testthat)
 import(utils)
 importFrom(R6,R6Class)
+importFrom(Stats,runif)
