### 2018-06-27 c499f7f
Corrected Observations from devtools::check
