+++ NAMESPACE
@@ -17,4 +17,4 @@ import(stringr)
 import(testthat)
 import(utils)
 importFrom(R6,R6Class)
-importFrom(Stats,runif)
+importFrom(stats,runif)
