### 2018-06-27 304ddf5
Corrected Observations from devtools::check
