### 2018-06-27 a3fd847
Removed example for avoding error in CHECK. #15
