### 2018-06-28 7d4136e
now it downloads the full database from an external repository. Also the XML export function has been disabled and pushed to v0.2.6
