+++ man/getDatabaseVersion.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/db-lib.R
+\name{getDatabaseVersion}
+\alias{getDatabaseVersion}
+\title{getDatabaseVersion}
+\usage{
+getDatabaseVersion()
+}
+\description{
+Obtains the code version from the database version file
+}
