### 2018-06-29 c95cbdd
changed the package version identification procedure
