### 2018-06-29 7ca9bfd
changed the package version identification procedure
