### 2018-06-29 350e301
changed the package version identification procedure
