### 2018-06-29 e2a87f8
changed the package version identification procedure
