### 2018-06-29 fe47db5
[bugfix] For DB regeneration
