### 2018-06-29 4596782
[bugfix] For DB regeneration
