+++ R/ledger-lib.R
@@ -202,7 +202,7 @@ ScraperLedger.class <- R6::R6Class("ScraperLedger",
    },
    loadPreloadedData = function(){
      self$preloaded.data.filename <- getPreloadedDataFilename()
-     self$preloaded.data <- utils::read.csv(preloaded.data.filename)
+     self$preloaded.data <- utils::read.csv(self$preloaded.data.filename)
      self$preloaded.data
    },
    getSizeToTimeScrape = function(sources, time2scrape = 60){
