+++ man/checkPackageVersion.Rd
@@ -1,11 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/db-lib.R
-\name{checkPackageVersion}
-\alias{checkPackageVersion}
-\title{checkPackageVersion}
-\usage{
-checkPackageVersion()
-}
-\description{
-Obtains code version from configuration text file
-}
