### 2018-06-29 5aaa634
[bugfix] in package initialization
