+++ R/zzz.R
@@ -30,13 +30,14 @@
   if (!file.exists(getPreloadedDataFilename())){
     downloadRPolyhedraSupportingFiles()
   }
+  .polyhedra <- NULL
   if (file.exists(polyhedra.rds.file)) {
     polyhedra.candidate <- readRDS(polyhedra.rds.file)
     if (isCompatiblePolyhedraRDS(polyhedra.candidate, halts = TRUE)){
       .polyhedra <- polyhedra.candidate
     }
   }
-  else{
+  if (is.null(.polyhedra)){
     .polyhedra <- PolyhedraDatabase.class$new()
   }
   assign(".polyhedra", value = .polyhedra, envir = parent.env(environment()))
