### 2018-06-29 cf2c5be
[bugfix] isCompatiblePolyhedraRDS
