### 2018-06-30 0cf9bb9
added a message banner to ask users to download supporting data
