### 2018-06-30 9003fa8
added a message banner to ask users to download supporting data
