+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(ScraperLedger.class)
+export(checkDownloadFilesUserAcceptance)
 export(downloadRPolyhedraSupportingFiles)
 export(getAvailablePolyhedra)
 export(getAvailableSources)
