### 2018-06-30 a07a7e5
added travis awareness for the user download message
