+++ .travis.yml
@@ -7,5 +7,8 @@ before_install:
   - tlmgr install index
 r_packages:
   - covr
+env:
+  - ON_TRAVIS = 1
 after_success:
   - Rscript -e 'library(covr); codecov()'
+
