### 2018-06-30 c8bb580
added travis awareness for the user download message
