### 2018-07-02 c422a2e
generating pkg-minimal
