+++ R/zzz.R
@@ -19,18 +19,36 @@
 #' @param libname The library name
 #' @param pkgname The package name
 .onLoad <- function(libname, pkgname) {
-  #debug settings to run the logger.
-  polyhedra.rds.file <- getPolyhedraRDSPath()
 
   #setup Available sources
   .available.sources <- list()
   .available.sources[["netlib"]] <- PolyhedronScraperConfigurationNetlib.class$new()
   .available.sources[["dmccooey"]] <- PolyhedronScraperConfigurationDmccoey.class$new()
   assign(".available.sources", value = .available.sources, envir = parent.env(environment()))
+  .available.scrapping.conf <- list()
+  .available.scrapping.conf[["dev-minimal"]] <- list(max.quant.config.schedule = 0,
+                                                     max.quant.scrape = 10,
+                                                     time2scrape.source = 20,
+                                                      #20 seconds of building/scraping polyhedra database for reasonable devs timing
+                                                     retry.scrape = FALSE)
+  .available.scrapping.conf[["pkg-minimal"]] <- list(max.quant.config.schedule = 0,
+                                                     max.quant.scrape = 0,
+                                                     time2scrape.source = 80,
+                                                     retry.scrape = FALSE)
+  .available.scrapping.conf[["fulldb"]] <- list(max.quant.config.schedule = 0,
+                                                     max.quant.scrape = 0,
+                                                     time2scrape.source = 0,
+                                                     retry.scrape = FALSE)
+  assign(".available.scrapping.conf", value = .available.sources, envir = parent.env(environment()))
+  if(!exists(".data.env"))
+  {
+    selectDataEnv()
+  }
   if (!file.exists(getPreloadedDataFilename())){
     downloadRPolyhedraSupportingFiles()
   }
   .polyhedra <- NULL
+  polyhedra.rds.file <- getPolyhedraRDSPath()
   if (file.exists(polyhedra.rds.file)) {
     polyhedra.candidate <- readRDS(polyhedra.rds.file)
     if (isCompatiblePolyhedraRDS(polyhedra.candidate, halts = TRUE)){
@@ -40,12 +58,9 @@
   if (is.null(.polyhedra)){
     .polyhedra <- PolyhedraDatabase.class$new()
   }
+
   assign(".polyhedra", value = .polyhedra, envir = parent.env(environment()))
-  scrapePolyhedraSources(max.quant.config.schedule = 0,
-                         max.quant.scrape = 0,
-                         time2scrape.source = 0,
-                         #30 seconds of building/scraping polyhedra database for reasonable devs timing
-                         sources.config = .available.sources,
-                         retry.scrape = FALSE)
+  scrapePolyhedra(.available.scrapping.conf[["pkg-minimal"]],
+                  sources.config = .available.sources)
 }
 
