+++ man/selectDataEnv.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/db-lib.R
+\name{selectDataEnv}
+\alias{selectDataEnv}
+\title{selectDataEnv}
+\usage{
+selectDataEnv()
+}
+\value{
+.data.env
+}
+\description{
+Asks the user to download the supporting files
+}
