+++ man/checkDownloadFilesUserAcceptance.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/db-lib.R
-\name{checkDownloadFilesUserAcceptance}
-\alias{checkDownloadFilesUserAcceptance}
-\title{checkDownloadFilesUserAcceptance}
-\usage{
-checkDownloadFilesUserAcceptance()
-}
-\value{
-TRUE if user accepts the download, FALSE otherwise
-}
-\description{
-Asks the user to download the supporting files
-}
