+++ NAMESPACE
@@ -1,7 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(ScraperLedger.class)
-export(checkDownloadFilesUserAcceptance)
+export(askDataDirEnvironment)
 export(downloadRPolyhedraSupportingFiles)
 export(getAvailablePolyhedra)
 export(getAvailableSources)
@@ -9,7 +9,10 @@ export(getPolyhedraRDSPath)
 export(getPolyhedron)
 export(getPreloadedDataFilename)
 export(polyhedronToXML)
+export(scrapePolyhedra)
 export(scrapePolyhedraSources)
+export(selectDataEnv)
+export(setDataDirEnvironment)
 import(R6)
 import(XML)
 import(futile.logger)
