### 2018-07-02 a2a76ba
generating pkg-minimal
