+++ R/db-lib.R
@@ -4,12 +4,57 @@
 #'
 #' This function is used internally to determine whether the package
 #' is compiled in source or package directory.
-getDataDir <- function() {
-  home.dir <- file.path(path.expand("~"), ".R", "Rpolyhedra")
-  if(dir.exists(home.dir) == FALSE) {
-    dir.create(home.dir, recursive=TRUE, showWarnings = FALSE)
+getDataDir <- function(data.env=.data.env) {
+  data.dir <- ""
+  if(data.env == "HOME")
+  {
+    data.dir <- file.path(path.expand("~"), ".R", "Rpolyhedra")
+    if(dir.exists(data.dir) == FALSE) {
+      dir.create(data.dir, recursive=TRUE, showWarnings = FALSE)
+    }
   }
-  home.dir
+  else{
+    data.dir <- getPackageDir()
+  }
+  data.dir
+}
+
+#' setDataDirEnv
+#'
+#' Sets the path of Rpolyhedra data dir.
+#'
+#' This function is used to set the data directories either to the package or the user home directory.
+#'
+#' @param env The type of environment to work with. Values are "PACKAGE" or "HOME" and it defaults to package
+#' @export
+setDataDirEnvironment <- function(env="PACKAGE") {
+  if(env=="PACKAGE")
+    .data.env <- "PACKAGE"
+  else if(env=="HOME")
+    .data.env <- "HOME"
+  else
+    stop("Possible values are PACKAGE and HOME")
+  assign(".data.env", value = .data.env, envir = parent.env(environment()))
+  .data.env
+}
+
+#' setDataDirEnv
+#'
+#' Sets the path of Rpolyhedra data dir.
+#'
+#' This function is used to set the data directories either to the package or the user home directory.
+#'
+#' @param env The type of environment to work with. Values are "PACKAGE" or "HOME" and it defaults to package
+#' @export
+askDataDirEnvironment <- function(env="PACKAGE") {
+  if(env=="PACKAGE")
+    .data.env <- "PACKAGE"
+  else if(env=="HOME")
+    .data.env <- "HOME"
+  else
+    stop("Possible values are PACKAGE and HOME")
+  assign(".data.env", value = .data.env, envir = parent.env(environment()))
+  .data.env
 }
 
 #' getPackageDir
@@ -18,10 +63,10 @@ getDataDir <- function() {
 
 getPackageDir <- function(){
   home.dir <- find.package("Rpolyhedra", lib.loc = NULL, quiet = TRUE)
-  data.subdir <- "inst/extdata/"
-  if (!dir.exists(file.path(home.dir, "/", data.subdir)))
+  data.subdir <- file.path("inst", "extdata/")
+  if (!dir.exists(file.path(home.dir, data.subdir)))
     data.subdir <- "extdata/"
-  file.path(home.dir, "/", data.subdir)
+  file.path(home.dir, data.subdir)
 }
 
 #' getPolyhedraRDSPath
@@ -47,32 +92,33 @@ getPreloadedDataFilename <- function(polyhedra_preloaded_data = "polyhedra.prelo
   file.path(getDataDir(),polyhedra_preloaded_data)
 }
 
-#' checkDownloadFilesUserAcceptance
+#' selectDataEnv
 #'
 #' Asks the user to download the supporting files
 #'
-#' @return TRUE if user accepts the download, FALSE otherwise
+#' @return .data.env
 #' @export
-checkDownloadFilesUserAcceptance <- function() {
+selectDataEnv <- function() {
   if(!is.na(Sys.getenv(x = "ON_TRAVIS", unset=NA)))
     return(TRUE)
-  #gettext("rpoly.already_defined", domain = "R-Rpolyhedra")
[…]
