### 2018-07-02 8ad3989
V0.2.5 with pkg-minimal db
