+++ R/db-lib.R
@@ -161,23 +161,45 @@ copyFilesToExtData <- function(force = FALSE){
   polyhedra.ledger.scraped <- polyhedra.ledger[polyhedra.ledger$status=="scraped",]
   data.env.home <- getDataDir(data.env ="HOME")
   data.env.package <- getDataDir(data.env ="PACKAGE")
-  #clean dirs
   dir.create(data.env.package,showWarnings = FALSE, recursive = TRUE)
+  #check existing sources
+  existing <- FALSE
   for (source in names(.available.sources)){
     source.config <- .available.sources[[source]]
-    source.dir <- source.config$getBaseDir(data.env.package)
-    file.remove(source.dir)
-    dir.create(source.dir, showWarnings = FALSE, recursive = TRUE)
+    dest.dir <- source.config$getBaseDir(data.env.package)
+    if (file.exists(dest.dir)){
+      existing <- TRUE
+    }
+  }
+  if (existing & !force){
+    stop(paste("Cannot copy files: they exists in destination. Call the function with force=TRUE or remove them manually"))
+  }
+  #clean dirs
+  for (source in names(.available.sources)){
+    source.config <- .available.sources[[source]]
+    dest.dir <- source.config$getBaseDir(data.env.package)
+    if (file.exists(dest.dir)){
+      unlink(dest.dir,recursive = TRUE)
+    }
+    dir.create(dest.dir, showWarnings = FALSE, recursive = TRUE)
   }
   #copy files
   #copy version
   file.copy(file.path(data.env.home,"version"),data.env.package,overwrite = TRUE)
+  file.copy(file.path(data.env.home,"polyhedra.preloaded.data.csv"),data.env.package,overwrite = TRUE)
 
   #copy polyhedra source files
+  cont <- 0
   for (i in c(1:nrow(polyhedra.ledger.scraped))){
     current.polyhedron <-polyhedra.ledger.scraped[i,]
-    #TODO copy polyhedra sources files
+    source.config <- .available.sources[[current.polyhedron$source]]
+    dest.dir <- source.config$getBaseDir(data.env.package)
+    if (file.copy(file.path(source.config$getBaseDir(data.env.home),current.polyhedron$filename),
+                  dest.dir)){
+        cont <- cont+1
+    }
   }
+  futile.logger::flog.info(paste("Copied",cont,"polyhedra sources files to",data.env.package))
   #copy RDS
   file.copy(file.path(data.env.home,"polyhedra.RDS"),data.env.package)
   self
@@ -287,18 +309,6 @@ PolyhedronScraperConfigurationDmccoey.class <- R6::R6Class("PolyhedronScraperCon
       polyhedra.dir   <- self$getBaseDir(home.dir.data)
       polyhedra.files <- dir(polyhedra.dir)
       polyhedra.files <- polyhedra.files[grep("\\.txt", polyhedra.files)]
-      geodesic.files  <- polyhedra.files[grep("geodesic",polyhedra.files,ignore.case = TRUE)]
-      geodesic.files  <- geodesic.files[-grep("Dual",geodesic.files,ignore.case = TRUE)]
-      regexp.natnum   <- "([0-9]+)"
-      geodesic.order  <- data.frame(filename = geodesic.files,
-                                    class= sub(regexp.natnum,"",geodesic.files),
-                                    order= as.numeric(str_extract(geodesic.files,regexp.natnum)),
-                                    stringsAsFactors = FALSE)
-      geodesic.order  <- geodesic.order[order(geodesic.order$order,geodesic.order$class),]
-      geodesic.order  <- geodesic.order[1:50,]
[…]
