+++ inst/extdata/dmccooey.com/polyhedra/45TruncatedDeltoidalHexecontahedron1.txt
@@ -1,0 +1,369 @@
+4-5-Truncated Deltoidal Hexecontahedron with truncation depth chosen
+to make the resulting heptagons have 5 equal (long) sides plus 2 equal
+(short) sides.
+
+C0  = 0.7071067811865475244008443621048
+C1  = 0.809016994374947424102293417183
+C2  = 0.839114039465866199798439681667
+C3  = 1.05762205169027673519809028420
+C4  = 1.65851990681992624152212694280
+C5  = 1.68035577065427768818890211041
+C6  = 1.71126842688626593458628975545003
+C7  = 1.76472883287682425959893464630
+C8  = 1.98323684510123479499858524883
+C9  = 2.06482181747953969678671193650
+C10 = 2.18035577065427768818890211041
+C11 = 2.50183784192836076758601314156
+C12 = 2.68354158025294918820491233722
+C13 = 2.85539123252163452978643532261
+C14 = 2.98937276502922511229119552760
+C15 = 3.34095188139422696738445282323
+C16 = 3.4759972597630901941852244017499
+C17 = 3.527889744485591856746494291496
+C18 = 3.91301328421191126498452560681
+C19 = 4.027889744485591856746494291496
+C20 = 4.04805866258077449178529718533
+C21 = 4.33690673886053928084878770868
+C22 = 4.34206148707287542972703928001
+C23 = 4.56665965940790046437272507806
+
+C0  = sqrt(2) / 2
+C1  = (1 + sqrt(5)) / 4
+C2  = (20 + 5 * sqrt(2) - 5 * sqrt(10) + 6 * sqrt(5 * (5 - sqrt(5)))) / 40
+C3  = (10 + 3 * sqrt(5 * (5 - sqrt(5)))) / 20
+C4  = (80 + 105 * sqrt(2) + 20 * sqrt(5) - 15 * sqrt(10)
+    + 6 * sqrt(5 * (65 + 19 * sqrt(5)))) / 220
+C5  = (5 * sqrt(2) + sqrt(10 * (5 + 2 * sqrt(5)))) / 10
+C6  = (5 + 5 * sqrt(5) + 3 * sqrt(5 * (5 + sqrt(5)))) / 20
+C7  = (10 + 10 * sqrt(2) + 3 * sqrt(5 * (5 - sqrt(5)))) / 20
+C8  = (20 + 15 * sqrt(2) + 5 * sqrt(10) + 6 * sqrt(5 * (5 - sqrt(5)))) / 40
+C9  = (5 + 5 * sqrt(2) + 5 * sqrt(5) + 3 * sqrt(5 * (5 + sqrt(5)))) / 20
+C10 = (5 + 5 * sqrt(2) + sqrt(10 * (5 + 2 * sqrt(5)))) / 10
+C11 = (5 + 5 * sqrt(5) + 5 * sqrt(10) + 3 * sqrt(5 * (5 + sqrt(5)))) / 20
+C12 = (90 + 15 * sqrt(2) + 50 * sqrt(5) + 45 * sqrt(10)
+    + 6 * sqrt(5 * (145 + 61 * sqrt(5)))) / 220
+C13 = (5 + 5 * sqrt(2) + 5 * sqrt(5) + 5 * sqrt(10)
+    + 3 * sqrt(5 * (5 + sqrt(5)))) / 20
[…]
