### 2018-07-02 b80cf92
V0.2.5 with pkg-minimal db
