+++ tests/testthat/test_rpolyhedra_lib.R
@@ -13,7 +13,7 @@ test_that("Scrape test rpolyhedra 5 regular solids", {
 context("scrape available polyhedra")
 test_that("Scrape test rpolyhedra for 12% of available polyhedra", {
   tasks <- .polyhedra$generateTestTasks(TestTaskClass = PolyhedronTestTaskScrape.class,
-                                        max.quant = getPercentilPolyhedraQuant(0.12))
+                                        max.quant = getPercentilPolyhedraQuant(0.12,50))
 
   for (task in tasks){
     task$run()
@@ -40,7 +40,7 @@ test_that("test getting properties of scraped regular solids", {
 context("Edges Consistency")
 test_that("test check edges consistency for 7% of available polyhedra", {
   tasks <- .polyhedra$generateTestTasks(TestTaskClass = PolyhedronTestTaskEdgesConsistency.class,
-                                        max.quant = getPercentilPolyhedraQuant(0.12))
+                                        max.quant = getPercentilPolyhedraQuant(0.12,50))
 
   for (task in tasks){
     task$run()
