+++ man/getPercentilPolyhedraQuant.Rd
@@ -5,10 +5,12 @@
 \title{getPercentilPolyhedraQuant
 returns polyhedra quantity of parameter percentil}
 \usage{
-getPercentilPolyhedraQuant(percentil)
+getPercentilPolyhedraQuant(percentil, quant.min = 100)
 }
 \arguments{
 \item{percentil}{is the percentil which must be applied to estimate the figure}
+
+\item{quant.min}{minimum quantity of files to return}
 }
 \description{
 getPercentilPolyhedraQuant
