+++ inst/extdata/dmccooey.com/polyhedra/6TruncatedTetrakisHexahedron2.txt
@@ -1,0 +1,95 @@
+6-Truncated Tetrakis Hexahedron (canonical)
+
+C0 = 0.908248290463863016366214012451 = (3 + sqrt(6)) / 6
+C1 = 1.724744871391589049098642037353 = (1 + sqrt(6)) / 2
+C2 = 2.13299316185545206546485604980  = (3 + 4 * sqrt(6)) / 6
+C3 = 2.58711730708738357364796305603  = 3 * (1 + sqrt(6)) / 4
+
+V0  = ( 0.0,  0.0,   C3)
+V1  = ( 0.0,  0.0,  -C3)
+V2  = (  C3,  0.0,  0.0)
+V3  = ( -C3,  0.0,  0.0)
+V4  = ( 0.0,   C3,  0.0)
+V5  = ( 0.0,  -C3,  0.0)
+V6  = (  C0,   C0,   C2)
+V7  = (  C0,   C0,  -C2)
+V8  = (  C0,  -C0,   C2)
+V9  = (  C0,  -C0,  -C2)
+V10 = ( -C0,   C0,   C2)
+V11 = ( -C0,   C0,  -C2)
+V12 = ( -C0,  -C0,   C2)
+V13 = ( -C0,  -C0,  -C2)
+V14 = (  C2,   C0,   C0)
+V15 = (  C2,   C0,  -C0)
+V16 = (  C2,  -C0,   C0)
+V17 = (  C2,  -C0,  -C0)
+V18 = ( -C2,   C0,   C0)
+V19 = ( -C2,   C0,  -C0)
+V20 = ( -C2,  -C0,   C0)
+V21 = ( -C2,  -C0,  -C0)
+V22 = (  C0,   C2,   C0)
+V23 = (  C0,   C2,  -C0)
+V24 = (  C0,  -C2,   C0)
+V25 = (  C0,  -C2,  -C0)
+V26 = ( -C0,   C2,   C0)
+V27 = ( -C0,   C2,  -C0)
+V28 = ( -C0,  -C2,   C0)
+V29 = ( -C0,  -C2,  -C0)
+V30 = (  C1,  0.5,   C1)
+V31 = (  C1,  0.5,  -C1)
+V32 = (  C1, -0.5,   C1)
+V33 = (  C1, -0.5,  -C1)
+V34 = ( -C1,  0.5,   C1)
+V35 = ( -C1,  0.5,  -C1)
+V36 = ( -C1, -0.5,   C1)
+V37 = ( -C1, -0.5,  -C1)
+V38 = (  C1,   C1,  0.5)
+V39 = (  C1,   C1, -0.5)
+V40 = (  C1,  -C1,  0.5)
+V41 = (  C1,  -C1, -0.5)
+V42 = ( -C1,   C1,  0.5)
+V43 = ( -C1,   C1, -0.5)
+V44 = ( -C1,  -C1,  0.5)
+V45 = ( -C1,  -C1, -0.5)
+V46 = ( 0.5,   C1,   C1)
+V47 = ( 0.5,   C1,  -C1)
+V48 = ( 0.5,  -C1,   C1)
+V49 = ( 0.5,  -C1,  -C1)
+V50 = (-0.5,   C1,   C1)
+V51 = (-0.5,   C1,  -C1)
+V52 = (-0.5,  -C1,   C1)
+V53 = (-0.5,  -C1,  -C1)
+
+Faces:
+{  6, 30, 14, 38, 22, 46 }
+{  7, 47, 23, 39, 15, 31 }
+{  8, 48, 24, 40, 16, 32 }
+{  9, 33, 17, 41, 25, 49 }
[…]
