+++ inst/extdata/dmccooey.com/polyhedra/10TruncatedTriakisIcosahedron2.txt
@@ -1,0 +1,250 @@
+10-Truncated Triakis Icosahedron (canonical)
+
+C0  = 0.762865556059566803012834542424
+C1  = 0.951056516295153572116439333379
+C2  = 1.39365488447020512065367684917
+C3  = 1.45105651629515357211643933338
+C4  = 1.53884176858762670128514528802
+C5  = 2.18539891484612096230950299909
+C6  = 2.25498097166009987658865538962
+C7  = 2.30170732464719350429797983044
+C8  = 2.34785876296257412538743870520
+C9  = 2.77318416713859409147820895373
+C10 = 2.98989828488278027340158462140
+C11 = 3.25276384094234707641441916382
+C12 = 3.29891527925772769750387803858
+C13 = 3.64863585613030499724233223878
+C14 = 3.72424068343374766359464828711
+C15 = 3.88670053155020082667258399322
+
+C0  = (10 + sqrt(10 * (5 - sqrt(5)))) / 20
+C1  = sqrt(2 * (5 + sqrt(5))) / 4
+C2  = (15 - sqrt(5) + 4 * sqrt(5 * (25 + 2 * sqrt(5)))) / 44
+C3  = (2 + sqrt(2 * (5 + sqrt(5)))) / 4
+C4  = sqrt(5 + 2 * sqrt(5)) / 2
+C5  = (5 + 5 * sqrt(5) + 4 * sqrt(5 * (5 + 2 * sqrt(5)))) / 20
+C6  = (5 + 7 * sqrt(5) + 2 * sqrt(10 * (85 + 31 * sqrt(5)))) / 44
+C7  = (10 + sqrt(10 * (65 + 29 * sqrt(5)))) / 20
+C8  = (1 + sqrt(5) + 2 * sqrt(5 + 2 * sqrt(5))) / 4
+C9  = (5 + 5 * sqrt(5) + sqrt(10 * (85 + 31 * sqrt(5)))) / 20
+C10 = (2 + sqrt(2 * (25 + 11 * sqrt(5)))) / 4
+C11 = (5 + 4 * sqrt(5 * (5 + 2 * sqrt(5)))) / 10
+C12 = (1 + sqrt(5) + sqrt(2 * (25 + 11 * sqrt(5)))) / 4
+C13 = (10 + 3 * sqrt(5) + sqrt(10 * (205 + 89 * sqrt(5)))) / 22
+C14 = (5 + 5 * sqrt(5) + 2 * sqrt(5 * (85 + 38 * sqrt(5)))) / 20
+C15 = (1 + sqrt(5) + 4 * sqrt(5 + 2 * sqrt(5))) / 4
+
+V0   = ( 0.5,  0.0,  C15)
+V1   = ( 0.5,  0.0, -C15)
+V2   = (-0.5,  0.0,  C15)
+V3   = (-0.5,  0.0, -C15)
+V4   = ( C15,  0.5,  0.0)
+V5   = ( C15, -0.5,  0.0)
+V6   = (-C15,  0.5,  0.0)
+V7   = (-C15, -0.5,  0.0)
+V8   = ( 0.0,  C15,  0.5)
+V9   = ( 0.0,  C15, -0.5)
+V10  = ( 0.0, -C15,  0.5)
+V11  = ( 0.0, -C15, -0.5)
+V12  = (  C0,   C1,  C14)
+V13  = (  C0,   C1, -C14)
+V14  = (  C0,  -C1,  C14)
+V15  = (  C0,  -C1, -C14)
+V16  = ( -C0,   C1,  C14)
+V17  = ( -C0,   C1, -C14)
[…]
