+++ README.md
@@ -20,6 +20,9 @@ Install the R package using the following commands on the R console:
 install.packages(c("futile.logger", "rgl", "stringr", "R6", "testthat", "devtools"))
 devtools::install_github("qbotics/Rpolyhedra")
 library(Rpolyhedra)
+# if want to switch to fullDB in user filespace, it will download the full database
+switchToFullDatabase()
+
 ```
 
 # A simple example of 5 regular polyhedra
