+++ NAMESPACE
@@ -13,6 +13,7 @@ export(scrapePolyhedra)
 export(scrapePolyhedraSources)
 export(selectDataEnv)
 export(setDataDirEnvironment)
+export(switchToFullDatabase)
 import(R6)
 import(XML)
 import(futile.logger)

[…]
