### 2018-07-02 56bfda2
V0.2.5 with pkg-minimal db
