+++ man/isCompatiblePolyhedraRDS.Rd
@@ -7,7 +7,9 @@
 isCompatiblePolyhedraRDS(.polyhedra.candidate = .polyhedra, halts = FALSE)
 }
 \arguments{
-\item{.polyhedra.candidate}{current polyhedra database candidate for checking compatibility}
+\item{.polyhedra.candidate}{polyhedra db to test}
+
+\item{halts}{indicates whether it has to halt execution when it is not compatible}
 }
 \description{
 Tests if the polyhedra RDS is compatible with the current format
