### 2018-07-02 e4055f7
V0.2.5 with pkg-minimal db
