### 2018-07-03 60b479c
Corrected v0.2.5 from fulldb branch
