### 2018-07-03 f737781
minor changes removing debug info
