+++ NAMESPACE
@@ -11,7 +11,6 @@ export(getPreloadedDataFilename)
 export(polyhedronToXML)
 export(scrapePolyhedra)
 export(scrapePolyhedraSources)
-export(selectDataEnv)
 export(setDataDirEnvironment)
 export(switchToFullDatabase)
 import(R6)
