### 2018-07-03 ee4d23e
minor changes removing debug info
