### 2018-07-03 c7b7a1c
environment changes to allow the modification of package wide variables
