### 2018-07-03 27105ab
environment changes to allow the modification of package wide variables
