### 2018-07-03 87af4ad
solved a problem with downloadRPolyhedraSupportingFiles not being called when selecting fullDB
