### 2018-07-03 1f11691
solved a problem with downloadRPolyhedraSupportingFiles not being called when selecting fullDB
