+++ man/getUserEnv.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/db-lib.R
+\name{getUserEnv}
+\alias{getUserEnv}
+\title{getUserEnv}
+\usage{
+getUserEnv()
+}
+\description{
+Gets a writable environment for the package
+}
