+++ man/getDataEnv.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/db-lib.R
+\name{getDataEnv}
+\alias{getDataEnv}
+\title{getDataEnv}
+\usage{
+getDataEnv()
+}
+\value{
+.data.env
+}
+\description{
+get the current .data.env value
+}
