### 2018-07-03 ef755b0
[bugfix] can change .data.env but have to access without exporting selectDataEnv
