+++ NAMESPACE
@@ -1,7 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
 export(ScraperLedger.class)
-export(askDataDirEnvironment)
 export(downloadRPolyhedraSupportingFiles)
 export(getAvailablePolyhedra)
 export(getAvailableSources)
@@ -11,6 +10,7 @@ export(getPreloadedDataFilename)
 export(polyhedronToXML)
 export(scrapePolyhedra)
 export(scrapePolyhedraSources)
+export(selectDataEnv)
 export(setDataDirEnvironment)
 export(switchToFullDatabase)
 import(R6)
