### 2018-07-03 420a56f
Solved the issue indicated in #17"
