### 2018-07-04 f0377fd
now it reloads the .polyhedra object after switchToFullDatabase()
