### 2018-07-04 e0b6805
now it reloads the .polyhedra object after switchToFullDatabase()
