### 2018-07-04 91ad6de
changes to better support memory management
