+++ R/db-lib.R
@@ -88,32 +88,39 @@ getPreloadedDataFilename <- function(polyhedra_preloaded_data = "polyhedra.prelo
 #'
 #' Asks the user where to set the system variable .data.env
 #'
+#' @param env The environment to run on, can be PACKAGE, HOME or NULL. If null, it asks the user for a an Environment.
 #' @return .data.env
 #' @import futile.logger
-selectDataEnv <- function() {
-  if(!is.na(Sys.getenv(x = "ON_TRAVIS", unset=NA)))
-    return(TRUE)
-  accept.option <- readline(prompt="Full Database needs to download data to home folder. Agree [y/n]?:")
-  if(tolower(accept.option[1]) == "n") {
-    futile.logger::flog.info("Working on demo DB. You can call switchToFullDatabase to use the full database.")
-    setDataDirEnvironment("PACKAGE")
-  }
-  if(tolower(accept.option[1]) == "y") {
-    setDataDirEnvironment("HOME")
-  }
-  while(!tolower(accept.option[1]) %in% c("n","y")) {
-    accept.option <- readline(prompt="Unknown option. Agree [y/n]?:")
+selectDataEnv <- function(env=NA) {
+  if(is.na(env)) {
+    if(!is.na(Sys.getenv(x = "ON_TRAVIS", unset=NA)))
+      return(TRUE)
+    accept.option <- readline(prompt="Full Database needs to download data to home folder. Agree [y/n]?:")
     if(tolower(accept.option[1]) == "n") {
-      futile.logger::flog.info("Working on demo DB. You can call selectDataEnv to use the full database.")
+      futile.logger::flog.info("Working on demo DB. You can call switchToFullDatabase to use the full database.")
       setDataDirEnvironment("PACKAGE")
-    } else if(tolower(accept.option[1]) == "y") {
+    }
+    if(tolower(accept.option[1]) == "y") {
       setDataDirEnvironment("HOME")
     }
+    while(!tolower(accept.option[1]) %in% c("n","y")) {
+      accept.option <- readline(prompt="Unknown option. Agree [y/n]?:")
+      if(tolower(accept.option[1]) == "n") {
+        futile.logger::flog.info("Working on demo DB. You can call selectDataEnv to use the full database.")
+        setDataDirEnvironment("PACKAGE")
+      } else if(tolower(accept.option[1]) == "y") {
+        setDataDirEnvironment("HOME")
+      }
+    }
+  } else {
+    Rpolyhedra::setDataDirEnvironment(env)
   }
+  #loads the database
   data.env <-getDataEnv()
   if (data.env=="HOME"){
     downloadRPolyhedraSupportingFiles()
   }
+  .polyhedra <- NULL
   polyhedra.rds.file <- getPolyhedraRDSPath()
   if (file.exists(polyhedra.rds.file)) {
     polyhedra.candidate <- readRDS(polyhedra.rds.file)
@@ -124,17 +131,29 @@ selectDataEnv <- function() {
   if (is.null(.polyhedra)){
     .polyhedra <- PolyhedraDatabase.class$new()
   }
+
+  assign(".polyhedra", value = .polyhedra, envir = getUserEnv())
+  scrapePolyhedra(.available.scrapping.conf[["pkg-minimal"]],
+                  sources.config = .available.sources)
   data.env
 }
 
 #' getDataEnv
 #'
-#' get the current .data.env value
+#' Gets the current .data.env value
 #'
 #' @return .data.env
-
 getDataEnv <- function() {
   get(".data.env", envir = getUserEnv())
+}
+
+#' getPolyhedraObject
+#'
+#' Gets the polyhedra object
+#'
+#' @return .polyhedra
+getPolyhedraObject <- function() {
+  get(".polyhedra", envir = getUserEnv())
 }
 
[…]
