+++ man/switchToFullDatabase.Rd
@@ -6,6 +6,9 @@
 \usage{
 switchToFullDatabase()
 }
+\arguments{
+\item{env}{The environment to run on, can be PACKAGE, HOME or NA. If null, it asks the user for a an Environment.}
+}
 \description{
 Prompts user for changing database to fulldb in user filespace
 }
