### 2018-07-04 9a31b90
changes to better support memory management
