### 2018-07-04 71f3cde
added documentation to switchToFullDatabase and modified the tests to call getPolyhedraObject()
