+++ tests/testthat/test_rpolyhedra_lib.R
@@ -3,7 +3,7 @@ context("Rpolyhedra-test")
 # Scrape tests
 context("regular solids")
 test_that("Scrape test rpolyhedra 5 regular solids", {
-  tasks <- .polyhedra$generateTestTasks(sources = "netlib",TestTaskClass = PolyhedronTestTaskScrape.class,
+  tasks <- getPolyhedraObject()$generateTestTasks(sources = "netlib",TestTaskClass = PolyhedronTestTaskScrape.class,
                                         max.quant = 5)
   for (task in tasks){
     task$run()
@@ -12,7 +12,7 @@ test_that("Scrape test rpolyhedra 5 regular solids", {
 
 context("scrape available polyhedra")
 test_that("Scrape test rpolyhedra for 12% of available polyhedra", {
-  tasks <- .polyhedra$generateTestTasks(TestTaskClass = PolyhedronTestTaskScrape.class,
+  tasks <- getPolyhedraObject()$generateTestTasks(TestTaskClass = PolyhedronTestTaskScrape.class,
                                         max.quant = getPercentilPolyhedraQuant(0.12,50))
 
   for (task in tasks){
@@ -39,7 +39,7 @@ test_that("test getting properties of scraped regular solids", {
 
 context("Edges Consistency")
 test_that("test check edges consistency for 7% of available polyhedra", {
-  tasks <- .polyhedra$generateTestTasks(TestTaskClass = PolyhedronTestTaskEdgesConsistency.class,
+  tasks <- getPolyhedraObject()$generateTestTasks(TestTaskClass = PolyhedronTestTaskEdgesConsistency.class,
                                         max.quant = getPercentilPolyhedraQuant(0.12,50))
 
   for (task in tasks){
