### 2018-07-05 3ec3fe7
solved a documentation problem, removed unneeded exports, and added support for a persistent file path storage
