+++ NAMESPACE
@@ -1,16 +1,9 @@
 # Generated by roxygen2: do not edit by hand
 
-export(ScraperLedger.class)
 export(downloadRPolyhedraSupportingFiles)
 export(getAvailablePolyhedra)
 export(getAvailableSources)
-export(getPolyhedraRDSPath)
 export(getPolyhedron)
-export(getPreloadedDataFilename)
-export(polyhedronToXML)
-export(scrapePolyhedra)
-export(scrapePolyhedraSources)
-export(setDataDirEnvironment)
 export(switchToFullDatabase)
 import(R6)
 import(XML)
