### 2018-07-05 6f73c01
solved a documentation problem, removed unneeded exports, and added support for a persistent file path storage
