### 2018-07-06 7c26273
[bugfix] reading .data.env from userspace when it is available
