+++ man/getDataDir.Rd
@@ -9,6 +9,9 @@ getDataDir(data.env = getDataEnv())
 \arguments{
 \item{data.env}{enviroment where data directory must be returned}
 }
+\value{
+dir where the package access polyhedra database
+}
 \description{
 Gets the path of Rpolyhedra data dir.
 }
