### 2018-07-06 7daefb1
[bugfix] reading .data.env from userspace when it is available
