### 2018-07-06 1df85c6
[bugfix] does not create home directory until user confirmation
