+++ R/db-lib.R
@@ -37,9 +37,6 @@ getDataDir <- function(data.env=getDataEnv()) {
   if(data.env == "HOME")
   {
     data.dir <- getUserSpace()
-    if(!dir.exists(data.dir)) {
-      dir.create(data.dir, recursive=TRUE, showWarnings = FALSE)
-    }
   }
   else{
     data.dir <- getPackageDir()
@@ -155,10 +152,20 @@ selectDataEnv <- function(env=NA) {
     setDataDirEnvironment(env)
   }
   #loads the database
-  data.env <-getDataEnv()
-  if (data.env=="HOME"){
+  .data.env <-getDataEnv()
+  if (.data.env=="HOME"){
     downloadRPolyhedraSupportingFiles()
   }
+  .data.env
+}
+
+#' updatePolyhedraDatabase
+#'
+#' Function for initializing database
+#'
+#' @return .data.env
+
+updatePolyhedraDatabase <- function(){
   .polyhedra <- NULL
   polyhedra.rds.file <- getPolyhedraRDSPath()
   if (file.exists(polyhedra.rds.file)) {
@@ -172,9 +179,10 @@ selectDataEnv <- function(env=NA) {
   }
 
   assign(".polyhedra", value = .polyhedra, envir = getUserEnv())
-  scrapePolyhedra(.available.scrapping.conf[["pkg-minimal"]],
+  .available.sources <- get(".available.sources", envir = asNamespace("Rpolyhedra"))
+  .available.scrapping.conf <- get(".available.scrapping.conf", envir = asNamespace("Rpolyhedra"))
+  scrapePolyhedra(scrape.conf = .available.scrapping.conf[["dev-minimal"]],
                   sources.config = .available.sources)
-  data.env
 }
 
 #' getDataEnv
@@ -210,11 +218,19 @@ downloadRPolyhedraSupportingFiles <- function(){
       URL <- paste("https://api.github.com/repos/qbotics/RpolyhedraDB/zipball/v", package.version, sep="")
       td <- tempdir()
       zipFile <- tempfile(tmpdir=td, fileext=".zip")
+      #download file to tempfile
       download.file(URL, destfile = zipFile, mode="wb")
       utils::unzip(zipfile = zipFile, exdir = td)
       tmp.db.path <- list.files(path = td, pattern="qbotics*")[1]
       files.to.copy <- list.files(file.path(td, tmp.db.path))
-      file.copy(from = file.path(td,tmp.db.path, files.to.copy), to=getDataDir(), recursive = TRUE)
+      #create dir
+      data.dir <- getUserSpace()
+      if(!dir.exists(data.dir)) {
+        dir.create(data.dir, recursive=TRUE, showWarnings = FALSE)
+      }
+      #copy files
+      file.copy(from = file.path(td,tmp.db.path, files.to.copy), to=data.dir, recursive = TRUE)
+      #delete tmp path
       unlink(file.path(td,tmp.db.path), recursive=TRUE)
       return(TRUE)
     }
