### 2018-07-06 69db8fa
pkg-minimal for v0.2.6
