+++ R/db-lib.R
@@ -50,7 +50,7 @@ getDataDir <- function(data.env=getDataEnv()) {
 #' @return The environment filepath
 
 getEnvironmentFilepath <- function(){
-  file.path(getDataDir("HOME"), "environment.txt")
+  file.path(getDataDir("HOME"), "Rpolyhedra.env")
 }
 
 #' setDataDirEnvironment
@@ -132,28 +132,33 @@ selectDataEnv <- function(env=NA) {
     if(!is.na(Sys.getenv(x = "ON_TRAVIS", unset=NA)))
       return(TRUE)
     accept.option <- readline(prompt="Full Database needs to download data to home folder. Agree [y/n]?:")
-    if(tolower(accept.option[1]) == "n") {
-      futile.logger::flog.info("Working on demo DB. You can call switchToFullDatabase to use the full database.")
-      setDataDirEnvironment("PACKAGE")
-    }
-    if(tolower(accept.option[1]) == "y") {
-      setDataDirEnvironment("HOME")
-    }
-    while(!tolower(accept.option[1]) %in% c("n","y")) {
-      accept.option <- readline(prompt="Unknown option. Agree [y/n]?:")
-      if(tolower(accept.option[1]) == "n") {
+    retry <- TRUE
+    while(retry) {
+      answer <- tolower(accept.option[1])
+      if(answer== "n") {
         futile.logger::flog.info("Working on demo DB. You can call selectDataEnv to use the full database.")
         setDataDirEnvironment("PACKAGE")
-      } else if(tolower(accept.option[1]) == "y") {
+        retry <- FALSE
+      } else if(answer == "y") {
         setDataDirEnvironment("HOME")
+        retry <- FALSE
+      }
+      else{
+        retry <- TRUE
+      }
+      if (retry){
+        accept.option <- readline(prompt="Unknown option. Agree [y/n]?:")
       }
     }
-  } else {
-    setDataDirEnvironment(env)
   }
   #loads the database
   .data.env <-getDataEnv()
   if (.data.env=="HOME"){
+    #create dir
+    data.dir <- getUserSpace()
+    if(!dir.exists(data.dir)) {
+      dir.create(data.dir, recursive=TRUE, showWarnings = FALSE)
+    }
     downloadRPolyhedraSupportingFiles()
   }
   .data.env
@@ -181,7 +186,7 @@ updatePolyhedraDatabase <- function(){
   assign(".polyhedra", value = .polyhedra, envir = getUserEnv())
   .available.sources <- get(".available.sources", envir = asNamespace("Rpolyhedra"))
   .available.scrapping.conf <- get(".available.scrapping.conf", envir = asNamespace("Rpolyhedra"))
-  scrapePolyhedra(scrape.conf = .available.scrapping.conf[["pkg-minimal"]],
+  scrapePolyhedra(scrape.conf = .available.scrapping.conf[["fulldb"]],
                   sources.config = .available.sources)
 }
 
@@ -223,11 +228,6 @@ downloadRPolyhedraSupportingFiles <- function(){
       utils::unzip(zipfile = zipFile, exdir = td)
       tmp.db.path <- list.files(path = td, pattern="qbotics*")[1]
       files.to.copy <- list.files(file.path(td, tmp.db.path))
-      #create dir
-      data.dir <- getUserSpace()
-      if(!dir.exists(data.dir)) {
-        dir.create(data.dir, recursive=TRUE, showWarnings = FALSE)
-      }
       #copy files
       file.copy(from = file.path(td,tmp.db.path, files.to.copy), to=data.dir, recursive = TRUE)
       #delete tmp path
