### 2018-07-06 5094bf5
[bugfix] data.dir for downloading
