### 2018-07-06 6e740e4
pkg-minimal config for v0.2.6
