+++ R/db-lib.R
@@ -186,7 +186,7 @@ updatePolyhedraDatabase <- function(){
   assign(".polyhedra", value = .polyhedra, envir = getUserEnv())
   .available.sources <- get(".available.sources", envir = asNamespace("Rpolyhedra"))
   .available.scrapping.conf <- get(".available.scrapping.conf", envir = asNamespace("Rpolyhedra"))
-  scrapePolyhedra(scrape.conf = .available.scrapping.conf[["fulldb"]],
+  scrapePolyhedra(scrape.conf = .available.scrapping.conf[["pkg-minimal"]],
                   sources.config = .available.sources)
 }
 
