+++ man/getPackageDB.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/db-lib.R
+\name{getPackageDB}
+\alias{getPackageDB}
+\title{getPackageDB}
+\usage{
+getPackageDB()
+}
+\description{
+Obtains the database version from environment
+}
