### 2018-07-09 f35a879
added documentation changes to comply with cran
