+++ man/getDataDirEnvironment.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/db-lib.R
-\name{getDataDirEnvironment}
-\alias{getDataDirEnvironment}
-\title{getDataDirEnv}
-\usage{
-getDataDirEnvironment()
-}
-\value{
-the data dir environment
-}
-\description{
-Gets the data dir environment
-}
