### 2018-07-10 bbe0336
encapsulated environment functions
