+++ R/db-lib.R
@@ -1,3 +1,47 @@
+#' getPackageEnvir
+#'
+#' Obtains a variable from package environment
+getPackageEnvir <- function(variable.name){
+  get(variable.name,envir=asNamespace("Rpolyhedra"))
+}
+
+#' setPackageEnvir
+#'
+#' Set a variable from package environment
+setPackageEnvir <- function(variable.name, value){
+  assign(x = variable.name, value = value, envir = asNamespace("Rpolyhedra"))
+}
+
+
+#' getUserEnvir
+#'
+#' Gets a writable environment for the package
+#'
+getUserEnvir <- function(variable.name) {
+  get(x=variable.name , envir = getPackageEnvir("RpolyhedraEnv"))
+}
+
+#' setUserEnvir
+#'
+#' Sets variable in the writable environment for the package
+#'
+setUserEnvir <- function(variable.name, value) {
+  assign(x=variable.name , value = value, envir = getPackageEnvir("RpolyhedraEnv"))
+}
+
+
+#' getDataEnv
+#'
+#' Gets the current .data.env value
+#'
+#' @return .data.env
+getDataEnv <- function() {
+  #get(".data.env", envir = getUserEnv())
+  getUserEnvir(".data.env")
+  #dev-package-env
+}
+
+
 #' getUserSpace
 #'
 #' This function is used internally for accesing the local database path
@@ -18,9 +62,15 @@ initDataDirEnvironment <- function() {
   }else{
     .data.env <- readLines(environment.filepath)[1]
   }
-  assign(".data.env", value = .data.env, envir = getUserEnv())
+
+  setUserEnvir(".data.env", value = .data.env)
+  #assign(".data.env", value = .data.env, envir = getUserEnv())
+  #dev-package-env
+
   .data.env
 }
+
+
 
 
 #' getDataDir
@@ -31,6 +81,7 @@ initDataDirEnvironment <- function() {
 #' is compiled in source or package directory.
 #' @param data.env enviroment where data directory must be returned
 #' @return dir where the package access polyhedra database
+
 getDataDir <- function(data.env=getDataEnv()) {
   data.dir <- ""
   if(data.env == "HOME")
@@ -70,19 +121,15 @@ setDataDirEnvironment <- function(env="PACKAGE") {
   if (file.exists(getUserSpace())){
     write(.data.env, getEnvironmentFilepath())
   }
-  assign(".data.env", value = .data.env, envir = getUserEnv())
+
+  setUserEnvir(".data.env", value = .data.env)
+  #assign(".data.env", value = .data.env, envir = getUserEnv())
+  #dev-package-env
+
   .data.env
 }
 
 
-
-#' getUserEnv
-#'
-#' Gets a writable environment for the package
-#'
-getUserEnv <- function() {
-  get("RpolyhedraEnv",envir=asNamespace("Rpolyhedra"))
-}
 
 
 #' getPackageDir
@@ -168,11 +215,15 @@ selectDataEnv <- function(env=NA) {
 #'
 #' Function for initializing database
 #'
-#' @return .data.env
+#'
+#' @return polyhedra db object
 
 updatePolyhedraDatabase <- function(){
   .polyhedra <- NULL
+
   polyhedra.rds.file <- getPolyhedraRDSPath()
+
+
   if (file.exists(polyhedra.rds.file)) {
     polyhedra.candidate <- readRDS(polyhedra.rds.file)
     if (isCompatiblePolyhedraRDS(polyhedra.candidate, halts = TRUE)){
@@ -183,21 +234,17 @@ updatePolyhedraDatabase <- function(){
     .polyhedra <- PolyhedraDatabase.class$new()
   }
 
[…]
