+++ man/setUserEnvir.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/db-lib.R
+\name{setUserEnvir}
+\alias{setUserEnvir}
+\title{setUserEnvir}
+\usage{
+setUserEnvir(variable.name, value)
+}
+\description{
+Sets variable in the writable environment for the package
+}
