+++ man/getUserEnvir.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/db-lib.R
-\name{getUserEnv}
-\alias{getUserEnv}
-\title{getUserEnv}
+\name{getUserEnvir}
+\alias{getUserEnvir}
+\title{getUserEnvir}
 \usage{
-getUserEnv()
+getUserEnvir(variable.name)
 }
 \description{
 Gets a writable environment for the package
