+++ man/scrapePolyhedraSources.Rd
@@ -18,6 +18,9 @@ scrapePolyhedraSources(sources.config = .available.sources, max.quant.config.sch
 
 \item{retry.scrape}{should it retry scrape?}
 }
+\value{
+polyhedra db object
+}
 \description{
 Method for obtaining polyhedra objects from text files of
 different sources, scheduling and scraping
