### 2018-07-10 0076bff
encapsulated environment functions
