+++ man/setPackageEnvir.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/db-lib.R
+\name{setPackageEnvir}
+\alias{setPackageEnvir}
+\title{setPackageEnvir}
+\usage{
+setPackageEnvir(variable.name, value)
+}
+\description{
+Set a variable from package environment
+}
