+++ man/getPackageEnvir.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/db-lib.R
+\name{getPackageEnvir}
+\alias{getPackageEnvir}
+\title{getPackageEnvir}
+\usage{
+getPackageEnvir(variable.name)
+}
+\description{
+Obtains a variable from package environment
+}
