### 2018-07-10 73915c1
[bugfix] package documentation
