### 2018-07-10 d4bb347
[bugfix] package documentation
