+++ R/zzz.R
@@ -28,7 +28,7 @@
   #package version db
   .package.db <- list()
   .package.db[["0.2.5"]] <- "0.2.5"
-  .package.db[["0.2.6"]] <- "0.2.5"
+  .package.db[["0.2.6"]] <- "0.2.6"
 
   setPackageEnvir(variable.name = ".package.db", .package.db)
   #assign(".package.db", value = .package.db, envir = asNamespace("Rpolyhedra"))
