### 2018-07-11 5fde000
solved a problem with switchToFullDB in which forcing an environment did not change it
