### 2018-07-20 fc0558c
minor changes
