### 2018-07-20 ccd1c23
rpoly.polyhedron_invalid_face
