### 2018-07-22 c7c6727
rpoly.polyhedron_invalid_face
