+++ R/polyhedra-lib.R
@@ -534,7 +534,6 @@ triangulate = function(force = FALSE) {
             current.vertex <- nrow(private$vertices.rgl)
             # the description considers vector referencing starting in 0.
             if (length(face) < 3) {
-
               stop(paste(gettext("rpoly.polyhedron_invalid_face", domain = "R-Rpolyhedra"),length(face)))
             }
             if (length(face) == 3) {
