### 2018-09-18 7040f8f
Implementing #22. By now, polyhedron are serialized and deserialized in RDS
