+++ man/PolyhedronStateDeserializer.class.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/polyhedra-lib.R
+\docType{class}
+\name{PolyhedronStateDeserializer.class}
+\alias{PolyhedronStateDeserializer.class}
+\title{Polyhedron State Deserializer}
+\format{\code{\link{R6Class}} object.}
+\usage{
+PolyhedronStateDeserializer.class
+}
+\description{
+Polyhedron state for deserialize from database
+}
+\section{Fields}{
+
+\describe{
+\item{\code{serialized.polyhedron}}{polyhedron definition serialized}
+}}
+
+\section{Methods}{
+
+\describe{
+  \item{\code{initialize(serialized.polyhedron)}}{Initializes the object.}
+  \item{\code{scrape()}}{Returns a State Defined}
+
+}
+}
+
+\keyword{datasets}
