### 2018-09-18 0596c11
Implementing #22. By now, polyhedron are serialized and deserialized in RDS
