+++ R/polyhedra-lib.R
@@ -6,22 +6,22 @@
 #' \describe{
 #'   \item{\code{addError(current.error)}}{Adds an error to the error string and log it as info}
 #'   \item{\code{scrape()}}{Scrapes the polyhedra folder files}
-#'   \item{\code{geSolid()}}{returns the object corresponding to the solid}
+#'   \item{\code{getSolid()}}{returns the object corresponding to the solid}
 #'   \item{\code{buildRGL(size = 1, origin = c(0, 0, 0), normalize.size = TRUE)}}{creates a RGL representation of the object}
 #'   \item{\code{exportToXML()}}{Gets an XML representation out of the polyhedron object}
 #' }
 #' @field errors Errors string
 #' @field source polyhedron definition source
-#' @field number polyhedron number
+#' @field file.id polyhedron file id
 #' @format \code{\link{R6Class}} object.
 #' @docType class
 #' @import rgl
 #' @importFrom R6 R6Class
 PolyhedronState.class <- R6::R6Class("PolyhedronState", public = list(
-source = NA, number = NA, errors = "",
-initialize = function(source, number) {
+source = NA, file.id = NA, errors = "",
+initialize = function(source, file.id) {
     self$source <- source
-    self$number <- number
+    self$file.id <- file.id
     self
 },
 addError = function(current.error) {
@@ -52,8 +52,8 @@ exportToXML = function(){
 #'
 #' @section Methods:
 #' \describe{
-#'   \item{\code{initialize(number, netlib.p3.lines)}}{Initializes the object, taking the number and PDH file as parameters}
-#'   \item{\code{extract_fows_from_label(label.number, expected.label)}}{Extracts data from the label, taking the label number and the
+#'   \item{\code{initialize(file.id, netlib.p3.lines)}}{Initializes the object, taking the file.id and PDH file as parameters}
+#'   \item{\code{extract_rows_from_label(label.number, expected.label)}}{Extracts data from the label, taking the label number and the
 #'   expected label as parameters}
 #'   \item{\code{getLabels()}}{Gets the label from the polyhedron}
 #'   \item{\code{scrapeNet(net.txt, offset = 0) }}{Scrape the net model}
@@ -74,12 +74,12 @@ exportToXML = function(){
 PolyhedronStateNetlibScraper.class <- R6::R6Class("PolyhedronStateNetlibScraper",
 inherit = PolyhedronState.class,
 public = list(netlib.p3.lines = NA, labels.rows = NA, labels.map = NA, errors = "",
-initialize = function(number, netlib.p3.lines) {
-    super$initialize(source= "netlib",number = number)
+initialize = function(file.id, netlib.p3.lines) {
+    super$initialize(source= "netlib",file.id = file.id)
     self$netlib.p3.lines <- netlib.p3.lines
     self$labels.map <- list()
     self
-}, extract_fows_from_label = function(label.number, expected.label) {
+}, extract_rows_from_label = function(label.number, expected.label) {
     observer.label <- self$netlib.p3.lines[self$labels.rows[label.number]]
     observer.label <- sub("\\:", "", observer.label)
     if (observer.label != expected.label) {
@@ -171,8 +171,8 @@ initialize = function(number, netlib.p3.lines) {
 }, getDataFromLabel = function(label) {
     r <- self$labels.map[[label]]
     ret <- NULL
-    if (!is.null(r)) ret <- self$netlib.p3.lines[self$extract_fows_from_label(r,
-    label)]
+    if (!is.null(r)) ret <- self$netlib.p3.lines[
+                                  self$extract_rows_from_label(r,label)]
     ret
[…]
