+++ R/db-lib.R
@@ -252,8 +252,12 @@ updatePolyhedraDatabase <- function(){
   .available.sources <- getPackageEnvir(".available.sources")
   .available.scrapping.conf <- getPackageEnvir(".available.scrapping.conf")
 
-  scrapePolyhedra(scrape.config = .available.scrapping.conf[["pkg-minimal"]],
+
+  scrapePolyhedra(scrape.config = .available.scrapping.conf[["dev-minimal"]],
                   sources.config = .available.sources)
+  #Change when release version
+  # scrapePolyhedra(scrape.config = .available.scrapping.conf[["pkg-minimal"]],
+  #                 sources.config = .available.sources)
 }
 
 
@@ -394,7 +398,7 @@ PolyhedronScraperConfiguration.class <- R6::R6Class("PolyhedronScraperConfigurat
 #' \describe{
 #'   \item{\code{initialize()}}{initializes the object}
 #'   \item{\code{getPolyhedraFiles(home.dir.data)}}{returns the file names on the netlib database}
-#'   \item{\code{scrape(polyhedron.number, polyhedron.filename)}}{scrapes the object}
+#'   \item{\code{scrape(polyhedron.file.id, polyhedron.filename)}}{scrapes the object}
 #' }
 #'
 #' @format \code{\link{R6Class}} object.
@@ -415,12 +419,12 @@ PolyhedronScraperConfigurationNetlib.class <- R6::R6Class("PolyhedronScraperConf
       polyhedra.files <- polyhedra.files[order(as.numeric(polyhedra.files))]
       polyhedra.files
     },
-    scrape = function(polyhedron.number, polyhedron.filename){
+    scrape = function(polyhedron.file.id, polyhedron.filename){
       polyhedra.netlib.lines <- readLines(polyhedron.filename)
-      polyhedron.number <- polyhedron.filename
-      polyhedron.number <- strsplit(polyhedron.number,split = "/")[[1]]
-      polyhedron.number <- as.numeric(polyhedron.number[length(polyhedron.number)])
-      current.polyhedron <- Polyhedron.class$new(number=polyhedron.number)
+      polyhedron.file.id <- polyhedron.filename
+      polyhedron.file.id <- strsplit(polyhedron.file.id,split = "/")[[1]]
+      polyhedron.file.id <- polyhedron.file.id[length(polyhedron.file.id)]
+      current.polyhedron <- Polyhedron.class$new(file.id=polyhedron.file.id)
       current.polyhedron$scrapeNetlib(netlib.p3.lines = polyhedra.netlib.lines)
       futile.logger::flog.debug(paste("parsed", polyhedron.filename, "with name",
                                       current.polyhedron$state$name))
@@ -436,7 +440,7 @@ PolyhedronScraperConfigurationNetlib.class <- R6::R6Class("PolyhedronScraperConf
 #' \describe{
 #'   \item{\code{initialize()}}{initializes the object}
 #'   \item{\code{getPolyhedraFiles(home.dir.data)}}{returns the file names on the netlib database}
-#'   \item{\code{scrape(polyhedron.number, polyhedron.filename)}}{scrapes the object}
+#'   \item{\code{scrape(polyhedron.file.id, polyhedron.filename)}}{scrapes the object}
 #' }
 #'
 #' @format \code{\link{R6Class}} object.
@@ -462,9 +466,9 @@ PolyhedronScraperConfigurationDmccoey.class <- R6::R6Class("PolyhedronScraperCon
       polyhedra.files <- polyhedra.files[grep("\\.txt", polyhedra.files)]
       polyhedra.files
     },
-    scrape = function(polyhedron.number, polyhedron.filename){
[…]
