### 2018-09-27 0ced344
added changes to Rpolyhedra
