+++ NAMESPACE
@@ -7,6 +7,7 @@ export(getAvailableSources)
 export(getPolyhedron)
 export(switchToFullDatabase)
 import(R6)
+import(RJSONIO)
 import(XML)
 import(futile.logger)
 import(rgl)
