### 2018-09-27 d457377
added changes to Rpolyhedra
