+++ man/polyhedronToXML.Rd
@@ -1,17 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/xml-lib.R
-\name{polyhedronToXML}
-\alias{polyhedronToXML}
-\title{polyhedronToXML()}
-\usage{
-polyhedronToXML(polyhedron.state.defined)
-}
-\arguments{
-\item{polyhedron.state.defined}{the polyhedron to get a representation from}
-}
-\value{
-an XML document, ready to be converted to String with XML::saveXML()
-}
-\description{
-Gets an XML representation out of the polyhedron object
-}
