### 2018-09-27 56d128f
added an addPolyhedron working with json zip files
