+++ man/hydratePolyhedron.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/serialization-lib.R
+\name{hydratePolyhedron}
+\alias{hydratePolyhedron}
+\title{hydratePolyhedron}
+\usage{
+hydratePolyhedron(file.path)
+}
+\value{
+polyhedron.state.defined
+}
+\description{
+hydrates a polyhedron from the defined path
+}
