### 2018-09-27 ef27219
added an addPolyhedron working with json zip files
