+++ R/db-lib.R
@@ -383,7 +383,7 @@ PolyhedronScraperConfiguration.class <- R6::R6Class("PolyhedronScraperConfigurat
       self$name
     },
     getBaseDir = function(home.dir.data) {
-      file.path(home.dir.data, self$base.dir)
+      file.path(home.dir.data, "sources", self$base.dir)
     },
     getPolyhedraFiles = function(home.dir.data){
       stop(gettext("rpoly.abstract_class", domain = "R-Rpolyhedra"))
@@ -678,7 +678,7 @@ checkDatabaseVersion <- function(){
 #' \describe{
 #'   \item{\code{initialize()}}{Initializes the object}
 #'   \item{\code{existsSource(source)}}{Determines if the source exists on the database}
-#'   \item{\code{getSource(source, strict=False)}}{Retrieves a source by name}
+#'   \item{\code{getPolyhedraSourceDir(source)}}{Retrieves polyhedra dir of a source}
 #'   \item{\code{addSource(source)}}{Adds a new source to the database}
 #'   \item{\code{configPolyhedraRDSPath()}}{config path for rds database file}
 #'   \item{\code{existsPolyhedron(source,polyhedron.name)}}{Determines if the polyhedron exists on the database}
@@ -699,7 +699,7 @@ checkDatabaseVersion <- function(){
 #'
 #' @format \code{\link{R6Class}} object.
 #' @docType class
-#' @import futile.logger
+#' @import futile.logger RJSONIO
 #' @importFrom R6 R6Class
 PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
   public = list(
@@ -707,12 +707,10 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
     polyhedra.rds.file = NA,
     sources.config = NA,
     ledger         = NA,
-    data           = NA,
     initialize = function() {
       self$version        <- getDatabaseVersion()
       self$ledger         <- ScraperLedger.class$new()
       self$sources.config <- list()
-      self$data           <- list()
       self
     },
     getVersion = function(){
@@ -723,44 +721,46 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
       self$polyhedra.rds.file
     },
     existsSource = function(source){
-      ret <- FALSE
-      if (length(self$data)>0){
-        ret <- source %in% names(self$data)
-      }
-      ret
-    },
-    getSource = function(source, strict = FALSE) {
-      ret <- NULL
-      if (strict & !self$existsSource(source)){
-        stop(paste("Source",source,"not available in polyhedra database"))
-      }
-      self$data[[source]]
+      source %in% self$getAvailableSources()
     },
     addSourceConfig = function(source.config) {
       source <- source.config$getName()
       if (!self$existsSource(source)){
         self$sources.config[[source]] <- source.config
-        self$data[[source]] <- list()
+        self$getPolyhedraSourceDir(source)
       }
       self
     },
     existsPolyhedron = function(source = "netlib",polyhedron.name) {
       ret <- FALSE
-      source.data <- self$getSource(source)
-      if (length(source.data)>0){
-        ret <- polyhedron.name %in% names(source.data)
+      file.path <- self$getPolyhedronFilename(source = source, polyhedron.name = polyhedron.name)
+      ret <- !is.null(file.path)
+      if (ret){
+        ret <- file.exists(file.path)
       }
       ret
     },
+    getPolyhedraSourceDir = function(source, create.dir = TRUE){
+      ret <- file.path(getDataDir(), "polyhedra",source)
+      if (create.dir){
+        dir.create(ret, showWarnings = FALSE, recursive = TRUE)
+      }
+      ret
+    },
[…]
