+++ man/persistPolyhedron.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/serialization-lib.R
+\name{persistPolyhedron}
+\alias{persistPolyhedron}
+\title{persistPolyhedron}
+\usage{
+persistPolyhedron(polyhedron.state.defined, file.path)
+}
+\arguments{
+\item{polyhedron.state.defined}{the polyhedron to get a representation from}
+}
+\value{
+the path or None
+}
+\description{
+persists a polyhedron to the defined path in a Zipped json format
+}
+\examples{
+#get the representation of a cube (netlib library)
+library(Rpolyhedra)
+polyhedronToJSON(getPolyhedron("netlib", "cube")$state)
+
+}
