### 2018-09-27 8e3b027
now polyhedra database works with RDS in ascii per polyhedron
