### 2018-09-27 bd698ce
now polyhedra database works with RDS in ascii per polyhedron
