### 2018-09-27 4b5f381
added more functions to the package'
