### 2018-09-27 961d45a
added more functions to the package'
