### 2018-10-01 a1d7dbb
test working without comparing R6 objects
