+++ R/db-lib.R
@@ -253,11 +253,11 @@ updatePolyhedraDatabase <- function(){
   .available.scrapping.conf <- getPackageEnvir(".available.scrapping.conf")
 
 
-  scrapePolyhedra(scrape.config = .available.scrapping.conf[["dev-tetrahedron"]],
-                  sources.config = .available.sources)
+  #scrapePolyhedra(scrape.config = .available.scrapping.conf[["dev-tetrahedron"]],
+  #                sources.config = .available.sources)
   #Change when release version
-  # scrapePolyhedra(scrape.config = .available.scrapping.conf[["pkg-minimal"]],
-  #                 sources.config = .available.sources)
+  scrapePolyhedra(scrape.config = .available.scrapping.conf[["dev-minimal"]],
+                   sources.config = .available.sources)
 }
 
 
@@ -583,15 +583,7 @@ PolyhedronTestTaskScrape.class <- R6::R6Class("PolyhedronTestTaskScrape.class",
                 self$polyhedra.db$getPolyhedron(source = source,
                                                 polyhedron.name = scraped.name)
 
-      #debug gp
-      #print(scraped.polyhedron)
-      #print(expected.polyhedron)
-      scraped.polyhedron <<- scraped.polyhedron
-      expected.polyhedron <<- expected.polyhedron
-      #task <<- self
-
-      expect_equal(scraped.polyhedron, expected.polyhedron)
-
+      expected.polyhedron$getState()$expect_equal(scraped.polyhedron)
     }))
 
 #' PolyhedronTestTaskEdgesConsistency
@@ -700,7 +692,7 @@ checkDatabaseVersion <- function(){
 #'
 #' @format \code{\link{R6Class}} object.
 #' @docType class
-#' @import futile.logger RJSONIO
+#' @import futile.logger
 #' @importFrom R6 R6Class
 PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
   public = list(
@@ -759,7 +751,6 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
           stop(message)
         }
       }
-      #TODO use JSON
       #serialized.polyhedron <- source.data[[polyhedron.name]]
       ret <- NULL
       zip.filename <- file.path(data.dir, paste(polyhedron.name, ".RDS.zip", sep=""))
@@ -768,9 +759,9 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
       if(file.exists(zip.filename)) {
         tmp.dir <- file.path(tempdir=tempdir(), source)
         dir.create(tmp.dir, showWarnings = FALSE, recursive = TRUE)
-        json.filename <- paste(polyhedron.name, ".RDS", sep="")
-        tmp.filename <- file.path(tmp.dir, json.filename)
-        unzip(zipfile = zip.filename, files = json.filename, exdir = tmp.dir)
+        serialized.filename <- paste(polyhedron.name, ".RDS", sep="")
+        tmp.filename <- file.path(tmp.dir, serialized.filename)
+        unzip(zipfile = zip.filename, files = serialized.filename, exdir = tmp.dir)
         serialized.polyhedron <- readRDS(file = tmp.filename)
         unlink(tmp.filename)
       }
@@ -793,8 +784,8 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
         serialized.polyhedron <- polyhedron$state$serialize()
         tmp.dir <- file.path(tempdir=tempdir(), source)
         dir.create(tmp.dir, showWarnings = FALSE, recursive = TRUE)
-        json.filename <- paste(polyhedron.name, ".RDS", sep="")
[…]
