+++ NAMESPACE
@@ -5,9 +5,9 @@ export(downloadRPolyhedraSupportingFiles)
 export(getAvailablePolyhedra)
 export(getAvailableSources)
 export(getPolyhedron)
+export(polyhedronToXML)
 export(switchToFullDatabase)
 import(R6)
-import(RJSONIO)
 import(XML)
 import(futile.logger)
 import(rgl)
