### 2018-10-01 7dd7454
test working without comparing R6 objects
