+++ man/polyhedronToXML.Rd
@@ -1,23 +1,23 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/serialization-lib.R
-\name{polyhedronToJSON}
-\alias{polyhedronToJSON}
-\title{polyhedronToJSON()}
+\name{polyhedronToXML}
+\alias{polyhedronToXML}
+\title{polyhedronToXML()}
 \usage{
-polyhedronToJSON(polyhedron.state.defined)
+polyhedronToXML(polyhedron.state.defined)
 }
 \arguments{
 \item{polyhedron.state.defined}{the polyhedron to get a representation from}
 }
 \value{
-an JSON document, ready to be saved as a String File
+a XML document, ready to be saved as a String File
 }
 \description{
-Gets an JSON representation out of the polyhedron object
+Gets an XML representation out of the polyhedron object
 }
 \examples{
 #get the representation of a cube (netlib library)
 library(Rpolyhedra)
-polyhedronToJSON(getPolyhedron("netlib", "cube")$state)
+polyhedronToXML(getPolyhedron("netlib", "cube")$state)
 
 }
