+++ R/db-lib.R
@@ -398,7 +398,7 @@ PolyhedronScraperConfiguration.class <- R6::R6Class("PolyhedronScraperConfigurat
 #' \describe{
 #'   \item{\code{initialize()}}{initializes the object}
 #'   \item{\code{getPolyhedraFiles(home.dir.data)}}{returns the file names on the netlib database}
-#'   \item{\code{scrape(polyhedron.file.id, polyhedron.filename)}}{scrapes the object}
+#'   \item{\code{scrape(polyhedron.file.id, source.filename)}}{scrapes the object}
 #' }
 #'
 #' @format \code{\link{R6Class}} object.
@@ -419,14 +419,17 @@ PolyhedronScraperConfigurationNetlib.class <- R6::R6Class("PolyhedronScraperConf
       polyhedra.files <- polyhedra.files[order(as.numeric(polyhedra.files))]
       polyhedra.files
     },
-    scrape = function(polyhedron.file.id, polyhedron.filename){
-      polyhedra.netlib.lines <- readLines(polyhedron.filename)
-      polyhedron.file.id <- polyhedron.filename
+    scrape = function(polyhedron.file.id, source.filename){
+      #debug
+      print(paste("netlib",source.filename))
+
+      polyhedra.netlib.lines <- readLines(source.filename)
+      polyhedron.file.id <- source.filename
       polyhedron.file.id <- strsplit(polyhedron.file.id,split = "/")[[1]]
       polyhedron.file.id <- polyhedron.file.id[length(polyhedron.file.id)]
       current.polyhedron <- Polyhedron.class$new(file.id=polyhedron.file.id)
       current.polyhedron$scrapeNetlib(netlib.p3.lines = polyhedra.netlib.lines)
-      futile.logger::flog.debug(paste("parsed", polyhedron.filename, "with name",
+      futile.logger::flog.debug(paste("parsed", source.filename, "with name",
                                       current.polyhedron$state$name))
       current.polyhedron
     }
@@ -440,7 +443,7 @@ PolyhedronScraperConfigurationNetlib.class <- R6::R6Class("PolyhedronScraperConf
 #' \describe{
 #'   \item{\code{initialize()}}{initializes the object}
 #'   \item{\code{getPolyhedraFiles(home.dir.data)}}{returns the file names on the netlib database}
-#'   \item{\code{scrape(polyhedron.file.id, polyhedron.filename)}}{scrapes the object}
+#'   \item{\code{scrape(polyhedron.file.id, source.filename)}}{scrapes the object}
 #' }
 #'
 #' @format \code{\link{R6Class}} object.
@@ -466,11 +469,15 @@ PolyhedronScraperConfigurationDmccoey.class <- R6::R6Class("PolyhedronScraperCon
       polyhedra.files <- polyhedra.files[grep("\\.txt", polyhedra.files)]
       polyhedra.files
     },
-    scrape = function(polyhedron.file.id, polyhedron.filename){
-      polyhedra.dmccoey.lines <- readLines(polyhedron.filename)
+    scrape = function(polyhedron.file.id, source.filename){
+      #debug
+      print(paste("dmccooey",source.filename))
+      print(polyhedron.file.id)
+
+      polyhedra.dmccoey.lines <- readLines(source.filename)
       current.polyhedron <- Polyhedron.class$new(file.id=polyhedron.file.id)
       current.polyhedron$scrapeDmccoey(polyhedra.dmccoey.lines = polyhedra.dmccoey.lines)
-      futile.logger::flog.debug(paste("parsed", polyhedron.filename, "with name",
+      futile.logger::flog.debug(paste("parsed", source.filename, "with name",
                                       current.polyhedron$state$name))
[…]
