### 2018-10-01 5b4bdcc
CRC generated filename for scraped polyhedron
