### 2018-10-01 fb65995
CRC generated filename for scraped polyhedron
