### 2018-10-01 998c091
removed debugging infrastructure
