### 2018-10-01 6b40908
[bugfix] XML and db version to 0.2.7
