+++ R/ledger-lib.R
@@ -31,7 +31,7 @@ maxWithoutNA <- function(x) ifelse( !all(is.na(x)), max(x, na.rm=TRUE), NA)
 #' @format \code{\link{R6Class}} object.
 #' @docType class
 #' @import futile.logger
-#' @import utils
+#' @import utils digest
 #' @importFrom R6 R6Class
 #' @export
 ScraperLedger.class <- R6::R6Class("ScraperLedger",
