### 2018-10-01 b6975b2
[bugfix] XML and db version to 0.2.7
