+++ NAMESPACE
@@ -9,6 +9,7 @@ export(polyhedronToXML)
 export(switchToFullDatabase)
 import(R6)
 import(XML)
+import(digest)
 import(futile.logger)
 import(rgl)
 import(stringr)
