+++ changelog.txt
@@ -1,0 +1,146 @@
+bb302c6 (HEAD -> master, origin/dev-leobelen) changed R badge to 3.4.0 in Readme.md)
+d513dd5 (tag: v0.2.7) Release version 0.2.7
+6b40908 [bugfix] XML and db version to 0.2.7
+998c091 removed debugging infrastructure
+fb65995 CRC generated filename for scraped polyhedron
+c5f6625 Changed filename in ledger to source.filename when is name of scraped sources
+a1d7dbb test working without comparing R6 objects
+961d45a added more functions to the package'
+bd698ce now polyhedra database works with RDS in ascii per polyhedron
+56d128f added an addPolyhedron working with json zip files
+d457377 added changes to Rpolyhedra
+0596c11 (origin/serialize-R6) Implementing #22. By now, polyhedron are serialized and deserialized in RDS
+a5d70a9 Merge remote-tracking branch 'origin/dev-leobelen' into bugfix-R6
+c7c6727 (origin/dev-package-env) rpoly.polyhedron_invalid_face
+f1536cf rpoly.polyhedron_invalid_face
+27ea70a minor changes
+5fde000 solved a problem with switchToFullDB in which forcing an environment did not change it
+19a80f9 (tag: v0.2.6) regenerated db for v0.2.6
+73915c1 [bugfix] package documentation
+bbe0336 encapsulated environment functions
+c0c4465 Merge remote-tracking branch 'origin/dev-leobelen' into dev-package-env
+a890dfd removed methods from Imports in the DESCRIPTION file to comply with CRAN rules
+f35a879 added documentation changes to comply with cran
+d2c7c36 Update polyhedra database after switch
+43e12b1 v0.2.6 with db v0.2.5
+afd547d Merge remote-tracking branch 'origin/dev-package-env' into dev-package-env
+cfe8b67 pkg-minimal config for v0.2.6
+ddac524 [bugfix] data.dir for downloading
+482ca41 environment.txt to Rpolyhedra.env
+3142707 pkg-minimal for v0.2.6
+1df85c6 [bugfix] does not create home directory until user confirmation
+0e6cc46 db generated v.0.2.6
+f96fbc6 Generating db v0.2.6
+7daefb1 [bugfix] reading .data.env from userspace when it is available
+984f00b missing page added
+3ec3fe7 solved a documentation problem, removed unneeded exports, and added support for a persistent file path storage
+71f3cde added documentation to switchToFullDatabase and modified the tests to call getPolyhedraObject()
+9a31b90 changes to better support memory management
+f0377fd (origin/dev-remote-db) now it reloads the .polyhedra object after switchToFullDatabase()
+fe1f936 workaround for #17 until next release
+7212c9c #17 closed
+420a56f Solved the issue indicated in #17"
+15534a7 [bugfix] can change .data.env but have to access without exporting selectDataEnv
+1f11691 (tag: v0.2.5) solved a problem with downloadRPolyhedraSupportingFiles not being called when selecting fullDB
+c7b7a1c environment changes to allow the modification of package wide variables
+17fbc64 [bugfix] in filenames for polyhedra.RDS
+17ff330 moved on to file.path on db-lib.R
+f737781 minor changes removing debug info
+2e894dc Corrected v0.2.5 from fulldb branch
+da16a77 Merge remote-tracking branch 'origin/RpolyhedraFullDB' into RpolyhedraFullDB
+74e2561 fulldb for v0.2.5
+e4055f7 V0.2.5 with pkg-minimal db
+8ad3989 V0.2.5 with pkg-minimal db
+c422a2e generating pkg-minimal
+c8bb580 added travis awareness for the user download message
[…]
