### 2018-10-04 08735f1
move genchangelog.sh to utils for better hqndling
