+++ utils/genchangelog.sh
@@ -1,0 +1,1 @@
+git log --oneline --decorate  > ../changelog.txt
