### 2018-10-04 b57ff34
[bugfix] changed changelog file name to respond to CRAN standards
