### 2018-10-09 4624ab4
added git2r integration according with #2
