+++ man/getGitCommit.Rd
@@ -5,7 +5,10 @@
 \title{getGitCommit
 get the last git commit sha}
 \usage{
-getGitCommit()
+getGitCommit(long.version = FALSE)
+}
+\arguments{
+\item{long.version}{determines if the complete version of the sha will be returned.}
 }
 \value{
 String with git commit sha
