+++ DESCRIPTION
@@ -22,7 +22,8 @@ Imports: futile.logger,
          R6,
          testthat, 
          XML,
-         digest
+         digest,
+         git2r
 Collate:
   'db-lib.R'
   'polyhedra-lib.R'
