### 2018-10-09 f84cae1
added git2r integration according with #2
