+++ NAMESPACE
@@ -11,6 +11,7 @@ import(R6)
 import(XML)
 import(digest)
 import(futile.logger)
+import(git2r)
 import(rgl)
 import(stringr)
 import(testthat)
