### 2018-11-12 bdc138f
linted R files except polyhedra-lib.R
