+++ R/db-lib.R
@@ -4,7 +4,7 @@
 
 #' Obtains a variable from package environment
 getPackageEnvir <- function(variable.name){
-  get(variable.name,envir=asNamespace("Rpolyhedra"))
+  get(variable.name, envir = asNamespace("Rpolyhedra"))
 }
 
 #' setPackageEnvir
@@ -24,7 +24,7 @@ setPackageEnvir <- function(variable.name, value){
 #'
 #' @param variable.name  name of variable to be retrieved
 getUserEnvir <- function(variable.name) {
-  get(x=variable.name , envir = getPackageEnvir("RpolyhedraEnv"))
+  get(x = variable.name, envir = getPackageEnvir("RpolyhedraEnv"))
 }
 
 #' setUserEnvir
@@ -35,7 +35,8 @@ getUserEnvir <- function(variable.name) {
 #' @param value          variable value
 
 setUserEnvir <- function(variable.name, value) {
-  assign(x=variable.name , value = value, envir = getPackageEnvir("RpolyhedraEnv"))
+  assign(x = variable.name, value = value, envir =
+           getPackageEnvir("RpolyhedraEnv"))
 }
 
 
@@ -45,9 +46,7 @@ setUserEnvir <- function(variable.name, value) {
 #'
 #' @return .data.env
 getDataEnv <- function() {
-  #get(".data.env", envir = getUserEnv())
   getUserEnvir(".data.env")
-  #dev-package-env
 }
 
 
@@ -66,16 +65,13 @@ getUserSpace <- function(){
 #' @return the data dir environment
 initDataDirEnvironment <- function() {
   environment.filepath <- getEnvironmentFilepath()
-  if(!file.exists(environment.filepath)){
+  if (!file.exists(environment.filepath)){
     .data.env <- "PACKAGE"
-  }else{
+  } else {
     .data.env <- readLines(environment.filepath)[1]
   }
 
   setUserEnvir(".data.env", value = .data.env)
-  #assign(".data.env", value = .data.env, envir = getUserEnv())
-  #dev-package-env
-
   .data.env
 }
 
@@ -91,13 +87,11 @@ initDataDirEnvironment <- function() {
 #' @param data.env enviroment where data directory must be returned
 #' @return dir where the package access polyhedra database
 
-getDataDir <- function(data.env=getDataEnv()) {
+getDataDir <- function(data.env = getDataEnv()) {
   data.dir <- ""
-  if(data.env == "HOME")
-  {
+  if (data.env == "HOME") {
     data.dir <- getUserSpace()
-  }
-  else{
+  } else {
     data.dir <- getPackageDir()
   }
   data.dir
@@ -120,21 +114,18 @@ getEnvironmentFilepath <- function(){
 #'
 #' @param env The type of environment to work with. Values are "PACKAGE" or "HOME" and it defaults to package
 #' @return the curruent .data.env
-setDataDirEnvironment <- function(env="PACKAGE") {
-  if(env %in% c("PACKAGE","HOME")){
+setDataDirEnvironment <- function(env = "PACKAGE") {
+  if (env %in% c("PACKAGE", "HOME")) {
     .data.env <- env
   }
-  else{
+  else {
     stop("Possible values are PACKAGE and HOME")
   }
-  if (file.exists(getUserSpace())){
+  if (file.exists(getUserSpace())) {
     write(.data.env, getEnvironmentFilepath())
   }
 
   setUserEnvir(".data.env", value = .data.env)
-  #assign(".data.env", value = .data.env, envir = getUserEnv())
-  #dev-package-env
-
   .data.env
 }
 
@@ -144,7 +135,6 @@ setDataDirEnvironment <- function(env="PACKAGE") {
[…]
