### 2018-11-12 2b0a713
linted R files except polyhedra-lib.R
