+++ R/ledger-lib.R
@@ -2,7 +2,7 @@
 #' maxWithoutNA
 #' Function that returns NA if all elements are NA, and the max value not NA, if not.
 #' @param x vector parameter
-maxWithoutNA <- function(x) ifelse( !all(is.na(x)), max(x, na.rm=TRUE), NA)
+maxWithoutNA <- function(x) ifelse( !all(is.na(x)), max(x, na.rm = TRUE), NA)
 
 
 #' ScraperLedger
@@ -73,36 +73,48 @@ ScraperLedger.class <- R6::R6Class("ScraperLedger",
      r <- NULL
      default.status <- "queued"
      if (is.null(self$getIdFilename(source, source.filename))){
-       futile.logger::flog.info(paste("Adding Filename to ledger ",source, source.filename))
-       r <- nrow(self$df)+1
+       futile.logger::flog.info(paste("Adding Filename to ledger ",
+                                      source,
+                                      source.filename))
+       r <- nrow(self$df) + 1
        status.field <- "status"
-       self$countStatusUse(status.field = status.field, status = default.status)
+       self$countStatusUse(status.field = status.field,
+                           status = default.status)
        states.row <- which(self$states$status.field %in% status.field &
                              self$states$status %in% default.status)
-       file.id <- self$states[states.row,"count"]
+       file.id <- self$states[states.row, "count"]
 
        #obtain preloaded.time2scrape
        row.preloaded.t2s <- which(self$preloaded.data$source == source &
-                                    self$preloaded.data$filename==source.filename)
-       if (length(row.preloaded.t2s)==1){
-         preloaded.data <- self$preloaded.data[row.preloaded.t2s,]
-         self$df[r,c("preloaded.name")]<-
-           as.character(preloaded.data[,3])
-         self$df[r,c("preloaded.vertices","preloaded.faces","preloaded.time2scrape")]<-
-           c(preloaded.data[,4:6])
+                          self$preloaded.data$filename == source.filename)
+       if (length(row.preloaded.t2s) == 1){
+         preloaded.data <- self$preloaded.data[row.preloaded.t2s, ]
+         self$df[r, c("preloaded.name")] <-
+           as.character(preloaded.data[, 3])
+         self$df[r, c("preloaded.vertices",
+                     "preloaded.faces",
+                     "preloaded.time2scrape")] <-
+           c(preloaded.data[, 4:6])
        }
        else{
          preloaded.time2scrape <- NA
        }
-       self$df[r,c("id","source","source.filename","status")]<- c(r, source, source.filename, default.status)
+       self$df[r, c("id",
+                   "source",
+                   "source.filename",
+                   "status")] <- c(r,
+                                  source,
+                                  source.filename,
+                                  default.status)
        self$dirty <- TRUE
      }
      r
    },
    getIdFilename = function(source, source.filename){
-     r <- which(self$df$source == source & self$df$source.filename == source.filename)
-     if (length(r)>0){
-       self$df[r,"id"]
+     r <- which(self$df$source == source &
+                  self$df$source.filename == source.filename)
+     if (length(r) > 0){
+       self$df[r, "id"]
      }
      else{
        r <- NULL
@@ -110,30 +122,41 @@ ScraperLedger.class <- R6::R6Class("ScraperLedger",
      r
    },
    getCRCPolyhedronName = function(source, polyhedron.name){
-     r <- which(self$df$source == source & tolower(self$df$scraped.name) == tolower(polyhedron.name))
+     r <- which(self$df$source == source & tolower(self$df$scraped.name)
+                == tolower(polyhedron.name))
      ret <- NULL
-     if (length(r)>0){
[…]
