### 2018-11-12 8d127cd
corrections on linted files
