### 2018-11-12 cf9d665
changed all imports with importFrom to narrow down the namespace, except polyhedra-lib.R
