+++ R/db-lib.R
@@ -172,7 +172,7 @@ getPreloadedDataFilename <- function(polyhedra_preloaded_data =
 #' @usage
 #'     selectDataEnv(env=NA)
 #' @return .data.env
-#' @import futile.logger
+#' @importFrom futile.logger flog.info
 selectDataEnv <- function(env=NA) {
   if (is.na(env)) {
     if (!is.na(Sys.getenv(x = "ON_TRAVIS", unset = NA))) {
@@ -271,7 +271,9 @@ getPolyhedraObject <- function() {
 #' Downloads the files from the remote location
 #'
 #' @return TRUE if sucessfull, FALSE otherwise
-#' @import utils
+#' @importFrom utils unzip
+#' @importFrom utils download.file
+#' @importFrom utils zip
 #' @export
 downloadRPolyhedraSupportingFiles <- function(){
   if (checkDatabaseVersion() == "UPDATE") {
@@ -284,7 +286,7 @@ downloadRPolyhedraSupportingFiles <- function(){
       td <- tempdir()
       zipFile <- tempfile(tmpdir = td, fileext = ".zip")
       #download file to tempfile
-      download.file(URL, destfile = zipFile, mode = "wb")
+      utils::download.file(URL, destfile = zipFile, mode = "wb")
       utils::unzip(zipfile = zipFile, exdir = td)
       tmp.db.path <- list.files(path = td, pattern = "qbotics*")[1]
       files.to.copy <- list.files(file.path(td, tmp.db.path))
@@ -305,8 +307,7 @@ downloadRPolyhedraSupportingFiles <- function(){
 #'
 #' @param force indicate if existings directories must be overwritten
 #' @return TRUE if sucessfull
-#' @import utils
-#' @import futile.logger
+#' @importFrom futile.logger flog.info
 copyFilesToExtData <- function(force = FALSE){
   polyhedra.ledger <- getPolyhedraObject()$
     ledger$getAvailablePolyhedra(ret.fields = NULL)
@@ -373,7 +374,7 @@ copyFilesToExtData <- function(force = FALSE){
 #' \describe{
 #'   \item{\code{initialize()}}{Initializes the object}
 #' }
-#' @import     futile.logger
+#' @importFrom     futile.logger flog.info
 #' @importFrom R6 R6Class
 PolyhedronScraperConfiguration.class <- R6::R6Class(
   "PolyhedronScraperConfiguration",
@@ -409,7 +410,7 @@ PolyhedronScraperConfiguration.class <- R6::R6Class(
 #'
 #' @format \code{\link{R6Class}} object.
 #' @docType class
-#' @import futile.logger
+#' @importFrom futile.logger flog.info
 #' @importFrom R6 R6Class
 PolyhedronScraperConfigurationNetlib.class <- R6::R6Class(
   "PolyhedronScraperConfigurationNetlib",
@@ -452,7 +453,7 @@ PolyhedronScraperConfigurationNetlib.class <- R6::R6Class(
 #'
 #' @format \code{\link{R6Class}} object.
 #' @docType class
-#' @import stringr futile.logger
+#' @importFrom futile.logger flog.info
 #' @importFrom R6 R6Class
 PolyhedronScraperConfigurationDmccoey.class <- R6::R6Class(
   "PolyhedronScraperConfigurationDmccoey",
@@ -489,7 +490,7 @@ PolyhedronScraperConfigurationDmccoey.class <- R6::R6Class(
 #' get the last git commit sha
 #' @param long.version determines if the complete version of the sha will
 #'         be returned.
-#' @import git2r
+#' @importFrom git2r commits
 #' @return String with git commit sha
 #'
[…]
