+++ man/switchToFullDatabase.Rd
@@ -7,11 +7,13 @@
 switchToFullDatabase(env=NA)
 }
 \arguments{
-\item{env}{The environment to run on, can be PACKAGE, HOME or NA. If NA, it asks the user for a an Environment.}
+\item{env}{The environment to run on, can be PACKAGE,
+HOME or NA. If NA, it asks the user for a an Environment.}
 }
 \value{
 .data.env
 }
 \description{
-Prompts user for changing database to fulldb in user filespace
+Prompts user for changing database to fulldb in
+user filespace
 }
