+++ man/scrapePolyhedra.Rd
@@ -17,5 +17,6 @@ polyhedra db object
 }
 \description{
 Gets polyhedra objects from text files of
-different sources, scheduling and scraping using predefined configurations
+different sources, scheduling and scraping using
+predefined configurations
 }
