### 2018-11-12 abcd4e2
changed all imports with importFrom to narrow down the namespace, except polyhedra-lib.R
