+++ NAMESPACE
@@ -7,14 +7,19 @@ export(getAvailableSources)
 export(getPolyhedron)
 export(polyhedronToXML)
 export(switchToFullDatabase)
-import(R6)
 import(XML)
-import(digest)
 import(futile.logger)
-import(git2r)
 import(rgl)
 import(stringr)
 import(testthat)
-import(utils)
 importFrom(R6,R6Class)
+importFrom(XML,xmlTree)
+importFrom(digest,digest)
+importFrom(futile.logger,flog.info)
+importFrom(git2r,commits)
 importFrom(stats,runif)
+importFrom(utils,download.file)
+importFrom(utils,packageVersion)
+importFrom(utils,read.csv)
+importFrom(utils,unzip)
+importFrom(utils,zip)
