### 2018-11-12 7d5bf78
Implemented transformation matrix for polyhedron state
