+++ R/polyhedra-lib.R
@@ -7,7 +7,8 @@
 #'   \item{\code{addError(current.error)}}{Adds an error to the error string and log it as info}
 #'   \item{\code{scrape()}}{Scrapes the polyhedra folder files}
 #'   \item{\code{getSolid()}}{returns the object corresponding to the solid}
-#'   \item{\code{buildRGL(size = 1, origin = c(0, 0, 0), normalize.size = TRUE)}}{creates a RGL representation of the object}
+#'   \item{\code{applyTransformationMatrix(transformation.matrix)}}{Apply transformation matrix to polyhedron}
+#'   \item{\code{buildRGL(transformation.matrix)}}{creates a RGL representation of the object}
 #'   \item{\code{exportToXML()}}{Gets an XML representation out of the polyhedron object}
 #' }
 #' @field errors Errors string
@@ -38,7 +39,10 @@ getSolid = function() {
 checkEdgesConsistency=function(){
     stop(gettext("rpoly.abstract_class", domain = "R-Rpolyhedra"))
 },
-buildRGL = function(size = 1, origin = c(0, 0, 0), normalize.size = TRUE) {
+applyTransformationMatrix = function(transformation.matrix){
+  stop("Abstract class")
+},
+buildRGL = function(transformation.matrix) {
     stop(gettext("rpoly.abstract_class", domain = "R-Rpolyhedra"))
 },
 exportToXML = function(){
@@ -62,7 +66,8 @@ exportToXML = function(){
 #'   \item{\code{setupLabelsOrder()}}{Sets up the order of labels included in PHD file}
 #'   \item{\code{getDataFromLabel(label)}}{Gets data from the Label}
 #'   \item{\code{scrape()}}{Scrapes the data from the PHD file}
-#'   \item{\code{buildRGL(size = 1, origin = c(0, 0, 0), normalize.size = TRUE)}}{Builds the \code{RGL} mmodel}
+#'   \item{\code{applyTransformationMatrix(transformation.matrix)}}{Apply transformation matrix to polyhedron}
+#'   \item{\code{buildRGL(transformation.matrix)}}{Builds the \code{RGL} model}
 #' }
 #' @field netlib.p3.lines The path to the PHD files
 #' @field labels.rows Labels - row of appearance
@@ -214,8 +219,12 @@ initialize = function(file.id, netlib.p3.lines) {
                   vertices =vertices, net = net, solid = solid,
                   hinges = hinges, dih = dih)
     ret
-}, buildRGL = function(size = 1, origin = c(0, 0, 0), normalize.size = TRUE) {
-    stop(gettext("rpoly.not_implemented", domain = "R-Rpolyhedra"))
+},
+applyTransformationMatrix = function(transformation.matrix){
+  stop(gettext("rpoly.not_implemented", domain = "R-Rpolyhedra"))
+},
+buildRGL = function(transformation.matrix) {
+  stop(gettext("rpoly.not_implemented", domain = "R-Rpolyhedra"))
 },
 exportToXML = function(){
     stop(gettext("rpoly.not_implemented", domain = "R-Rpolyhedra"))
@@ -234,7 +243,8 @@ exportToXML = function(){
 #'   \item{\code{scrapeValues(values.lines)}}{Scrapes values}
 #'   \item{\code{scrapeVertices(vertices.lines)}}{Scrapes vertices}
 #'   \item{\code{scrapeFaces(face.lines)}}{Scrapes faces}
-#'   \item{\code{buildRGL(size = 1, origin = c(0, 0, 0), normalize.size = TRUE)}}{Builds the \code{RGL} implementation}
[…]
