### 2018-11-12 f6ceb23
Implemented transformation matrix for polyhedron state
