+++ man/PolyhedronStateDmccoeyScraper.class.Rd
@@ -19,7 +19,8 @@ Scrapes polyhedra from a dmccooey file format
   \item{\code{scrapeValues(values.lines)}}{Scrapes values}
   \item{\code{scrapeVertices(vertices.lines)}}{Scrapes vertices}
   \item{\code{scrapeFaces(face.lines)}}{Scrapes faces}
-  \item{\code{buildRGL(size = 1, origin = c(0, 0, 0), normalize.size = TRUE)}}{Builds the \code{RGL} implementation}
+  \item{\code{applyTransformationMatrix(transformation.matrix)}}{Apply transformation matrix to polyhedron}
+  \item{\code{buildRGL(transformation.matrix)}}{Builds the \code{RGL} model}
 }
 }
 
