### 2018-11-12 db97477
Example corrected for using transformation matrix
