### 2018-11-12 acbeefb
Example corrected for using transformation matrix
