+++ inst/doc/poly_original_help_message.Rmd
@@ -1,34 +1,0 @@
----
-title: "Original **poly** help message"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Poly - Original Help Message}
-  %\VignetteEngine{knitr::rmarkdown}
-  \usepackage[utf8]{inputenc}
----
-The original help message of the package states that:
-
-##*number*
-The polyhedron's number (written and read with the %d printf/scanf format).
-  
-##*name*
-The polyhedron's name is less than 128 characters long and is not capitalized.
-  
-##*symbol*
-The ```eqn(1)``` input for two symbols separated by a tab; the Johnson symbol, and the Schlafli
-symbol.
-  
-##*dual*
-The name of the dual polyhedron optionally followed by a horizontal tab and the number of the dual.
-  
-##*vertices*
-The first line contains the number of vertices. The vertices are arranged one per line as an (```x```, ```y```, ```z```) coordinate of white-space separated values (described below).  The vertices are implicitly numbered starting at zero.
-  
-##*net*
-The first line contains the number of faces and the maximum number of vertices in a face. The remaining lines are the faces in the planar net. Each face has a vertex count followed by the vertex numbers. The vertices are listed in counter-clockwise order as viewed from outside the polyhedron.
-  
-##*hinges*
-The first line contains the number of hinges in the planar net  The remaining lines are hinge connections. The format is ```face1``` ```side1``` ```face2``` ```side2``` ```value```.  Sides are numbered from zero. If the dihedral (```value```) is greater than J, it is a reflex or re-entrant hinge. solid the first line contains the number of faces and the maximum number of vertices in a face. The remaining lines are the faces in the 3D polyhedron.  Each face has a vertex count followed by the vertex numbers.  The vertices are listed in counter-clockwise order as viewed from outside the polyhedron.
-  
-##*dihedral*
-The first line contains the number of distinct dihedrals. Each dihedral starts on a new line and has a count and a value.  If the count is non-zero, then that many ```face edge``` pairs (one per line) follow the dihedral value.
