### 2018-11-12 ed29ab4
reopen #2 still an issue using gitr, using command line instead
