+++ NAMESPACE
@@ -1,5 +1,12 @@
 # Generated by roxygen2: do not edit by hand
 
+export(ScraperLedger.class)
+export(downloadRPolyhedraSupportingFiles)
+export(getAvailablePolyhedra)
+export(getAvailableSources)
+export(getPolyhedron)
+export(polyhedronToXML)
+export(switchToFullDatabase)
 import(XML)
 import(futile.logger)
 import(rgl)
