### 2018-11-12 ba29879
[bugfix] accessing to db.version. Always using getPackageDB
