### 2018-11-12 6991b7b
[bugfix] accessing to db.version. Always using getPackageDB
