### 2018-11-12 1513d8f
[bugfix] accessing to db.version. Always using getPackageDB
