### 2018-11-15 1cbe4ef
changes to documentation
