+++ R/polyhedra-lib.R
@@ -36,7 +36,7 @@ scrape = function() {
 getSolid = function() {
     stop(gettext("rpoly.abstract_class", domain = "R-Rpolyhedra"))
 },
-checkEdgesConsistency=function(){
+checkEdgesConsistency = function(){
     stop(gettext("rpoly.abstract_class", domain = "R-Rpolyhedra"))
 },
 applyTransformationMatrix = function(transformation.matrix){
@@ -76,33 +76,44 @@ exportToXML = function(){
 #' @docType class
 #' @import stringr futile.logger
 #' @importFrom R6 R6Class
-PolyhedronStateNetlibScraper.class <- R6::R6Class("PolyhedronStateNetlibScraper",
+PolyhedronStateNetlibScraper.class <- R6::R6Class(
+  "PolyhedronStateNetlibScraper",
 inherit = PolyhedronState.class,
-public = list(netlib.p3.lines = NA, labels.rows = NA, labels.map = NA, errors = "",
+public = list(netlib.p3.lines = NA,
+              labels.rows = NA,
+              labels.map = NA,
+              errors = "",
 initialize = function(file.id, netlib.p3.lines) {
-    super$initialize(source= "netlib",file.id = file.id)
+    super$initialize(source = "netlib", file.id = file.id)
     self$netlib.p3.lines <- netlib.p3.lines
     self$labels.map <- list()
     self
-}, extract_rows_from_label = function(label.number, expected.label) {
+},
+extract_rows_from_label = function(label.number, expected.label) {
     observer.label <- self$netlib.p3.lines[self$labels.rows[label.number]]
     observer.label <- sub("\\:", "", observer.label)
     if (observer.label != expected.label) {
-        current.error <- paste(self$errors, "In label#", label.number, "Expected label was",
-        expected.label, " and observed was", observer.label)
+        current.error <- paste(self$errors, "In label#",
+                        label.number, "Expected label was",
+                        expected.label, " and observed was", observer.label)
         self$addError(current.error)
     }
     first.data.row <- self$labels.rows[label.number]
     last.data.row <- self$labels.rows[label.number + 1]
     ret <- 0
     if (first.data.row > last.data.row) {
-        self$addError(paste("for label", expected.label, "no valid rows: (fr, lr)",
+        self$addError(paste("for label", expected.label,
+                            "no valid rows: (fr, lr)",
         first.data.row, last.data.row))
-    } else ret <- c((first.data.row + 1):(last.data.row - 1))
+    } else {
+      ret <- c( (first.data.row + 1): (last.data.row - 1))
+    }
     ret
-}, getLabels = function() {
+},
+getLabels = function() {
     self$netlib.p3.lines[self$labels.rows]
-}, scrapeNet = function(net.txt, offset = 0) {
+},
+scrapeNet = function(net.txt, offset = 0) {
     first.line <- strsplit(net.txt[1], split = " ")[[1]]
     faces <- as.numeric(first.line[1])
     max.degree <- as.numeric(first.line[2])
@@ -116,29 +127,40 @@ initialize = function(file.id, netlib.p3.lines) {
         cont <- cont + 1
     }
     net
-}, extractCFOutBrackets = function(x) {
+},
+extractCFOutBrackets = function(x) {
     open.bracket.pos <- which(strsplit(x, "")[[1]] == "[")
     ret <- x
     if (length(open.bracket.pos) > 0) {
         ret <- substr(x, 1, open.bracket.pos - 1)
     }
     ret
-}, scrapeVertices = function(vertices.txt) {
+},
+scrapeVertices = function(vertices.txt) {
     first.line <- strsplit(vertices.txt[1], split = " ")[[1]]
[…]
