+++ man/PolyhedronStateDmccoeyScraper.class.Rd
@@ -14,12 +14,14 @@ Scrapes polyhedra from a dmccooey file format
 \section{Methods}{
 
 \describe{
-  \item{\code{initialize(file.id, netlib.p3.lines)}}{Initializes the object, taking the file.id and PDH file as parameters}
+  \item{\code{initialize(file.id, netlib.p3.lines)}}{Initializes
+  the object, taking the file.id and PDH file as parameters}
   \item{\code{scrape()}}{Scrapes data from dmccooey file format}
   \item{\code{scrapeValues(values.lines)}}{Scrapes values}
   \item{\code{scrapeVertices(vertices.lines)}}{Scrapes vertices}
   \item{\code{scrapeFaces(face.lines)}}{Scrapes faces}
-  \item{\code{applyTransformationMatrix(transformation.matrix)}}{Apply transformation matrix to polyhedron}
+  \item{\code{applyTransformationMatrix(transformation.matrix)}}{Apply
+  transformation matrix to polyhedron}
   \item{\code{buildRGL(transformation.matrix)}}{Builds the \code{RGL} model}
 }
 }

[…]
