### 2018-11-15 2fb5359
added lintr to travis
