+++ .travis.yml
@@ -7,8 +7,11 @@ before_install:
   - tlmgr install index
 r_packages:
   - covr
+  - lintr
 env:
   - ON_TRAVIS = 1
 after_success:
+  - R CMD INSTALL $PKG_TARBALL
   - Rscript -e 'library(covr); codecov()'
+  - Rscript -e 'lintr::lint_package()'
 
