### 2018-11-30 475e6c5
changed 1:nrow by seq_len in db-lib.R to conform to ROpenSci rules
