### 2018-11-30 69cfa7e
Replaced 1:nrow...  for seq_along(nrow...
