### 2018-11-30 e2503df
Replaced 1:nrow...  for seq_along(nrow...
