### 2018-11-30 c952f02
[bugfix] seq_len instead of seq_along. Debugging problem when scraping faces
