### 2018-11-30 e1ed679
[bugfix] seq_len instead of seq_along. Debugging problem when scraping faces
