### 2018-11-30 b41ea92
[bugfixed] Was processing faces descriptor as a face
