### 2018-11-30 336ca31
seq_len instead of 1:nrow working
