### 2018-11-30 eb16005
seq_len instead of 1:nrow working
