### 2018-11-30 23b8ea9
replaced second vapply
