### 2018-11-30 675ec7d
changed sapply to vapply
