+++ R/db-lib.R
@@ -250,7 +250,7 @@ updatePolyhedraDatabase <- function(){
 
   #"dev-tetrahedron" "dev-minimal" "pkg-minimal" "fulldb"
   #Change when release version
-  scrapePolyhedra(scrape.config = .available.scrapping.conf[["dev-minimal"]],
+  scrapePolyhedra(scrape.config = .available.scrapping.conf[["pkg-minimal"]],
                    sources.config = .available.sources)
 }
 
@@ -957,7 +957,7 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
                                   polyhedron.file.id, source.filename){
         source <- source.config$getName()
         current.polyhedron <- NULL
-        #tryCatch({
+        tryCatch({
           self$ledger$updateStatus(source = source,
                                    source.filename = source.filename,
                                    status = "scraping")
@@ -976,16 +976,16 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
                                      source.filename = source.filename,
                                      status = "failed", obs = errors)
           }
-        # },
-        # error = function(e){
-        #   error <- paste(e$message, collapse = ",")
-        #   futile.logger::flog.error(paste("catched error", error))
-        #   assign("error", error, envir = parent.env(environment()))
-        #   self$ledger$updateStatus(source = source,
-        #                            source.filename,
-        #                            status = "exception",
-        #                            obs = error)
-        # })
+        },
+        error = function(e){
+          error <- paste(e$message, collapse = ",")
+          futile.logger::flog.error(paste("catched error", error))
+          assign("error", error, envir = parent.env(environment()))
+          self$ledger$updateStatus(source = source,
+                                   source.filename,
+                                   status = "exception",
+                                   obs = error)
+        })
         current.polyhedron
       }
       if (time2scrape.source > 0){
