### 2018-11-30 qbotics/Rpolyhedra#23 @kenarab
Make Rpolyhedra compliant with Ropensci requirements
Regarding [onboard issue](https://github.com/ropensci/onboarding/issues/157)
