### 2018-11-30 fe84f19
Examples using transformationMatrix
