### 2018-11-30 a2de0d0
Examples using transformationMatrix
