### 2018-11-30 c3f3d71
changes to conform with ropensci
