+++ NAMESPACE
@@ -1,16 +1,23 @@
 # Generated by roxygen2: do not edit by hand
 
-import(XML)
-import(futile.logger)
-import(rgl)
-import(stringr)
-import(testthat)
+export(ScraperLedger.class)
+export(downloadRPolyhedraSupportingFiles)
+export(getAvailablePolyhedra)
+export(getAvailableSources)
+export(getPolyhedron)
+export(polyhedronToXML)
+export(switchToFullDatabase)
 importFrom(R6,R6Class)
 importFrom(XML,xmlTree)
 importFrom(digest,digest)
+importFrom(futile.logger,flog.debug)
 importFrom(futile.logger,flog.info)
 importFrom(git2r,commits)
+importFrom(rgl,asHomogeneous)
+importFrom(rgl,identityMatrix)
+importFrom(rgl,transform3d)
 importFrom(stats,runif)
+importFrom(stringr,str_extract)
 importFrom(utils,download.file)
 importFrom(utils,packageVersion)
 importFrom(utils,read.csv)
