### 2018-11-30 2ca27b0
seq_len replaced 1:nrow
