### 2018-11-30 6ef2a7b
[bugfix] when no preloaded name. Included parameter for running specific source.filename scraper
