+++ R/db-lib.R
@@ -221,10 +221,11 @@ selectDataEnv <- function(env=NA) {
 #'
 #' Function for initializing database
 #'
+#' @param source.filenames if not null specify which source filenames to scrape
 #'
 #' @return polyhedra db object
 
-updatePolyhedraDatabase <- function(){
+updatePolyhedraDatabase <- function(source.filenames = NULL){
   .polyhedra <- NULL
 
   polyhedra.rds.file <- getPolyhedraRDSPath()
@@ -251,6 +252,7 @@ updatePolyhedraDatabase <- function(){
   #"dev-tetrahedron" "dev-minimal" "pkg-minimal" "fulldb"
   #Change when release version
   scrapePolyhedra(scrape.config = .available.scrapping.conf[["pkg-minimal"]],
+                  source.filenames = source.filenames,
                    sources.config = .available.sources)
 }
 
@@ -703,9 +705,9 @@ checkDatabaseVersion <- function(){
 #'   \item{\code{addPolyhedron(source,polyhedron,overwrite)}}{Adds a polyhedron
 #'   by source and name, if orverwrite is TRUE, it will update any existing one
 #'   by that source and name}
-#'   \item{\code{configPolyhedraSource(source.config, max.quant)}}{Scrapes all
+#'   \item{\code{configPolyhedraSource(source.config, source.filenames, max.quant)}}{Scrapes all
 #'   polyhedra in the given directory for adding to db or testing}
-#'   \item{\code{schedulePolyhedraSources(sources.config,max.quant,
+#'   \item{\code{schedulePolyhedraSources(sources.config, source.filenames, max.quant,
 #'   test)}}{Scrapes files applying parameter sources.config}
 #'   \item{\code{getAvailablePolyhedra(sources,search.string)}}{Retrieves
 #'   all polyhedron within the source those names match with search.string}
@@ -851,13 +853,16 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
                                scraped.polyhedron = polyhedron)
       polyhedron
     },
-    configPolyhedraSource = function(source.config, max.quant = 0) {
+    configPolyhedraSource = function(source.config, source.filenames= NULL, max.quant = 0) {
       source <- source.config$getName()
       home.dir.data <- getDataDir()
       self$configPolyhedraRDSPath()
       futile.logger::flog.debug(paste("configuring source", source))
       polyhedra.dir   <- source.config$getBaseDir(home.dir.data)
       polyhedra.files <- source.config$getPolyhedraFiles(home.dir.data)
+      if (!is.null(source.filenames)){
+        polyhedra.files <- polyhedra.files[polyhedra.files %in% source.filenames]
+      }
       if (length(polyhedra.files) > 0) {
         if (max.quant > 0){
           polyhedra.files <- polyhedra.files[1:min(max.quant,
@@ -957,7 +962,7 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
                                   polyhedron.file.id, source.filename){
         source <- source.config$getName()
         current.polyhedron <- NULL
-        tryCatch({
+        #tryCatch({
           self$ledger$updateStatus(source = source,
                                    source.filename = source.filename,
                                    status = "scraping")
@@ -976,16 +981,16 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
                                      source.filename = source.filename,
                                      status = "failed", obs = errors)
           }
-        },
-        error = function(e){
-          error <- paste(e$message, collapse = ",")
-          futile.logger::flog.error(paste("catched error", error))
-          assign("error", error, envir = parent.env(environment()))
-          self$ledger$updateStatus(source = source,
-                                   source.filename,
-                                   status = "exception",
[…]
