+++ NAMESPACE
@@ -1,12 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-export(ScraperLedger.class)
-export(downloadRPolyhedraSupportingFiles)
-export(getAvailablePolyhedra)
-export(getAvailableSources)
-export(getPolyhedron)
-export(polyhedronToXML)
-export(switchToFullDatabase)
 importFrom(R6,R6Class)
 importFrom(XML,xmlTree)
 importFrom(digest,digest)
