### 2018-11-30 26c2d92
[bugfix] when no preloaded name. Included parameter for running specific source.filename scraper
