### 2018-11-30 b6f46a6
[bugfix] When no solid definition or other errors in polyhedron does not continue scraping process
