### 2018-11-30 4ca56d8
[bugfix] When no solid definition or other errors in polyhedron does not continue scraping process
