### 2018-11-30 7130be3
[bugfix] Check when no solid definition
