### 2018-11-30 ccb8284
Updating ledger info=>debug
