### 2018-11-30 ea3f5e8
After lintr
