### 2018-11-30 qbotics/Rpolyhedra#23 @kenarab
All required changes solved in ea3f5e8e2adc288f5d558eaf3438201369519760
