### 2018-11-30 #157 @kenarab
Hello @noamross. We are finally uploading a new version with all observations you made implemented.

Here is a [link to the current release](https://github.com/qbotics/Rpolyhedra/releases/tag/v0.3.0) (v0.3.0)

We included a function `switchToFullDatabase()` for switching to fullDB, which downloads from a separate repository. By default, the packages installs with a minimal db with 120 polyhedra. There are many configurations possible for users but also for development proposes.
All polyhedra is scraped in its own serialized file, so different packing strategies could be implemented in a future, including a scrape-on-demand, for example.

Other interesting feature that worth mentioning is  polyhedra RGL models are now modified with transformationMatrix, so common strategies for managing openGL models can be applied.

The rest of modifications are technical, We think V0.3.0 are a good starting point for reviewers and doesn't worth for reviewers to check it out.

Thank you very much for your time and guidance. We hope the package is Ropensci compliant, and we can finish the process to publish within the project.

About JOSS, we didn't upload any paper because we don't think the contribution in this package has to be published in a peer reviewed paper at this point on time.
Best,

Leonardo & Alejandro.
