### 2018-12-01 4d9e423
Corrections in README.md
