### 2018-12-03 3a035ff
Added getName to state objects
