### 2018-12-03 0fe1066
Added getName to state objects
