+++ man/PolyhedronStateDmccoeyScraper.class.Rd
@@ -17,6 +17,7 @@ Scrapes polyhedra from a dmccooey file format
   \item{\code{initialize(file.id, netlib.p3.lines)}}{Initializes
   the object, taking the file.id and PDH file as parameters}
   \item{\code{scrape()}}{Scrapes data from dmccooey file format}
+  \item{\code{getName()}}{returns polyhedron name}
   \item{\code{scrapeValues(values.lines)}}{Scrapes values}
   \item{\code{scrapeVertices(vertices.lines)}}{Scrapes vertices}
   \item{\code{scrapeFaces(face.lines)}}{Scrapes faces}
