### 2018-12-03 #157 @noamross
Thank you @kenarab.  I see that your overall test coverage [is still low at about 30%](https://codecov.io/github/qbotics/Rpolyhedra).  Is there a reason for this? Sometimes this occurs because some of the package-building code ends up in the package itself.  But coverage less than 75% or so is problematic.
