### 2018-12-03 c261910
first draft of xml support added #15
