### 2018-12-03 87c6be1
first draft of xml support added #15
