### 2018-12-03 86893c3
Revert "Add code coverage badge to README"
