+++ tests/testthat/test_rpolyhedra_lib.R
@@ -95,5 +95,3 @@ context("XML export")
 test_that("test xml can be exported", {
   polyhedronToXML(getPolyhedron("netlib", "cube")$getState())
   })
-
-
