### 2018-12-06 #157 @kenarab
Hi Noam.

The code not covered is infrastructure for the core functionality. That’s why we didn’t included it in test cases, but of course it is a good practice to be covered.

We are coding the new test cases in the next days.

We are now in ic18 conference presenting Rpolyhedra.

Thank you,
All the best,
Ale.

El 3 dic. 2018, a la(s) 13:05, Noam Ross <notifications@github.com> escribió:

> Thank you @kenarab. I see that your overall test coverage is still low at about 30%. Is there a reason for this? Sometimes this occurs because some of the package-building code ends up in the package itself. But coverage less than 75% or so is problematic.
>
