### 2018-12-21 f236140
[bugfix] When running testcase for db, works in pretend mode
