+++ R/db-lib.R
@@ -249,7 +249,7 @@ updatePolyhedraDatabase <- function(source.filenames = NULL){
 
   #"dev-tetrahedron" "dev-minimal" "pkg-minimal" "fulldb"
   #Change when release version
-  scrapePolyhedra(scrape.config = .available.scrapping.conf[["pkg-minimal"]],
+  scrapePolyhedra(scrape.config = .available.scrapping.conf[["dev-minimal"]],
                   source.filenames = source.filenames,
                    sources.config = .available.sources)
 }
@@ -808,12 +808,13 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
       ret
     },
     addPolyhedron = function(source="netlib", source.filename,
-                             polyhedron, overwrite = FALSE) {
+                             polyhedron, overwrite = FALSE,
+                             pretend = TRUE) {
       polyhedron.name <- polyhedron$getName()
       data.dir <- self$getPolyhedraSourceDir(source = source)
       prev.data <- self$getPolyhedron(source = source,
                                       polyhedron.name = polyhedron.name)
-      if (!overwrite & !is.null(prev.data)){
+      if (!overwrite & !is.null(prev.data) & !pretend){
         futile.logger::flog.info(paste("Polyhedron",
                                        polyhedron.name,
                                        "in source",
@@ -821,29 +822,31 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
                                        "already in database"))
       }
       else {
-        crc.name <- self$ledger$getCRCPolyhedronName(source = source,
-                          polyhedron.name = polyhedron.name)
-        serialized.polyhedron <- polyhedron$state$serialize()
-        tmp.dir <- file.path(tempdir = tempdir(), source)
-        dir.create(tmp.dir, showWarnings = FALSE, recursive = TRUE)
-        serialized.filename <- paste(crc.name, ".RDS", sep = "")
-        tmp.filename <- file.path(tmp.dir, serialized.filename)
-        saveRDS(object = serialized.polyhedron, ascii = TRUE,
-                file = tmp.filename)
-        zip(zipfile = file.path(data.dir, paste(crc.name,
-                                                ".RDS.zip",
-                                                sep = "")),
-            files = tmp.filename, flags = "-j")
-        unlink(tmp.filename)
-        futile.logger::flog.info(paste("Added polyhedron in file",
-                                       polyhedron.name,
-                                       "#|n",
-                                       polyhedron$file.id,
-                                       polyhedron.name,
-                                       "in source",
-                                       source,
-                                       "to database with CRC",
-                                       crc.name))
+        if (!pretend){
+          crc.name <- self$ledger$getCRCPolyhedronName(source = source,
+                                                       polyhedron.name = polyhedron.name)
+          serialized.polyhedron <- polyhedron$state$serialize()
+          tmp.dir <- file.path(tempdir = tempdir(), source)
+          dir.create(tmp.dir, showWarnings = FALSE, recursive = TRUE)
+          serialized.filename <- paste(crc.name, ".RDS", sep = "")
+          tmp.filename <- file.path(tmp.dir, serialized.filename)
+          saveRDS(object = serialized.polyhedron, ascii = TRUE,
+                  file = tmp.filename)
+          zip(zipfile = file.path(data.dir, paste(crc.name,
+                                                  ".RDS.zip",
+                                                  sep = "")),
+              files = tmp.filename, flags = "-j")
+          unlink(tmp.filename)
+          futile.logger::flog.info(paste("Added polyhedron in file",
+                                         polyhedron.name,
+                                         "#|n",
+                                         polyhedron$file.id,
+                                         polyhedron.name,
+                                         "in source",
+                                         source,
[…]
