### 2018-12-21 8dfecf2
[bugfix] When running testcase for db, works in pretend mode
