### 2018-12-21 f630d77
added a graceful download failure message
