### 2018-12-22 31f742c
[bugfix] with null symbols when no symbol label in netlib. More coverage on testcases
