### 2018-12-22 acf4bec
expect_equal in rgl context
