### 2018-12-24 cbb5a1e
download files now has better failover messages
