+++ tests/testthat/test_db_lib.R
@@ -13,10 +13,10 @@ test_that("test on db lib functions", {
   testthat::expect(!is.null(getPackageDir()), failure_message = "getPackageDir cannot be null")
   testthat::expect(!is.null(getPolyhedraRDSPath()), failure_message = "getPolyhedraRDSPath cannot be null")
   testthat::expect(!is.null(getPreloadedDataFilename()), failure_message = "getPreloadedDataFilename cannot be null")
-  testthat::expect_equal(selectDataEnv(env = "PACKAGE"), "PACKAGE")
+  testthat::expect_equal(selectDataEnv(env = "PACKAGE"), "SUCCESS")
   testthat::expect(!is.null(updatePolyhedraDatabase()), failure_message = "updatePolyhedraDatabase cannot be null")
   testthat::expect(!is.null(getPolyhedraObject()), failure_message = "getPolyhedraObject cannot be null")
-  testthat::expect(downloadRPolyhedraSupportingFiles()==TRUE)
+  testthat::expect(downloadRPolyhedraSupportingFiles()=="SUCCESS")
   testthat::expect_error(copyFilesToExtData(FALSE))
   testthat::expect(!is.null(getGitCommit()))
   testthat::expect(!is.null(getPackageVersion()))
