### 2018-12-24 a6b5af9
download files now has better failover messages
