### 2018-12-24 b741b4d
Code reorganization
