+++ R/public-lib.R
@@ -1,0 +1,185 @@
+#' scrapePolyhedra()
+#'
+#' Gets polyhedra objects from text files of
+#' different sources, scheduling and scraping using
+#' predefined configurations
+#'
+#' @param scrape.config predefined configuration for scraping
+#' @param source.filenames if not null specify which source filenames to scrape
+#' @param sources.config the sources that will be used by the function
+#' @return polyhedra db object
+scrapePolyhedra <- function(scrape.config,
+                            source.filenames = NULL,
+                            sources.config = getPackageEnvir(".available.sources")){
+  scrapePolyhedraSources(max.quant.config.schedule =
+                           scrape.config[["max.quant.config.schedule"]],
+                         max.quant.scrape = scrape.config[["max.quant.scrape"]],
+                         time2scrape.source = scrape.config[["time2scrape.source"]],
+                         sources.config = sources.config,
+                         source.filenames = source.filenames,
+                         retry.scrape = scrape.config[["retry.scrape"]])
+
+}
+
+#' scrapePolyhedraSources()
+#'
+#' Method for obtaining polyhedra objects from text files of
+#' different sources, scheduling and scraping
+#'
+#' @param sources.config the sources that will be used by the function
+#' @param max.quant.config.schedule number of files to schedule
+#' @param max.quant.scrape number of files scrape
+#' @param time2scrape.source time applied to scrape source
+#' @param source.filenames if not null specify which source filenames to scrape
+#' @param retry.scrape should it retry scrape?
+#' @return polyhedra db object
+#' @usage
+#'     scrapePolyhedraSources(sources.config =
+#'          getPackageEnvir(".available.sources"),
+#'     max.quant.config.schedule = 0,
+#'     max.quant.scrape = 0, time2scrape.source = 30,
+#'     source.filenames = NULL, retry.scrape = FALSE)
+scrapePolyhedraSources <- function(sources.config =
+                                     getPackageEnvir(".available.sources"),
+                                   max.quant.config.schedule = 0,
+                                   max.quant.scrape = 0,
+                                   time2scrape.source = 30,
+                                   source.filenames = NULL,
+                                   retry.scrape = FALSE){
+  futile.logger::flog.debug(paste("Scheduling",
+                                  max.quant.config.schedule,
+                                  "polyhedra for scraping"))
+  getPolyhedraObject()$schedulePolyhedraSources(sources.config =
+                                                  sources.config,
+                                                source.filenames =
+                                                  source.filenames,
+                                                max.quant = max.quant.config.schedule,
+                                                pretend = FALSE)
+  if (retry.scrape){
+    mode <- "scrape.retry"
+  }
+  else {
+    mode <- "scrape.queued"
+  }
+  futile.logger::flog.debug(paste("Scraping",
+                                  max.quant.scrape,
+                                  "polyhedra up to",
+                                  time2scrape.source, "seconds"))
+  getPolyhedraObject()$scrape(mode = mode,
+                              max.quant = max.quant.scrape,
+                              time2scrape.source = time2scrape.source,
+                              pretend = FALSE)
+  #All files not scraped in building, marked as skipped
+  still.queued <- which(getPolyhedraObject()$ledger$df$status == "queued")
+  if (length(still.queued) > 0){
+    apply(getPolyhedraObject()$ledger$df[still.queued, ], MARGIN = 1,
+          FUN = function(x) {
[…]
