+++ man/scrapePolyhedra.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/db-lib.R
+% Please edit documentation in R/public-lib.R
 \name{scrapePolyhedra}
 \alias{scrapePolyhedra}
 \title{scrapePolyhedra()}
