### 2018-12-24 4680d5a
Code reorganization
