+++ R/env-lib.R
@@ -1,0 +1,279 @@
+#' getPackageEnvir
+#'
+#' @param variable.name  name of variable to be retrieved
+
+#' Obtains a variable from package environment
+getPackageEnvir <- function(variable.name){
+  get(variable.name, envir = asNamespace("Rpolyhedra"))
+}
+
+#' setPackageEnvir
+#'
+#' Set a variable from package environment
+#' @param variable.name  name of variable to be set
+#' @param value          variable value
+
+setPackageEnvir <- function(variable.name, value){
+  assign(x = variable.name, value = value, envir = asNamespace("Rpolyhedra"))
+}
+
+
+#' getUserEnvir
+#'
+#' Gets a writable environment for the package
+#'
+#' @param variable.name  name of variable to be retrieved
+getUserEnvir <- function(variable.name) {
+  get(x = variable.name, envir = getPackageEnvir("RpolyhedraEnv"))
+}
+
+#' setUserEnvir
+#'
+#' Sets variable in the writable environment for the package
+#'
+#' @param variable.name  name of variable to be set
+#' @param value          variable value
+
+setUserEnvir <- function(variable.name, value) {
+  assign(x = variable.name, value = value, envir =
+           getPackageEnvir("RpolyhedraEnv"))
+}
+
+
+#' getDataEnv
+#'
+#' Gets the current .data.env value
+#'
+#' @return .data.env
+getDataEnv <- function() {
+  getUserEnvir(".data.env")
+}
+
+
+#' getUserSpace
+#'
+#' This function is used internally for accesing the local database path
+#' @return path of user space
+getUserSpace <- function(){
+  file.path(path.expand("~"), ".R", "Rpolyhedra")
+}
+
+#' initDataDirEnvironment
+#'
+#' initialize data enviornment
+#'
+#' @return the data dir environment
+initDataDirEnvironment <- function() {
+  environment.filepath <- getEnvironmentFilepath()
+  if (!file.exists(environment.filepath)){
+    .data.env <- "PACKAGE"
+  } else {
+    .data.env <- readLines(environment.filepath)[1]
+  }
+
+  setUserEnvir(".data.env", value = .data.env)
+  .data.env
+}
+
+
+
+
+#' getDataDir
+#'
+#' Gets the path of Rpolyhedra data dir.
+#'
+#' This function is used internally to determine whether the package
+#' is compiled in source or package directory.
+#' @param data.env enviroment where data directory must be returned
+#' @return dir where the package access polyhedra database
+
+getDataDir <- function(data.env = getDataEnv()) {
+  data.dir <- ""
+  if (data.env == "HOME") {
+    data.dir <- getUserSpace()
+  } else {
+    data.dir <- getPackageDir()
+  }
+  data.dir
+}
+
+#' getEnvironmentFilepath
+#'
+#' Gets the filename where package data environment is persisted
+#' @return The environment filepath
+
+getEnvironmentFilepath <- function(){
+  file.path(getDataDir("HOME"), "Rpolyhedra.env")
+}
+
+#' setDataDirEnvironment
+#'
+#' Sets the path of Rpolyhedra data dir.
+#'
+#' This function is used to set the data directories either to the package or the user home directory.
+#'
+#' @param env The type of environment to work with. Values are "PACKAGE" or "HOME" and it defaults to package
+#' @return the curruent .data.env
+setDataDirEnvironment <- function(env = "PACKAGE") {
+  if (env %in% c("PACKAGE", "HOME")) {
+    .data.env <- env
+  }
+  else {
[…]
