+++ R/package-lib.R
@@ -1,0 +1,379 @@
+#' getPreloadedDataFilename
+#'
+#' Gets the path of Polyhedra preloaded data CSV file
+#'
+#' @param polyhedra_preloaded_data filename of polyhedra preloaded data csv
+#' @return the path to the Polyhedra database file
+getPreloadedDataFilename <- function(polyhedra_preloaded_data =
+                                       "polyhedra.preloaded.data.csv"){
+  file.path(getDataDir(), polyhedra_preloaded_data)
+}
+
+#' updatePolyhedraDatabase
+#'
+#' Function for initializing database
+#'
+#' @param source.filenames if not null specify which source filenames to scrape
+#'
+#' @return polyhedra db object
+
+updatePolyhedraDatabase <- function(source.filenames = NULL){
+  .polyhedra <- NULL
+
+  polyhedra.rds.file <- getPolyhedraRDSPath()
+
+
+  if (file.exists(polyhedra.rds.file)) {
+    polyhedra.candidate <- readRDS(polyhedra.rds.file)
+    if (isCompatiblePolyhedraRDS(polyhedra.candidate, halts = TRUE)){
+      .polyhedra <- polyhedra.candidate
+    }
+  }
+  if (is.null(.polyhedra)){
+    .polyhedra <- PolyhedraDatabase.class$new()
+  }
+
+  setUserEnvir(".polyhedra", value = .polyhedra)
+
+  .available.sources <- getPackageEnvir(".available.sources")
+  .available.scrapping.conf <- getPackageEnvir(".available.scrapping.conf")
+
+
+  #"dev-tetrahedron" "dev-minimal" "pkg-minimal" "fulldb"
+  #Change when release version
+  scrapePolyhedra(scrape.config = .available.scrapping.conf[["pkg-minimal"]],
+                  source.filenames = source.filenames,
+                  sources.config = .available.sources)
+}
+
+#' downloadRPolyhedraSupportingFiles
+#'
+#' Downloads the files from the remote location
+#'
+#' @return TRUE if sucessfull, FALSE otherwise
+#' @importFrom utils unzip
+#' @importFrom utils download.file
+#' @importFrom utils zip
+#' @export
+downloadRPolyhedraSupportingFiles <- function(){
+  retVal <- "SUCCESS"
+  if (checkDatabaseVersion() == "UPDATE") {
+    if (getDataEnv() == "HOME"){
+      db.version <- getPackageDB()
+      URL <- paste(
+        "https://api.github.com/repos/qbotics/RpolyhedraDB/zipball/v",
+        db.version,
+        sep = "")
+      td <- tempdir()
+      zipFile <- tempfile(tmpdir = td, fileext = ".zip")
+      #download file to tempfile
+      oldw <- getOption("warn")
+      options(warn = -1)
+      retVal <-  tryCatch({
+        utils::download.file(URL, destfile = zipFile, mode = "wb")
+        "SUCCESS"
+      }, error = function(e) {
+        "NOT_AVAILABLE"
+      })
+      options(warn = oldw)
+      if(retVal  == "SUCCESS")
+      {
+        utils::unzip(zipfile = zipFile, exdir = td)
+        tmp.db.path <- list.files(path = td, pattern = "qbotics*")[1]
+        files.to.copy <- list.files(file.path(td, tmp.db.path))
+        #copy files
+        file.copy(from = file.path(td, tmp.db.path, files.to.copy),
+                  to = getUserSpace(), recursive = TRUE)
+        #delete tmp path
+        unlink(file.path(td, tmp.db.path), recursive = TRUE)
+      }
+
+    }
+  }
+  retVal
+}
+
+#' copyFilesToExtData
+#'
[…]
