+++ tests/testthat.R
@@ -5,4 +5,7 @@ library(futile.logger)
 library(rgl)
 
 
+#Change threshold to ERROR. Comment out/change if verbosity required for development
+futile.logger::flog.threshold(futile.logger::ERROR)
+
 test_check("Rpolyhedra")
