### 2018-12-24 b77d211
linted R/ files
