### 2018-12-24 1cda363
linted R/ files
