+++ man/getDatabaseVersion.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/env-lib.R
+% Please edit documentation in R/package-lib.R
 \name{getDatabaseVersion}
 \alias{getDatabaseVersion}
 \title{getDatabaseVersion}
