### 2018-12-24 b8b4a37
Beter coverage in public-lib. Pushed funcionality to db
