### 2018-12-24 b2eaf1d
Beter coverage in public-lib. Pushed funcionality to db
