### 2018-12-24 ce9b9e9
Commented mocks in test package
