### 2018-12-24 d2ccef6
More coverage on db-lib
