+++ tests/testthat.R
@@ -6,7 +6,6 @@ library(rgl)
 
 
 #Change threshold to ERROR. Comment out/change if verbosity required for development
-#futile.logger::flog.threshold(futile.logger::ERROR)
-futile.logger::flog.threshold(futile.logger::INFO)
+futile.logger::flog.threshold(futile.logger::ERROR)
 
 test_check("Rpolyhedra")
