### 2018-12-24 bb98b68
mockup fixed for test_package_lib.R
