### 2018-12-24 0c9f515
A little more test coverage
