+++ NAMESPACE
@@ -4,6 +4,7 @@ export(ScraperLedger.class)
 export(downloadRPolyhedraSupportingFiles)
 export(getAvailablePolyhedra)
 export(getAvailableSources)
+export(getPolyhedraObject)
 export(getPolyhedron)
 export(polyhedronToXML)
 export(switchToFullDatabase)
