### 2018-12-24 a4f19e7
A little more test coverage
