+++ tests/testthat.R
@@ -7,5 +7,5 @@ library(rgl)
 
 #Change threshold to ERROR. Comment out/change if verbosity required for development
 futile.logger::flog.threshold(futile.logger::ERROR)
-
+c
 test_check("Rpolyhedra")
