### 2018-12-24 7f7a560
Trying to extend coverage on db-lib
