### 2018-12-26 2920da6
lintr code
