+++ tests/testthat/test_package_lib.R
@@ -1,15 +1,21 @@
 context("package-lib")
 test_that("test on package lib functions", {
-  testthat::expect(!is.null(getPreloadedDataFilename()), failure_message = "getPreloadedDataFilename cannot be null")
-  testthat::expect(!is.null(updatePolyhedraDatabase()), failure_message = "updatePolyhedraDatabase cannot be null")
+  testthat::expect(!is.null(getPreloadedDataFilename()),
+                   failure_message = "getPreloadedDataFilename cannot be null")
+  testthat::expect(!is.null(updatePolyhedraDatabase()),
+                   failure_message = "updatePolyhedraDatabase cannot be null")
   with_mock(
-       "Rpolyhedra::getUserSpace" = function(){file.path(getUserSpace(),".tmp/")},
+       "Rpolyhedra::getUserSpace" = function(){
+            file.path(getUserSpace(),".tmp/")
+         },
        testthat::expect(
        with_mock(
-         "Rpolyhedra::getDataEnv" = function(){"HOME"},
+         "Rpolyhedra::getDataEnv" = function(){
+           "HOME"},
          with_mock(
            "Rpolyhedra::checkDatabaseVersion" = function(){"UPDATE"},
-           downloadRPolyhedraSupportingFiles() %in% c("SUCCESS", "NOT_AVAILABLE")
+           downloadRPolyhedraSupportingFiles() %in%
+                  c("SUCCESS", "NOT_AVAILABLE")
      )), failure_message = "downloadRPolyhedraSupportingFiles error"))
 
    with_mock(
