### 2018-12-26 460878e
lintr code
