+++ tests/testthat/test_serialization_lib.R
@@ -1,5 +1,7 @@
 context("serialization-lib XML export")
 test_that("test xml can be exported", {
-  xml.polyhedron <- polyhedronToXML(getPolyhedron("netlib", "tetrahedron")$getState())
-  expect_equal(as.numeric(nchar(XML::getChildrenStrings(xml.polyhedron))), 406)
+  xml.polyhedron <- polyhedronToXML(
+          getPolyhedron("netlib", "tetrahedron")$getState())
+  expect_equal(
+    as.numeric(nchar(XML::getChildrenStrings(xml.polyhedron))), 406)
 })
