+++ R/db-lib.R
@@ -27,13 +27,22 @@
 #'   the polyhedron exists on the database}
 #'   \item{\code{getPolyhedron(source, polyhedron.name, strict)}}{Retrieves
 #'   a polyhedron by source and name}
-#'   \item{\code{addPolyhedron(source,polyhedron,overwrite)}}{Adds a polyhedron
+#'   \item{\code{addPolyhedron(source, polyhedron, overwrite, saveOnChange = FALSE)}}{Adds a polyhedron
 #'   by source and name, if orverwrite is TRUE, it will update any existing one
 #'   by that source and name}
 #'   \item{\code{configPolyhedraSource(source.config, source.filenames, max.quant)}}{Scrapes all
 #'   polyhedra in the given directory for adding to db or testing}
 #'   \item{\code{schedulePolyhedraSources(sources.config, source.filenames, max.quant,
 #'   test)}}{Scrapes files applying parameter sources.config}
+#'   \item{\code{cover(sources, covering.code, polyhedra.names = NULL,
+#'                     max.quant = 0, saveOnChange = FALSE, seed = NULL)}{Cover all polyhedron with specified code}
+#'   \item{\code{scrape(mode = "scrape.queued",
+#'                      sources = names(self$sources.config),
+#'                      max.quant = 0, time2scrape.source = 30,
+#'                      saveOnChange = FALSE, skip.still.queued = FALSE)}
+#'                      {Scrape file with specified parameters}
+#'
+#'   \item{\code{saveRDS = function(saveOnChange = TRUE)}{ Save state in file when specified}
 #'   \item{\code{getAvailablePolyhedra(sources,search.string)}}{Retrieves
 #'   all polyhedron within the source those names match with search.string}
 #' }
@@ -136,12 +145,12 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
     },
     addPolyhedron = function(source="netlib", source.filename,
                              polyhedron, overwrite = FALSE,
-                             pretend = TRUE) {
+                             saveOnChange = FALSE) {
       polyhedron.name <- polyhedron$getName()
       data.dir <- self$getPolyhedraSourceDir(source = source)
       prev.data <- self$getPolyhedron(source = source,
                                       polyhedron.name = polyhedron.name)
-      if (!overwrite & !is.null(prev.data) & !pretend){
+      if (!overwrite & !is.null(prev.data) & !saveOnChange){
         futile.logger::flog.info(paste("Polyhedron",
                                      polyhedron.name,
                                      "in source",
@@ -156,7 +165,7 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
         dir.create(tmp.dir, showWarnings = FALSE, recursive = TRUE)
         serialized.filename <- paste(crc.name, ".RDS", sep = "")
         tmp.filename <- file.path(tmp.dir, serialized.filename)
-        if (!pretend){
+        if (!saveOnChange){
           saveRDS(object = serialized.polyhedron, ascii = TRUE,
                   file = tmp.filename)
           zip(zipfile = file.path(data.dir, paste(crc.name,
@@ -165,7 +174,7 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
               files = tmp.filename, flags = "-j")
           unlink(tmp.filename)
         }
-        futile.logger::flog.info(paste(ifelse(pretend,"[pretend]",""),
+        futile.logger::flog.info(paste(ifelse(saveOnChange, "[saveOnChange]", ""),
                                        "Added polyhedron in file",
                                        polyhedron.name,
                                        "#|n",
@@ -184,7 +193,7 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
     },
[…]
