+++ R/env-lib.R
@@ -237,4 +237,3 @@ getGitCommit <- function(long.version = FALSE){
   git.sha
 }
 
-
