### 2018-12-26 e083ae8
lintr and improved package coverage
