+++ R/db-lib.R
@@ -27,7 +27,7 @@
 #'   the polyhedron exists on the database}
 #'   \item{\code{getPolyhedron(source, polyhedron.name, strict)}}{Retrieves
 #'   a polyhedron by source and name}
-#'   \item{\code{addPolyhedron(source, polyhedron, overwrite, saveOnChange = FALSE)}}{Adds a polyhedron
+#'   \item{\code{addPolyhedron(source, polyhedron, overwrite, save.on.change = FALSE)}}{Adds a polyhedron
 #'   by source and name, if orverwrite is TRUE, it will update any existing one
 #'   by that source and name}
 #'   \item{\code{configPolyhedraSource(source.config, source.filenames, max.quant)}}{Scrapes all
@@ -35,14 +35,14 @@
 #'   \item{\code{schedulePolyhedraSources(sources.config, source.filenames, max.quant,
 #'   test)}}{Scrapes files applying parameter sources.config}
 #'   \item{\code{cover(sources, covering.code, polyhedra.names = NULL,
-#'                     max.quant = 0, saveOnChange = FALSE, seed = NULL)}{Cover all polyhedron with specified code}
+#'                     max.quant = 0, save.on.change = FALSE, seed = NULL)}{Cover all polyhedron with specified code}
 #'   \item{\code{scrape(mode = "scrape.queued",
 #'                      sources = names(self$sources.config),
 #'                      max.quant = 0, time2scrape.source = 30,
-#'                      saveOnChange = FALSE, skip.still.queued = FALSE)}
+#'                      save.on.change = FALSE, skip.still.queued = FALSE)}
 #'                      {Scrape file with specified parameters}
 #'
-#'   \item{\code{saveRDS = function(saveOnChange = TRUE)}{ Save state in file when specified}
+#'   \item{\code{saveRDS = function(save.on.change = TRUE)}{ Save state in file when specified}
 #'   \item{\code{getAvailablePolyhedra(sources,search.string)}}{Retrieves
 #'   all polyhedron within the source those names match with search.string}
 #' }
@@ -145,12 +145,12 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
     },
     addPolyhedron = function(source="netlib", source.filename,
                              polyhedron, overwrite = FALSE,
-                             saveOnChange = FALSE) {
+                             save.on.change = FALSE) {
       polyhedron.name <- polyhedron$getName()
       data.dir <- self$getPolyhedraSourceDir(source = source)
       prev.data <- self$getPolyhedron(source = source,
                                       polyhedron.name = polyhedron.name)
-      if (!overwrite & !is.null(prev.data) & !saveOnChange){
+      if (!overwrite & !is.null(prev.data) & !save.on.change){
         futile.logger::flog.info(paste("Polyhedron",
                                      polyhedron.name,
                                      "in source",
@@ -165,7 +165,7 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
         dir.create(tmp.dir, showWarnings = FALSE, recursive = TRUE)
         serialized.filename <- paste(crc.name, ".RDS", sep = "")
         tmp.filename <- file.path(tmp.dir, serialized.filename)
-        if (!saveOnChange){
+        if (!save.on.change){
           saveRDS(object = serialized.polyhedron, ascii = TRUE,
                   file = tmp.filename)
           zip(zipfile = file.path(data.dir, paste(crc.name,
@@ -174,7 +174,8 @@ PolyhedraDatabase.class <- R6::R6Class("PolyhedraDatabase",
               files = tmp.filename, flags = "-j")
           unlink(tmp.filename)
         }
-        futile.logger::flog.info(paste(ifelse(saveOnChange, "[saveOnChange]", ""),
[…]
