### 2018-12-26 9a04f08
lintr and improved package coverage
