+++ R/env-lib.R
@@ -236,4 +236,3 @@ getGitCommit <- function(long.version = FALSE){
   }
   git.sha
 }
-
