+++ NAMESPACE
@@ -1,5 +1,13 @@
 # Generated by roxygen2: do not edit by hand
 
+export(ScraperLedger.class)
+export(downloadRPolyhedraSupportingFiles)
+export(getAvailablePolyhedra)
+export(getAvailableSources)
+export(getPolyhedraObject)
+export(getPolyhedron)
+export(polyhedronToXML)
+export(switchToFullDatabase)
 import(XML)
 importFrom(R6,R6Class)
 importFrom(XML,newXMLNode)
