### 2018-12-26 79c6601
[bugfix] rds was not saving. More coverage and check, goodpractices, lintr, etc.
