+++ tests/testthat/test_package_lib.R
@@ -1,10 +1,10 @@
 
 context("package-lib")
 
-setup({tmp.package.dir <<-
-        file.path(getDataDir(data.env = "PACKAGE"), ".tmp/")
-       tmp.home.dir <<- file.path(getUserSpace(), ".tmp/")
-       })
+setup(tmp.package.dir <<-
+          file.path(getDataDir(data.env = "PACKAGE"), ".tmp/"))
+setup(tmp.home.dir <<- file.path(getUserSpace(), ".tmp/"))
+
 
 test_that("test on package lib functions", {
   testthat::expect(!is.null(getPreloadedDataFilename()),
@@ -31,7 +31,8 @@ test_that("test on package lib functions", {
      )), failure_message = "downloadRPolyhedraSupportingFiles error"))
 
   testthat::expect_equal(copyFilesToExtData(force = FALSE,
-                                            source.folder = getDataDir(data.env =  "HOME"),
+                                            source.folder =
+                                              getDataDir(data.env =  "HOME"),
                                             dest.folder = tmp.package.dir),
                                                    TRUE)
 
@@ -43,9 +44,6 @@ test_that("test on package lib functions", {
   testthat::expect(switchToFullDatabase(env = "PACKAGE") == "PACKAGE")
 })
 
-teardown({
-  unlink(tmp.package.dir, recursive = TRUE)
-  unlink(tmp.home.dir, recursive = TRUE)
-  rm("tmp.package.dir")
-  rm("tmp.home.dir")
-})
+teardown(unlink(tmp.package.dir, recursive = TRUE))
+teardown(unlink(tmp.home.dir, recursive = TRUE))
+
