### 2018-12-26 9d301f2
[bugfix] rds was not saving. More coverage and check, goodpractices, lintr, etc.
