+++ NAMESPACE
@@ -8,8 +8,8 @@ export(getPolyhedraObject)
 export(getPolyhedron)
 export(polyhedronToXML)
 export(switchToFullDatabase)
-import(XML)
 importFrom(R6,R6Class)
+importFrom(XML,newXMLDoc)
 importFrom(XML,newXMLNode)
 importFrom(digest,digest)
 importFrom(futile.logger,flog.debug)
