### 2018-12-26 8c062cd
[bugfix] teardown environment
