### 2018-12-26 2c7d800
Changes in getPolyhedron doc
