### 2018-12-26 252541d
Changes in getPolyhedron doc
