### 2018-12-26 afea65d
[bugfix] Wasnt saving polyhedra
