### 2018-12-26 #157 @kenarab
Hi Noam. We just finnished a sprint for code coverage. Now is above 75%.
Will be better in the future but maybe should be enough for Ropensci.
Also checked  goodpractices, lintr, etc.

I hope you a great 2019.

Best, Ale.

El jue., 6 dic. 2018 a las 13:10, alejandro baranek (<
alejandrobaranek@gmail.com>) escribió:

> Hi Noam.
>
> The code not covered is infrastructure for the core functionality. That’s
> why we didn’t included it in test cases, but of course it is a good
> practice to be covered.
>
> We are coding the new test cases in the next days.
>
> We are now in ic18 conference presenting Rpolyhedra.
>
> Thank you,
> All the best,
> Ale.
>
> El 3 dic. 2018, a la(s) 13:05, Noam Ross <notifications@github.com>
> escribió:
>
> Thank you @kenarab <https://github.com/kenarab>. I see that your overall
> test coverage is still low at about 30%
> <https://codecov.io/github/qbotics/Rpolyhedra>. Is there a reason for
> this? Sometimes this occurs because some of the package-building code ends
> up in the package itself. But coverage less than 75% or so is problematic.
>
