### 2018-12-27 fa6d99e
Changes in public functions
