+++ NAMESPACE
@@ -1,5 +1,9 @@
 # Generated by roxygen2: do not edit by hand
 
+export(getAvailablePolyhedra)
+export(getAvailableSources)
+export(getPolyhedron)
+export(switchToFullDatabase)
 importFrom(R6,R6Class)
 importFrom(XML,newXMLDoc)
 importFrom(XML,newXMLNode)
