### 2018-12-27 93c4a1f
Changes in public functions
