+++ R/serialization-lib.R
@@ -41,7 +41,6 @@ convertDFToXML <- function(df, name, node) {
 #'
 #' @importFrom XML newXMLDoc
 #' @importFrom XML newXMLNode
-#' @export
 polyhedronToXML <- function(polyhedron.state.defined,
                             is.transformed.vertices=TRUE) {
   # Start empty XML document tree
