+++ R/ledger-lib.R
@@ -34,7 +34,6 @@ maxWithoutNA <- function(x) ifelse( !all(is.na(x)), max(x, na.rm = TRUE), NA)
 #' @importFrom utils read.csv
 #' @importFrom digest digest
 #' @importFrom R6 R6Class
-#' @export
 ScraperLedger.class <- R6::R6Class("ScraperLedger",
  public = list(
    states = NULL,
