### 2018-12-27 3619a0e
Added testcase for checking db size and existing files
