### 2018-12-27 162eb60
After lintr
