### 2018-12-29 #157 @noamross
Thank you @kenarab .  As it's been a while since I originally contacted them I will see whether the original reviewers are still available. Also, a few notes, which I will ask the reviewers to consider. I would recommend not incorporating them until reviews are in, as well, so you can assess the feedback all together.

- Function documentation is pretty sparse, in that most function descriptions are one-line.  However you also include documentation for a large number of non-exported functions that are developer facing.  This makes for confusing navigation of help.  I'd suggest leaving this documentation only in the source code and not in man pages, which you can do with the `@noRd` tag.  Then focus on expanding the documentation of the smaller number of exported functions.  For instance, there is not documentation on the R6 methods available for the Polyhedron class.
- Similarly, the vignette includes calls to non-exported functions and has little explanatory text.  Removing the section on non-exported function to a developer's vignette or appendix that makes it clear these are internal, package-building functions would improve clarity, and the vignette should spend more effort on explaining the exported functions, the R6 class and how and when to use to use them.
- As I mentioned above, the vignette is more useful if it is a fully compiled R Markdown with inputs and outputs.  **rgl** can do this with WebGL.  For instance, the following code will produce R markdown HTML output with manipulatable WebGL:

    library(rgl)
    library(Rpolyhedra)
    setupKnitr()
    pol <- getPolyhedron(polyhedron.name = "tetrahedron")$getRGLModel()
    shade3d(pol, color = "red")
