### 2018-12-31 4a56377
comment changes as suggested by Noam Ross
