+++ man/getPolyhedraRDSPath.Rd
@@ -1,17 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/env-lib.R
-\name{getPolyhedraRDSPath}
-\alias{getPolyhedraRDSPath}
-\title{getPolyhedraRDSPath}
-\usage{
-getPolyhedraRDSPath(polyhedra_rds_filename = "polyhedra.RDS")
-}
-\arguments{
-\item{polyhedra_rds_filename}{filename of polyhedra database}
-}
-\value{
-the path to the Polyhedra database file
-}
-\description{
-Gets the path of Polyhedra RDS database file
-}
