+++ man/switchToFullDatabase.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/package-lib.R
+% Please edit documentation in R/public-lib.R
 \name{switchToFullDatabase}
 \alias{switchToFullDatabase}
 \title{switchToFullDatabase()}
