### 2018-12-31 b0d2f80
comment changes as suggested by Noam Ross
