+++ R/ledger-lib.R
@@ -1,7 +1,9 @@
-
 #' maxWithoutNA
+#'
 #' Function that returns NA if all elements are NA, and the max value not NA, if not.
+#'
 #' @param x vector parameter
+#' @noRd
 maxWithoutNA <- function(x) ifelse( !all(is.na(x)), max(x, na.rm = TRUE), NA)
 
 
@@ -34,6 +36,7 @@ maxWithoutNA <- function(x) ifelse( !all(is.na(x)), max(x, na.rm = TRUE), NA)
 #' @importFrom utils read.csv
 #' @importFrom digest digest
 #' @importFrom R6 R6Class
+#' @noRd
 ScraperLedger.class <- R6::R6Class("ScraperLedger",
  public = list(
    states = NULL,
