+++ man/initDataDirEnvironment.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/env-lib.R
-\name{initDataDirEnvironment}
-\alias{initDataDirEnvironment}
-\title{initDataDirEnvironment}
-\usage{
-initDataDirEnvironment()
-}
-\value{
-the data dir environment
-}
-\description{
-initialize data enviornment
-}

[…]
