### 2019-01-01 ec1b7fa
improved the documentation as indicated in #157
