+++ man/getPackageEnvir.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/env-lib.R
-\name{getPackageEnvir}
-\alias{getPackageEnvir}
-\title{getPackageEnvir}
-\usage{
-getPackageEnvir(variable.name)
-}
-\arguments{
-\item{variable.name}{name of variable to be retrieved}
-}
-\description{
-Obtains a variable from package environment
-}
