### 2019-01-01 f95f704
improved the documentation as indicated in #157
