### 2019-01-03 62f309d
pkg-minimal for v0.4.0
