### 2019-01-03 19687fe
[bugfix] in copyFilesToExtData
