+++ tests/testthat/test_package_lib.R
@@ -2,12 +2,13 @@
 context("package-lib")
 
 setup(assign(".tmp.package.dir",
-          file.path(getDataDir(data.env = "PACKAGE"), ".tmp/"),
+          file.path(getDataDir(data.env = "PACKAGE"), ".tmp"),
           env = parent.frame()))
 setup(assign(".tmp.home.dir",
-             file.path(getUserSpace(), ".tmp/"),
+             file.path(getUserSpace(), ".tmp"),
              env = parent.frame()))
-
+setup(dir.create(.tmp.package.dir, recursive = TRUE, showWarnings = FALSE))
+setup(dir.create(.tmp.home.dir, recursive = TRUE, showWarnings = FALSE))
 
 test_that("test on package lib functions", {
   testthat::expect(!is.null(getPreloadedDataFilename()),
@@ -33,12 +34,12 @@ test_that("test on package lib functions", {
                   c("SUCCESS", "NOT_AVAILABLE")
      )), failure_message = "downloadRPolyhedraSupportingFiles error"))
 
+
   testthat::expect_equal(copyFilesToExtData(force = FALSE,
                                             source.folder =
-                                              getDataDir(data.env =  "HOME"),
+                                              getDataDir(data.env =  "PACKAGE"),
                                             dest.folder = .tmp.package.dir),
                                                    TRUE)
-
 
   testthat::expect(!is.null(getPackageVersion()))
   testthat::expect(!is.null(getPackageDB()))
