### 2019-01-03 17bdb99
[bugfix] getBaseDir setup in package-lib
