### 2019-01-03 8ca8ec2
added windows/linux/macos targets to travis
