### 2019-01-03 509d93a
removed windows/linux/macos targets to travis, for now it only checks on the default os (linux)
