+++ .travis.yml
@@ -1,10 +1,6 @@
 language: r
 cache: packages
 warnings_are_errors: false
-os:
-  - linux
-  - osx
-  - windows
 r:
   - release
 before_install:
