### 2019-01-03 d3ad9be
Improvement to publishing process. New parameter in selectDataEnv for not downloading HOME database
