### 2019-01-03 8a930c4
Improvement to publishing process. New parameter in selectDataEnv for not downloading HOME database
