### 2019-01-03 dda26e5
Changes in selectDataEnv return value (Now returns getDataEnv())
