### 2019-01-03 4f4a7ab
lintr
