### 2019-01-03 #157 @kenarab
Hello @noamross:

Thank you for your time and your coaching and tips! We applied all of them. You can see them in this [branch](https://github.com/qbotics/Rpolyhedra/tree/dev-0.4.0). Also we just uploaded a new release (v0.4.0) to CRAN.

We hope reviewers (originals or new ones) can take the time to assess Rpolyhedra.

Happy new year for all the Ropensci team.

Best, Ale.
