### 2019-01-04 350341f
removed unneeded code for codecov.io
