+++ .codecov.yml
@@ -1,15 +1,0 @@
-codecov:
- notify:
- require_ci_to_pass: yes
-coverage:
- precision: 2
- round: down
- range: “75…100”
-status:
- project: yes
- patch: yes
- changes: no
-comment:
- layout: “reach, diff, flags, files, footer”
- behavior: default
- require_changes: no
