### 2019-01-06 #157 @noamross
@yulijia and @schloerke have agreed to review.  Due date: 2019-01-30.
