### 2019-01-30 #157 @yulijia
Estimated hours spent reviewing: 2.5 hours

---

I used Fedora release 28 and I was able to install the package from Github without any problem by devtools.
- Function and variable naming is proper.
- README.rmd is clear and easy to follow steps for installations.
- `devtools::check` passes without any warnings or any errors.

This is quite a nice job, all demos are run successfully and very beautiful. The code is neat, but some of the package help pages make me confused. The following comments are addressing minor things that I didn't understand, it may need to be improved.

1. The help page of `Rpolyhedra-package` and `.onLoad()` are same. What does `.onLoad()` used for?

2. How to use `scrapePolyhedra()` and `scrapePolyhedraSources()`? Any examples in the help page? Or maybe they are internal functions?
