### 2019-02-01 0ba8e49
copy in vignette intro
much more descriptive
