### 2019-02-01 1253c37
fix table alignment with non perfect markdown preview
