+++ README.md
@@ -1,6 +1,5 @@
  # Rpolyhedra
 
-  | Release | Usage | Development |
  <!-- Polyhedra database scraped from publically available sources using R6 objects and 'rgl' visualizing capabilities. -->
 
  This package is a curation made based on the poly package found on http://www.netlib.org/polyhedra/ ([Original Help message](poly_original_help_message.html)), and the polyhedra database found on http://dmccooey.com/polyhedra, both of which provide polyhedra databases on its own format. As such, Rpolyhedra provides with the following:
@@ -10,8 +9,9 @@
  1. An R6 polyhedron representation with 'rgl' package visualizing capabilites.
 
 
+| Release | Usage | Development |
 |:--------|:------|:------------|
- [![](https://badges.ropensci.org/157_status.svg)](https://github.com/ropensci/onboarding/issues/157)| [![minimal R version](https://img.shields.io/badge/R%3E%3D-3.4.0-blue.svg)](https://cran.r-project.org/) | [![Travis](https://travis-ci.org/qbotics/Rpolyhedra.svg?branch=master)](https://travis-ci.org/qbotics/Rpolyhedra) |
+| [![](https://badges.ropensci.org/157_status.svg)](https://github.com/ropensci/onboarding/issues/157)| [![minimal R version](https://img.shields.io/badge/R%3E%3D-3.4.0-blue.svg)](https://cran.r-project.org/) | [![Travis](https://travis-ci.org/qbotics/Rpolyhedra.svg?branch=master)](https://travis-ci.org/qbotics/Rpolyhedra) |
 | [![CRAN](http://www.r-pkg.org/badges/version/Rpolyhedra)](https://cran.r-project.org/package=Rpolyhedra) | | [![codecov](https://codecov.io/gh/qbotics/Rpolyhedra/branch/master/graph/badge.svg)](https://codecov.io/gh/qbotics/Rpolyhedra) |
 
 # How to get started
@@ -39,9 +39,9 @@ To get started execute the following commands:
 ```R
 # 1.  Obtain 5 regular solids
 polyhedra.2.draw <- getAvailablePolyhedra(source = "netlib")
-polyhedra.2.draw <- polyhedra.2.draw %>% 
+polyhedra.2.draw <- polyhedra.2.draw %>%
                         filter(scraped.name %in%
-                            c("tetrahedron", "octahedron", "cube", 
+                            c("tetrahedron", "octahedron", "cube",
                                "icosahedron", "dodecahedron"))
 
 # 2. Setup colors and scales
@@ -61,14 +61,14 @@ for (i in seq_len(n)) {
   polyhedron.row <- polyhedra.2.draw[i,]
   polyhedron.name <- polyhedron.row$scraped.name
   polyhedron <- getPolyhedron(source = polyhedron.row$source, polyhedron.name)
-  
+
   # Setup angles, position into transformationMatrix
   current.angle <- i/n * 2 * pi
   tm <- rotationMatrix(current.angle, 1, 0, 0)
   x.pos <- round(polyhedron.scale * sin(current.angle), 2)
   y.pos <- round(polyhedron.scale * cos(current.angle), 2)
   tm <- tm %*% translationMatrix(x.pos, y.pos, 0)
-  
+
   # Render
   print(paste("Drawing ", polyhedron.name, " rotated ", round(current.angle, 2),
               " in (1,0,0) axis. Translated to (", x.pos, ",", y.pos, ",0)",
@@ -78,13 +78,13 @@ for (i in seq_len(n)) {
 }
 
 ```
-## sources 
+## sources
 ### netlib
  Includes 142 polyhedra definitions.
[…]
