### 2019-02-01 e130210
add itemize to package doc file. Separate `.onLoad` from `-package`
