+++ man/Rpolyhedra.Rd
@@ -1,31 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/zzz.R
-\docType{package}
-\name{Rpolyhedra}
-\alias{Rpolyhedra}
-\alias{.onLoad}
-\alias{Rpolyhedra-package}
-\title{Rpolyhedra}
-\usage{
-.onLoad(libname, pkgname)
-}
-\arguments{
-\item{libname}{The library name}
-
-\item{pkgname}{The package name}
-}
-\description{
-polyhedra database
-}
-\details{
-A polyhedra database scraped from:
-* http://paulbourke.net/dataformats/phd/: PHD files as R6 objects and 'rgl'
-         visualizing capabilities. The PHD format was created to describe the geometric
-         polyhedra definitions derived mathematically <http://www.netlib.org/polyhedra/>
-         by Andrew Hume and by the Kaleido program of Zvi Har'El.
-* http://dmccooey.com/Polyhedra: Polyhedra text datafiles.
-}
-\author{
-Alejandro Baranek <abaranek@dc.uba.ar>, Leonardo Javier Belen <leobelen@gmail.com>
-Executes code while loading the package.
-}
