### 2019-02-01 c1b75ff
display the final output in the example
