### 2019-02-01 4082794
do not install.packages in readme.  add `build_opts = NULL`
