+++ README.md
@@ -116,3 +116,5 @@ Ubuntu
 ```bash
 sudo apt-get install r-cran-rgl
 ```
+
+Please note that the 'Rpolyhedra' project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project, you agree to abide by its terms.
