### 2019-02-02 #157 @kenarab
Thank you for your feedback and advices @yulijia and @schloerke!
I just been father of tweens on last wednesday (1-30-2019) so I been full offline for some days. In next days we will analyse your feedback and write back.

We merged the PR  @schloerke made, but wondering about which style code did you use. We are based on [google style code (GSC)](https://google.github.io/styleguide/Rguide.xml). I don't know if RopenSci have a style code down to the level of GSC. Capital Letter starting convention was defined for R6 classes, as there was no reference to R6 classes in GSC.

It may be possible we have some unconsistency in the code, we will check. Maybe will change lintr config @schloerke PR us.

About broken tests and switchToFullDatabase is strange it is not working: we are using Travis CI and working on MacOS for developing and for every release we run all kinds of checks (A call to switchToFullDatabase is included in testcases) for it to work before publishing.
And then, for publishing the package on CRAN all test should work ok.

In which branch did you were working whe you had tests problems?

Thank you! We will reach you on next days.
