### 2019-02-02 #157 @kenarab
about @yulijia question 2:
```
How to use scrapePolyhedra() and scrapePolyhedraSources()? Any examples in the help page? Or maybe they are internal functions?
```

They are development functions. We applied ` @noRd` as @noamross suggested so there are no documentation.
For a package like this, intended to be reproducible, we would prefer have a call like @devRd (A vigentte for developers) for any dev or collaborator who wants to reproduce the backstage would be able to do it.
Reproducible features for the  scraping process is not finished yet. We will research on this while working on this review.
