### 2019-02-04 #157 @noamross
@kenarab First, congratulations on becoming a new father, twice! 👶👶  Thank you @yulijia and @schloerke for your reviews.

Regarding code style: consistency is generally our rule.  The pattern we find most frequent is people use one form for user-facing functions, a different form for internal functions and R6 or S4 types.   With the split in this package user-facing and developer-facing functions, I mostly think that is important that there is consistency within each of these.

Per the reviewer's questions above, the CONTRIBUTING.md file would be an excellent place to explain the overall scraping/developer workflow, referring the reader to the documentation in the source code for details on specific functions.

@sckott is our expert on mocking test functions, he might have some insight into the test failure above.
