### 2019-02-04 f24505d
update objects titles in documentation
