+++ R/test-lib.R
@@ -1,4 +1,4 @@
-#' PolyhedronTestTask
+#' Polyhedron test task
 #'
 #' Is an abstract class for specifying TestTask to make R6 iteration methods
 #' like cover complaint with testhat infrastructure
@@ -28,7 +28,7 @@ public = list(
     stop(gettext("rpoly.abstract_class", domain = "R-Rpolyhedra"))
   }))
 
-#' PolyhedronTestTaskScrape
+#' Polyhedron test task scrape
 #'
 #' A Test task for comparing a new scrape with an already scraped polyhedron
 #' in the database
@@ -88,7 +88,7 @@ PolyhedronTestTaskScrape.class <- R6::R6Class("PolyhedronTestTaskScrape.class",
       expected.polyhedron$getState()$expect_equal(scraped.polyhedron)
     }))
 
-#' PolyhedronTestTaskEdgesConsistency
+#' Polyhedron test task edges consistency
 #'
 #' A Test task for running edges consistency test for current polyhedron
 #'
@@ -116,7 +116,7 @@ PolyhedronTestTaskEdgesConsistency.class <- R6::R6Class(
     }))
 
 
-#' getPercentilPolyhedraQuant
+#' Get percentil polyhedra quantity
 #'
 #' Returns polyhedra quantity of parameter percentil
 #' @param percentil is the percentil which must be applied
