+++ R/public-lib.R
@@ -1,4 +1,4 @@
-#' getPolyhedraObject
+#' Get a polyhedra object
 #'
 #' Return the polyhedra database handler.
 #' @seealso PolyhedraDatabase.class
@@ -9,7 +9,7 @@ getPolyhedraObject <- function() {
   getUserEnvir(".polyhedra")
 }
 
-#' scrapePolyhedra()
+#' Scrape polyedra objects
 #'
 #' Gets polyhedra objects from text files of
 #' different sources, scheduling and scraping using
@@ -32,7 +32,7 @@ scrapePolyhedra <- function(scrape.config,
 
 }
 
-#' scrapePolyhedraSources()
+#' Scrape polyhedra sources
 #'
 #' Scrapes polyhedra objects from text files of
 #' different sources, in order to make them available to the
@@ -87,7 +87,7 @@ scrapePolyhedraSources <- function(sources.config =
 }
 
 
-#' getAvailableSources()
+#' Get available sources
 #'
 #' Gets the list of names of available sources in database to be used later as
 #' references to the package.
@@ -111,7 +111,7 @@ getAvailableSources <- function(){
   getPolyhedraObject()$getAvailableSources()
 }
 
-#' getAvailablePolyhedra()
+#' Get available polyhedra
 #'
 #' Gets the list of names of available polyhedra and its status in
 #' the polyhedra database, which can be later called with getPolyhedron
@@ -143,7 +143,7 @@ getAvailablePolyhedra <- function(sources =
                       search.string = search.string)
 }
 
-#' getPolyhedron()
+#' Get polyhedron
 #'
 #' Gets a polyhedron from the database. It returns an R6 Class
 #' with all its characteristics and functions.
@@ -189,7 +189,7 @@ getPolyhedron <- function(source = "netlib", polyhedron.name) {
   ret
 }
 
-#' switchToFullDatabase()
+#' Switch to full database
 #'
 #' Prompts user for changing database to fulldb in
 #' user filespace. Also, allows the user to switch back to
