### 2019-02-04 #157 @schloerke
Updating here as well. My personal configuration had `~/.R` require sudo to make new folders.  This is non standard.  By fixing my folder permissions, it fixed the function issue and test suite failure.  Sorry for the odd scare!
