### 2019-02-04 ee84ec9
lintr corrections
