### 2019-02-04 443516e
Corrected lintr bot (Not present in local lintr)
