+++ tests/testthat/test_db_lib.R
@@ -2,16 +2,18 @@
 context("db-lib")
 test_that("create minimal db", {
   initDataDirEnvironment()
-  expect_equal(selectDataEnv("PACKAGE"), "PACKAGE")
+  test_that::expect_equal(selectDataEnv("PACKAGE"), "PACKAGE")
   source.config.netlib <- PolyhedronScraperConfigurationNetlib.class$new()
   source.config.dmccooey <- PolyhedronScraperConfigurationDmccoey.class$new()
-  expect_equal(source.config.netlib$getBaseDir("destdir"),
+  test_that::expect_equal(source.config.netlib$getBaseDir("destdir"),
                file.path("destdir", "sources", "www.netlib.org", "polyhedra"))
-  expect_equal(source.config.dmccooey$getBaseDir("destdir"),
+  test_that::expect_equal(source.config.dmccooey$getBaseDir("destdir"),
                file.path("destdir", "sources", "dmccooey.com", "polyhedra"))
-  expect_equal(length(source.config.netlib$getPolyhedraFiles(getDataDir())),
+  test_that::expect_equal(
+              length(source.config.netlib$getPolyhedraFiles(getDataDir())),
                119)
-  expect_equal(length(source.config.dmccooey$getPolyhedraFiles(getDataDir())),
+  test_that::expect_equal(
+              length(source.config.dmccooey$getPolyhedraFiles(getDataDir())),
                9)
 
   db <- PolyhedraDatabase.class$new()
@@ -31,7 +33,7 @@ test_that("create minimal db", {
             skip.still.queued = FALSE, save.on.change = FALSE)
   db$scrape(mode = "scrape.queued", sources = "dmccooey", max.quant = 3,
             skip.still.queued = TRUE, save.on.change = FALSE)
-  expect_equal(db$getAvailablePolyhedra()$scraped.name,
+  test_that::expect_equal(db$getAvailablePolyhedra()$scraped.name,
                c("tetrahedron", "octahedron", "cube",
                  "10-truncated triakis icosahedron (canonical)",
     "4-5-truncated deltoidal hexecontahedron with truncation depth chosen",
@@ -41,7 +43,7 @@ test_that("create minimal db", {
   tasks <- db$
     generateTestTasks(TestTaskClass = PolyhedronTestTaskScrape.class,
                       max.quant = 3)
-  expect_equal(length(tasks), 3)
+  test_that::expect_equal(length(tasks), 3)
   #TODO
   #testRR
 })
