### 2019-02-04 b6085da
fixed testthat reference
