+++ tests/testthat/test_db_lib.R
@@ -2,17 +2,17 @@
 context("db-lib")
 test_that("create minimal db", {
   initDataDirEnvironment()
-  test_that::expect_equal(selectDataEnv("PACKAGE"), "PACKAGE")
+  testthat::expect_equal(selectDataEnv("PACKAGE"), "PACKAGE")
   source.config.netlib <- PolyhedronScraperConfigurationNetlib.class$new()
   source.config.dmccooey <- PolyhedronScraperConfigurationDmccoey.class$new()
-  test_that::expect_equal(source.config.netlib$getBaseDir("destdir"),
+  testthat::expect_equal(source.config.netlib$getBaseDir("destdir"),
                file.path("destdir", "sources", "www.netlib.org", "polyhedra"))
-  test_that::expect_equal(source.config.dmccooey$getBaseDir("destdir"),
+  testthat::expect_equal(source.config.dmccooey$getBaseDir("destdir"),
                file.path("destdir", "sources", "dmccooey.com", "polyhedra"))
-  test_that::expect_equal(
+  testthat::expect_equal(
               length(source.config.netlib$getPolyhedraFiles(getDataDir())),
                119)
-  test_that::expect_equal(
+  testthat::expect_equal(
               length(source.config.dmccooey$getPolyhedraFiles(getDataDir())),
                9)
 
@@ -33,7 +33,7 @@ test_that("create minimal db", {
             skip.still.queued = FALSE, save.on.change = FALSE)
   db$scrape(mode = "scrape.queued", sources = "dmccooey", max.quant = 3,
             skip.still.queued = TRUE, save.on.change = FALSE)
-  test_that::expect_equal(db$getAvailablePolyhedra()$scraped.name,
+  testthat::expect_equal(db$getAvailablePolyhedra()$scraped.name,
                c("tetrahedron", "octahedron", "cube",
                  "10-truncated triakis icosahedron (canonical)",
     "4-5-truncated deltoidal hexecontahedron with truncation depth chosen",
@@ -43,7 +43,7 @@ test_that("create minimal db", {
   tasks <- db$
     generateTestTasks(TestTaskClass = PolyhedronTestTaskScrape.class,
                       max.quant = 3)
-  test_that::expect_equal(length(tasks), 3)
+  testthat::expect_equal(length(tasks), 3)
   #TODO
   #testRR
 })
