+++ tests/testthat/test_ledger_lib.R
@@ -32,13 +32,13 @@ test_that("create minimal ledger", {
   # Execute functions
   ledger.crc   <-  ledger$getCRCPolyhedronName(source = "netlib",
                                                polyhedron.name = "tetrahedron")
-  test_that::expect_equal(ledger$getAvailableSources(), "netlib")
-  test_that::expect_equal(ledger$getAvailablePolyhedra()$scraped.name,
+  testthat::expect_equal(ledger$getAvailableSources(), "netlib")
+  testthat::expect_equal(ledger$getAvailablePolyhedra()$scraped.name,
                                         c("tetrahedron", "octahedron", "cube"))
-  test_that::expect_equal(
+  testthat::expect_equal(
             unique(ledger$getFilenamesStatus(status = "scraped")$status),
             "scraped")
-  test_that::expect_equal(
+  testthat::expect_equal(
             unique(ledger$getFilenamesStatusMode(mode = "test")$status),
             "scraped")
 
