### 2019-02-04 ca93ab7
fixed testthat reference
