+++ tests/testthat/test_package_lib.R
@@ -10,7 +10,7 @@ setup(assign(".tmp.home.dir",
 setup(dir.create(.tmp.package.dir, recursive = TRUE, showWarnings = FALSE))
 setup(dir.create(.tmp.home.dir, recursive = TRUE, showWarnings = FALSE))
 
-test_that("test on package lib functions", {
+testthat::test_that("test on package lib functions", {
   testthat::expect(!is.null(getPreloadedDataFilename()),
                    failure_message = "getPreloadedDataFilename cannot be null")
   testthat::expect(!is.null(updatePolyhedraDatabase()),
