### 2019-02-04 15e0b01
with_mock false snake case correction
