+++ tests/testthat/test_package_lib.R
@@ -15,17 +15,17 @@ testthat::test_that("test on package lib functions", {
                    failure_message = "getPreloadedDataFilename cannot be null")
   testthat::expect(!is.null(updatePolyhedraDatabase()),
                    failure_message = "updatePolyhedraDatabase cannot be null")
-  with_mock(
+  testthat::with_mock(
        "Rpolyhedra::getUserSpace" = function(){
             .tmp.home.dir
          },
        testthat::expect(
-       with_mock(
+         testthat::with_mock(
          "Rpolyhedra::getDataEnv" = function(){
            "HOME"
            }
          ,
-         with_mock(
+         testthat::with_mock(
            "Rpolyhedra::checkDatabaseVersion" = function(){
              "UPDATE"
              }
