### 2019-02-04 #157 @kenarab
Thank you @noamross! The twins are in neonatology, but they are very good. They are growing really fast and will go to home soon.

Dear @yulijia and @schloerke and @noamross . We implemented most of suggestions.
* We configured lintr for CamelCase and DotCase. Removed Snake case and found some internal inconsistencies, but most were related to dependencies, specifically testthat. For avoiding lintr messages added testthat spacename in that callings. But it is specific to some lintr version and others no.
* About code style we still using google code style. With exception of R6 objects, as discussed previously. This was included in CONTRIBUTING.md as matters to developers only
* We included a CONTRIBUTING.md which explain some basics of the package for developers. We will improve the description in next versions. For example, for people wanting to program scrapers for other sources.
* About @yulijia question `The help page of Rpolyhedra-package and .onLoad() are same. What does .onLoad() used for?`: @schloerke the other reviewer made a PR fixing this documentation issue. .onLoad function is a standard function for R package with initialization scripts regarding a package.

Thank you very much for your time.

Best, Ale
