### 2019-02-14 #157 @noamross
Thank you all.  @yulijia and  @schloerke, please let us know if @kenarab's changes have addressed all your comments. If you have any additional comments on interface or use cases that you may not have gotten to while working through testing and documentation issues, feel free to provide these, as well.
