### 2019-02-14 #157 @schloerke
Good to me!

Took me a couple times to realize the Google Coding Style is a mixed style to help identify a variable vs a package function.  I've been trained on a "all or nothing" style (preferring snake case, which google does not like).
