### 2019-02-15 #157 @kenarab
Dear all:

We have made cleanup of Rpolyhedra repo. If you have uncommited code or PR,
please clone the repo again before making changes.

Best, Ale.

On Thu, Feb 14, 2019 at 12:35 PM Barret Schloerke <notifications@github.com>
wrote:

> Good to me!
>
> Took me a couple times to realize the Google Coding Style is a mixed style
> to help identify a variable vs a package function. I've been trained on a
> "all or nothing" style (preferring snake case, which google does not like).
>
