### 2019-02-25 4adb847
fixed typo on CONTIBUTING.md
