### 2019-02-25 a566df6
Compliant with ropensci/software-review#157
