+++ DESCRIPTION
@@ -4,7 +4,10 @@ Title: Polyhedra Database
 Version: 0.4.1
 Authors@R: c(
     person("Alejandro", "Baranek",,"abaranek@dc.uba.ar", c("aut","com","cre", "cph")),
-    person("Leonardo","Belen",,"leobelen@gmail.com", c("aut","com","cph"))
+    person("Leonardo","Belen",,"leobelen@gmail.com", c("aut","com","cph")),
+    person("Noam","Ross",,"noam.ross@gmail.com", c("rev")),
+    person("Barret","Schloerke",,"schloerke@gmail.com", c("rev")),
+    person("Lijia","Yu",,"yu@lijiayu.net", c("rev"))
     )
 Maintainer: Alejandro Baranek <abaranek@dc.uba.ar>
 Description: A polyhedra database scraped from various sources as R6 objects and 'rgl' visualizing capabilities.
