### 2019-02-25 ff497e0
added pkgdown doc folder
