### 2019-02-25 54c338f
updated example on vignette
