### 2019-02-25 #157 @kenarab
Thank you, Noam!

I looked through `?person`and saw there is no role for editor. So you will
not be mentioned in authors@R, ok?

Can you please invite @leobelen to RopenSci and add him to Rpolyhedra team?
I see have no privileges for doing and he is actually an author of the
package.

Best, Ale.
