### 2019-02-25 ffe177f
[bugfix] ^docs in .Rbuildignore
