### 2019-02-25 #157 @stefaniebutland
@kenarab  I'm rOpenSci's Community Manager, here to say we would love to feature a post about `Rpolyhedra` on our blog.

This link will give you many examples of blog posts by authors of peer-reviewed packages so you can get an idea of the style and length you prefer: https://ropensci.org/tags/software-peer-review/. Here are some technical and editorial guidelines for contributing a blog post: https://github.com/ropensci/roweb2#contributing-a-blog-post.

Please let me know when you might want to submit a draft. Happy to answer any questions.
