### 2019-03-07 892f8d9
added a NEWS.md file
