+++ NEWS.md
@@ -1,0 +1,32 @@
+Rpolyhedra 0.4.1
+============
+
+First version on [rOpenSci.org](http://www.rOpenSci.org). 
+
+### MINOR IMPROVEMENTS
+
+* Complies with all the prerequisites of rOpenSci and applies the suggestions made by rOpenSci reviwers. 
+* Fixes a test that writes on user space.
+* Integrates with codecov.io, which allows for better test coverage. 
+* Updated examples.
+
+Rpolyhedra 0.4.0
+============
+
+### NEW FEATURES
+
+* `Rpolyhedra` can export polyhedra definitions as XML.
+
+
+Rpolyhedra 0.3.0
+============
+
+### NEW FEATURES
+
+* `Rpolyhedra` now has a new database format based on ascii RDSs, which are meant to use less memory, for example when used in a Shiny App.
+* `Rpolyhedra` now uses a transformation matrix for general polyhedra manipulation.
+
+### MINOR IMPROVEMENTS
+
+* Applied suggestions from rOpenSci onboarding process. 
+
