### 2019-03-16 3e7a4b9
Changed BoundingBox to ConvexHull from geometry
