+++ tests/testthat/test_serialization_lib.R
@@ -3,5 +3,5 @@ testthat::test_that("test xml can be exported", {
   xml.polyhedron <- polyhedronToXML(
           getPolyhedron("netlib", "tetrahedron")$getState())
   testthat::expect_equal(
-    as.numeric(nchar(XML::getChildrenStrings(xml.polyhedron))), 406)
+    as.numeric(nchar(XML::getChildrenStrings(xml.polyhedron))), 408)
 })
