### 2019-03-16 7b88a56
[bugfix] geometry::convhulln wasn't imported in NAMESPACE
