+++ NAMESPACE
@@ -12,6 +12,7 @@ importFrom(XML,newXMLNode)
 importFrom(digest,digest)
 importFrom(futile.logger,flog.debug)
 importFrom(futile.logger,flog.info)
+importFrom(geometry,convhulln)
 importFrom(git2r,commits)
 importFrom(rgl,asHomogeneous)
 importFrom(rgl,identityMatrix)
