### 2019-03-18 c2a979a
buildRGL applies internal transformation matrix
