+++ vignettes/Rpolyhedra.Rmd
@@ -12,11 +12,11 @@ This package is a curation made based on the poly package found on http://www.ne
 
 1. A module to scrape the polyhedra for the different sources found with features for incremental correction of issues found and to be found in scraping process.
 1. A database of the scraped polyhedra.
-1. An R6 polyhedron representation with 'rgl' package visualizing capabilites. 
+1. An R6 polyhedron representation with 'rgl' package visualizing capabilities. 
 
 # Usage
 
-For final users, the package provides a common interface for accesing public polyhedra databases, analize properties, compare and visualize them with RGL. 
+For final users, the package provides a common interface for accessing public polyhedra databases, analyze properties, compare and visualize them with RGL. 
 
 For advanced users, the package provides a simplified set of R6 objects to scrape and compare polyhedra databases. 
 ```{r setup, include=FALSE}
@@ -43,7 +43,7 @@ cube <- getPolyhedron(source = "netlib", polyhedron.name = "cube")
 ```
 
 # A demo
-To try package funcionality, a simple demo can be executed which shows the 5 regular polyhedra.
+To try package functionality, a simple demo can be executed which shows the 5 regular polyhedra.
 
 ```{r demo, webgl=TRUE}
 # 1.  Obtain 5 regular solids
