+++ NEWS.md
@@ -1,3 +1,9 @@
+Rpolyhedra 0.4.2
+============
+
+### BUG FIXES
+
+
 Rpolyhedra 0.4.1
 ============
 
@@ -5,7 +11,7 @@ First version on [rOpenSci.org](http://www.rOpenSci.org).
 
 ### MINOR IMPROVEMENTS
 
-* Complies with all the prerequisites of rOpenSci and applies the suggestions made by rOpenSci reviwers. 
+* Complies with all the prerequisites of rOpenSci and applies the suggestions made by rOpenSci reviewers. 
 * Fixes a test that writes on user space.
 * Integrates with codecov.io, which allows for better test coverage. 
 * Updated examples.
