+++ DESCRIPTION
@@ -41,3 +41,4 @@ Collate:
   'zzz.R'
 BugReports: https://github.com/ropensci/Rpolyhedra/issues
 URL: https://github.com/ropensci/Rpolyhedra
+LANGUAGE: en-US
