+++ tests/testthat.R
@@ -2,6 +2,8 @@ library(Rpolyhedra)
 library(stringr)
 library(futile.logger)
 library(rgl)
+library(geometry)
+library(testthat)
 
 
 #Change threshold to ERROR. Comment out/change if verbosity required for development
