### 2019-03-26 fdff31a
NEWS for 0.4.2
