+++ DESCRIPTION
@@ -2,6 +2,7 @@ Package: Rpolyhedra
 Type: Package
 Title: Polyhedra Database
 Version: 0.4.2
+Language: en-US
 Authors@R: c(
     person("Alejandro", "Baranek",,"abaranek@dc.uba.ar", c("aut","com","cre", "cph")),
     person("Leonardo","Belen",,"leobelen@gmail.com", c("aut","com","cph")),
@@ -41,4 +42,3 @@ Collate:
   'zzz.R'
 BugReports: https://github.com/ropensci/Rpolyhedra/issues
 URL: https://github.com/ropensci/Rpolyhedra
-LANGUAGE: en-US
