+++ DESCRIPTION
@@ -28,7 +28,7 @@ Imports: R6,
          digest,
          testthat,
          futile.logger,
-         git2r,
+         git2r
 Collate:
   'Rpolyhedra-package.R'
   'polyhedra-lib.R'
